Multistage deterministic record linkage

Link records in ordered stages with flexible matching conditions.

links(
  criteria,
  sub_criteria = NULL,
  sn = NULL,
  strata = NULL,
  data_source = NULL,
  data_links = "ANY",
  display = "progress",
  group_stats = FALSE,
  expand = TRUE,
  shrink = FALSE
)

record_group(df, ..., to_s4 = TRUE)

Arguments

criteria
list of attributes to compare at each stage. Comparisons are done as an exact match i.e. (==). See Details.

sub_criteria
list of additional attributes to compare at each stage. Comparisons are done as an exact match or with user-defined logical tests function. See sub_criteria

sn
Unique numerical record identifier. Useful for creating familiar episode identifiers.

strata
Subsets. Record groups are tracked separately within each subset.

data_source
Unique data source identifier. Useful when the dataset contains data from multiple sources.

data_links
A set of data_sources required in each record group. A strata without records from these data sources will be skipped, and record groups without these will be unlinked. See Details.

display
The messages printed on screen. Options are; "none" (default) or, "progress" and "stats" for a progress update or a more detailed breakdown of the linkage process.

group_stats
If TRUE (default), group-specific information like record counts. See Value.

expand
If TRUE, allows increases in the size of a record group at subsequent stages of the linkage process.

shrink
If TRUE, allows reductions in the size of a record group at subsequent stages of the linkage process.

df
data.frame. One or more datasets appended together. See Details.

...
Arguments passed to links

to_s4
Data type of returned object. pid (TRUE) or data.frame (FALSE).

Value

pid objects or data.frame if to_s4 is FALSE)

sn - unique record identifier as provided (or generated)

pid | .Data - unique group identifier

link_id - unique record identifier of matching records

pid_cri - matching criteria

pid_dataset - data sources in each group

pid_total - number of records in each group

iteration - iteration of the process when each record was linked to its record group

Details

links() performs an ordered multistage deterministic linkage. The relevance or priority of each stage is determined by the order in which they have been listed.

sub_criteria specifies additional matching conditions for each stage (criteria) of the process. If sub_criteria is not NULL, only records with matching criteria and sub_criteria are linked. If a record has missing values for any criteria, that record is skipped at that stage, and another attempt is made at the next stage. If there are no matches for a record at every stage, that record is assigned a unique group ID.

By default, records are compared for an exact match. However, user-defined logical tests (function) are also permitted. The function must be able to compare two atomic vectors and return either TRUE or FALSE. The function must have two arguments - x for the attribute and y for what it'll be compared against.

A match at each stage is considered more relevant than a match at the next stage. Therefore, criteria should always be listed in order of decreasing relevance.

data_source - including this populates the pid_dataset slot. See Value.

data_links should be a list of atomic vectors with every element named "l" (links) or "g" (groups).

"l" - Record groups with records from every listed data source will be retained.

"g" - Record groups with records from any listed data source will be retained.

data_links is useful for skipping record groups that are not required.

record_group() as it existed before v0.2.0 has been retired. Its now exists to support previous code and arguments with minimal disruption. Please use links() moving forward.

See vignette("links") for more information.

See also

Examples

library(diyar)
# Exact match
links(criteria = c("Obinna","James","Ojay","James","Obinna"))

#> 
#> Checked `criteria 1`.
#> 
#> Summary.
#> Time elapsed: < 0.01 secs.
#> Groups:     3
#> Records:    5
#>    CRI 001: 4
#>    No Hits: 1

#> [1] "P.1 (CRI 001)" "P.2 (CRI 001)" "P.3 (No Hits)" "P.2 (CRI 001)"
#> [5] "P.1 (CRI 001)"

# User-defined tests using `sub_criteria()`
# Matching `sex` and + 20-year age gaps
age <- c(30, 28, 40, 25, 25, 29, 27)
sex <- c("M", "M", "M", "F", "M", "M", "F")
f1 <- function(x, y) (y - x) %in% 0:20
links(criteria = sex,
      sub_criteria = list(s1 = sub_criteria(age, funcs = f1)))

#> 
#> Checked `criteria 1`.
#> 
#> Summary.
#> Time elapsed: < 0.01 secs.
#> Groups:     2
#> Records:    7
#>    CRI 001: 7

#> [1] "P.1 (CRI 001)" "P.1 (CRI 001)" "P.1 (CRI 001)" "P.4 (CRI 001)"
#> [5] "P.1 (CRI 001)" "P.1 (CRI 001)" "P.4 (CRI 001)"

# Multistage linkage
# Relevance of matches: `forename` > `surname`
data(staff_records); staff_records

#>   r_id forename  surname sex    dataset
#> 1    1    James    Green   M Staff list
#> 2    2     <NA> Anderson   M Staff list
#> 3    3    Jamey    Green   M  Pay slips
#> 4    4              <NA>   F  Pay slips
#> 5    5  Derrick Anderson   M Staff list
#> 6    6  Darrack Anderson   M  Pay slips
#> 7    7 Christie    Green   F Staff list

links(criteria = list(staff_records$forename, staff_records$surname),
      data_source = staff_records$sex)

#> 
#> Checked `criteria 1`.
#> Checked `criteria 2`.
#> 
#> Summary.
#> Time elapsed: < 0.01 secs.
#> Groups:     3
#> Records:    7
#>    CRI 002: 6
#>    No Hits: 1

#> [1] "P.1 (CRI 002)" "P.2 (CRI 002)" "P.1 (CRI 002)" "P.4 (No Hits)"
#> [5] "P.2 (CRI 002)" "P.2 (CRI 002)" "P.1 (CRI 002)"

# Relevance of matches:
# `staff_id` > `age` AND (`initials`, `hair_colour` OR `branch_office`)
data(missing_staff_id); missing_staff_id

#>   r_id staff_id age initials hair_colour     branch_office source_1 source_2
#> 1    1       NA  30     G.D.       Brown Republic of Ghana        A        3
#> 2    2       NA  30     B.G.        Teal            France        A        1
#> 3    3       NA  30     X.P.        <NA>              <NA>        A        1
#> 4    4       NA  30     X.P.       Green              <NA>        B        1
#> 5    5       NA  30     <NA>       Green            France        A        1
#> 6    6        2  30     G.D.  Dark brown             Ghana        A        1
#> 7    7        2  30     G.D.       Brown Republic of Ghana        B        2

links(criteria = list(missing_staff_id$staff_id, missing_staff_id$age),
      sub_criteria = list(s2 = sub_criteria(missing_staff_id$initials,
                                          missing_staff_id$hair_colour,
                                          missing_staff_id$branch_office)),
      data_source = missing_staff_id$source_1)

#> 
#> Checked `criteria 1`.
#> Checked `criteria 2`.
#> 
#> Summary.
#> Time elapsed: < 0.01 secs.
#> Groups:     1
#> Records:    7
#>    CRI 001: 2
#>    CRI 002: 5

#> [1] "P.6 (CRI 002)" "P.6 (CRI 002)" "P.6 (CRI 002)" "P.6 (CRI 002)"
#> [5] "P.6 (CRI 002)" "P.6 (CRI 001)" "P.6 (CRI 001)"