Create epid and pid objects with index of matching records

Create epid and pid objects with index of matching records

make_episodes(
  x_pos,
  y_pos,
  x_val,
  date,
  case_nm,
  wind_id,
  wind_nm,
  from_last,
  data_source,
  data_links,
  iteration,
  options,
  episode_unit
)

make_pids(
  x_pos,
  y_pos,
  x_val,
  link_id,
  pid_cri,
  data_source,
  data_links,
  iteration
)

Arguments

x_pos

[integer]. Index of one half of a record pair.

y_pos

[integer]. Index of one half of a record pair.

x_val

[integer]. Value of one half of a record pair.

date

[date|datetime|integer|number_line]. Record date or period.

case_nm

[integer|character] Record type in regards to case assignment (sub_criteria[Encoded]).

wind_id

[integer]. Unique reference ID for each match.

wind_nm

[list]. Type of window i.e. "Case" or "Recurrence".

from_last

[logical]. Chronological order of episode tracking i.e. ascending (TRUE) or descending (FALSE).

iteration

The iteration when a record was matched to it's group (.Data).

options

[list]. Some options passed to the instance of episodes.

episode_unit

[character]. Time unit for case_length and recurrence_length. See episodes

link_id

[integer]. Unique reference ID for each match.

pid_cri

Match stage of the step-wise linkage.