Overlapping number line objects

Identify overlapping number_line objects

overlaps(x, y, methods = 8)

overlap(x, y)

exact(x, y)

reverse(x, y)

across(x, y)

chain(x, y)

aligns_start(x, y)

aligns_end(x, y)

inbetween(x, y)

overlap_method(x, y)

include_overlap_method(methods)

exclude_overlap_method(methods)

overlap_method_codes(methods)

Arguments

x
[number_line]

y
[number_line]

methods
[charater|integer]. Methods of overlap. Check different pairs of number_line objects by different methods. Options are "exact", "reverse", "inbetween", "across", "chain", "aligns_start" and "aligns_end". Combinations are also supported see diyar::overlap_methods$options.

Value

logical; character

Details

9 logical test;

exact() - Identical left and right points.

reverse() - Swapped left and right points.

inbetween() - start and end point of one number_line object is within the start and end point of another.

across() - start or end point of one number_line object is in between the start and end point of another.

chain() - endpoint of one number_line object is the same as the start point of another.

aligns_start() - identical start points only.

aligns_end() - identical end point only.

overlap() - any kind of overlap. A convenient method for "ANY" and "ALL" methods of overlap.

overlaps() - overlap by a specified combination of the methods.

Describe methods of overlap;

overlap_method() - Shows how a pair of number_line object has overlapped. Does not show "overlap" since overlap() is always TRUE when any other method is TRUE.

include_overlap_method() and exclude_overlap_method() - Conveniently create the required values for methods, and case_overlap_methods and recurrence_overlap_methods in episodes.

overlap_method_codes() - Numeric codes for the supported combination of overlap methods.

See also

Examples

a <- number_line(-100, 100)
b <- number_line(10, 11.2)
c <- number_line(100, 200)
d <- number_line(100, 120)
e <- number_line(50, 120)
g <- number_line(100, 100)
f <- number_line(120, 50)

overlaps(a, g)

#> [1] TRUE

overlaps(a, g, methods = "exact|chain")

#> [1] FALSE

overlap(a, b)

#> [1] TRUE

overlap(a, e)

#> [1] TRUE

exact(a, g)

#> [1] FALSE

exact(a, a)

#> [1] TRUE

reverse(e, e)

#> [1] FALSE

reverse(e, f)

#> [1] TRUE

across(a, b)

#> [1] FALSE

across(a, e)

#> [1] TRUE

chain(c, d)

#> [1] FALSE

chain(a, c)

#> [1] TRUE

aligns_start(c, d)

#> [1] TRUE

aligns_start(a, c)

#> [1] FALSE

aligns_end(d, e)

#> [1] TRUE

aligns_end(a, c)

#> [1] FALSE

inbetween(a, g)

#> [1] FALSE

inbetween(b, a)

#> [1] TRUE

overlap_method(a, c)

#> [1] "chain"

overlap_method(d, c)

#> [1] "aligns_start"

overlap_method(a, g)

#> [1] "aligns_end"

overlap_method(b, e)

#> [1] "none"

include_overlap_method("across")

#> [1] 1

include_overlap_method(c("across", "chain"))

#> [1] 12

exclude_overlap_method("across")

#> [1] 440

exclude_overlap_method(c("across", "chain"))

#> [1] 338

overlap_method_codes("across")

#> [1] 1

overlap_method_codes("across|chain|exact")

#> [1] 59