Predefined logical tests in diyar

A collection of predefined logical tests used with sub_criteria objects.

exact_match(x, y)

range_match(x, y, range = 10)

range_match_legacy(x, y)

prob_link(
  x,
  y,
  cmp_func,
  attr_threshold,
  score_threshold,
  probabilistic,
  return_weights
)

Arguments

x
Value of an attribute(s) to be compare against.

y
Value of an attribute(s) to be compare by.

range
Difference between y and x.

cmp_func
Logical tests such as string comparators. See links_wf_probabilistic.

attr_threshold
Matching set of weight thresholds for each result of cmp_func. See links_wf_probabilistic.

score_threshold
Score threshold determining matched or linked records. See links_wf_probabilistic.

probabilistic
If TRUE, matches determined through a score derived base on Fellegi-Sunter model for probabilistic linkage. See links_wf_probabilistic.

return_weights
If TRUE, returns the match-weights and score-thresholds for record pairs. See links_wf_probabilistic.

Details

exact_match() - test that x == y

range_match() - test that x \(\le\) y \(\le\) (x + range)

range_match_legacy() - test that overlap(as.number_line(x@gid), y) is TRUE.

prob_link() - Test that a record-sets x and y are from the same entity based on calculated weights and probability scores.

Examples

`exact_match`

#> function(x, y) {
#>   x == y & !is.na(x) & !is.na(y)
#> }
#> <bytecode: 0x000000002485e468>
#> <environment: namespace:diyar>

exact_match(x = 1, y = 1)

#> [1] TRUE

exact_match(x = 1, y = 2)

#> [1] FALSE

`range_match`

#> function(x, y, range = 10){
#>   x <- as.numeric(x); y <- as.numeric(y)
#>   (x <= y) & (y <= x + range) & !is.na(x) & !is.na(y)
#> }
#> <bytecode: 0x00000000244dd618>
#> <environment: namespace:diyar>

range_match(x = 10, y = 16, range = 6)

#> [1] TRUE

range_match(x = 16, y = 10, range = 6)

#> [1] FALSE

`range_match_legacy`

#> function(x, y) {
#>   lg <- overlap(as.number_line(x@gid), x)
#>   lg[is.na(lg)] <- F
#>   if(any(!lg)) {
#>     rng_i <- paste(head(which(!lg), 5), collapse = ", ", sep="")
#>     rng_v <- as.character(substitute(x))[!as.character(substitute(x)) %in% c("$","df2")]
#>     stop(paste0("Range matching error: Actual value (gid) is out of range in ", "[", rng_i, "]"))
#>   }
#>   overlap(as.number_line(x@gid), y)
#> }
#> <bytecode: 0x00000000240bf120>
#> <environment: namespace:diyar>

x_nl <- number_line(10, 16, gid = 10)
y_nl1 <- number_line(16, 10)
y_nl2 <- number_line(16, 10)

range_match_legacy(x = x_nl, y = y_nl1)

#> [1] TRUE

range_match_legacy(x = x_nl, y = y_nl2)

#> [1] TRUE