Episode grouping for record deduplication and case assignment

This function assigns unique episode indentifiers to chronological episodes

episode_group(df, sn = NULL, strata = NULL, date, case_length,
  episode_type = "fixed", episode_unit = "days", episodes_max = Inf,
  recurrence_length = NULL, rolls_max = Inf, data_source = NULL,
  custom_sort = NULL, from_last = FALSE, overlap_method = c("across",
  "inbetween", "aligns_start", "aligns_end", "chain"),
  bi_direction = FALSE, group_stats = FALSE, display = TRUE)

Arguments

df
Dataframe. One or more datasets appended together.

sn
Unique numeric record indentifier. Optional.

strata
Column names. Episodes will be unique to each strata. record_group can be used to create stratas

date
Record date or interval. date, datetime or interval objects.

case_length
Period from a "Case" within which another record of the same strata is considered a "Duplicate" record.

episode_type
"fixed" or "rolling".

episode_unit
Time units as supported by lubridate's duration function.

episodes_max
Maximum number of times to group episodes in each strata.

recurrence_length
Period from the last record of an episode within which another record of the same strata is considered a "Recurrent" record. If epidsode_type is "rolling" and recurrence_length is not supplied, the case_length is used.

rolls_max
Maximum number of recurrence permitted within each episode. Only used if episode_type is "rolling".

data_source
Unique dataset indentifier for the dataframe. Useful when dataframe contains multiple datsets.

custom_sort
If TRUE, "Case" assignment will be in preference to this sort order. Useful in specifying that episode grouping begins at a particular kind of record regardless of chronological order.

from_last
If TRUE, episode grouping will be backwards in time - starting at the most recent record and proceeding to the earliest. If FALSE, it'll be forward in time - starting at the earliest record and proceeding to the most recent one.

overlap_method
A set of methods for grouped intervals to overlap. Options are; "across", "aligns_start", "aligns_end", "inbetween", "chain". See overlap functions.

bi_direction
If FALSE, "Duplicate" records will be those within the case_length and recurrence_length, before or after the "Case" as determined by from_last. If TRUE, "Duplicate" records will be those on both sides of the "Case".

group_stats
If TRUE, output will include additional columns with useful stats for each episode.

display
If TRUE, status messages are printed on screen.

Value

Dataframe

sn - unique record identifier as provided

epid - unique episode indentifier

case_nm - record type in regards to case assignment

epid_dataset - datasets contained in each episode

epid_interval - Episode start and end dates. Lubridate's interval object.

epid_length - Difference between episode start and end dates. difftime object. If possible, the same unit supplied to episode_unit is used otherwise, a difference in days is returned

epid_total - number of records in each record group

Details

Episode grouping begins at a reference record ("Case") and proceeds forward or backward in time depending on from_last. If custom_sort is used, episode grouping can be forced to begin at certain record before proceeding forward or backwards in time. The maximun duration of a "fixed" episode is the case_length while, the maximum duration of a "rolling" episode is the case_length in addition to all recurrence periods. A recurrence period is the recurrence_length from the last record in an episode

See also

Examples

library(lubridate)

#> Warning: package 'lubridate' was built under R version 3.5.3

#> 
#> Attaching package: 'lubridate'

#> The following object is masked from 'package:base':
#> 
#>     date

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:lubridate':
#> 
#>     intersect, setdiff, union

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

library(diyar)

data(infections); infections

#> # A tibble: 11 x 4
#>    rd_id date       infection epi_len
#>    <int> <date>     <chr>       <dbl>
#>  1     1 2018-04-01 BSI            15
#>  2     2 2018-04-07 UTI            15
#>  3     3 2018-04-13 UTI            15
#>  4     4 2018-04-19 UTI            15
#>  5     5 2018-04-25 BSI            15
#>  6     6 2018-05-01 UTI            15
#>  7     7 2018-05-07 BSI            15
#>  8     8 2018-05-13 BSI            15
#>  9     9 2018-05-19 RTI            15
#> 10    10 2018-05-25 RTI            15
#> 11    11 2018-05-31 BSI            15

# 16-hour (difference of 15 hours) episodes, and the most recent record defined as the "Case"
epids <- episode_group(infections, sn = rd_id, date = date, case_length = epi_len,
from_last = TRUE, episode_unit = "hours", group_stats = TRUE)

#> Episode or recurrence window 1 .
#> 1 of 11 record(s) grouped into episodes. 10 records not yet grouped.
#> Episode or recurrence window 2 .
#> 1 of 10 record(s) grouped into episodes. 9 records not yet grouped.
#> Episode or recurrence window 3 .
#> 1 of 9 record(s) grouped into episodes. 8 records not yet grouped.
#> Episode or recurrence window 4 .
#> 1 of 8 record(s) grouped into episodes. 7 records not yet grouped.
#> Episode or recurrence window 5 .
#> 1 of 7 record(s) grouped into episodes. 6 records not yet grouped.
#> Episode or recurrence window 6 .
#> 1 of 6 record(s) grouped into episodes. 5 records not yet grouped.
#> Episode or recurrence window 7 .
#> 1 of 5 record(s) grouped into episodes. 4 records not yet grouped.
#> Episode or recurrence window 8 .
#> 1 of 4 record(s) grouped into episodes. 3 records not yet grouped.
#> Episode or recurrence window 9 .
#> 1 of 3 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 10 .
#> 1 of 2 record(s) grouped into episodes. 1 records not yet grouped.
#> Episode or recurrence window 11 .
#> 1 of 1 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 11 record(s) assinged a unique ID. 

left_join(infections, epids, by=c("rd_id"="sn"))

#> # A tibble: 11 x 9
#>    rd_id date       infection epi_len  epid case_nm epid_length epid_total
#>    <int> <date>     <chr>       <dbl> <dbl> <chr>   <time>           <int>
#>  1     1 2018-04-01 BSI            15     1 Case    0 hours              1
#>  2     2 2018-04-07 UTI            15     2 Case    0 hours              1
#>  3     3 2018-04-13 UTI            15     3 Case    0 hours              1
#>  4     4 2018-04-19 UTI            15     4 Case    0 hours              1
#>  5     5 2018-04-25 BSI            15     5 Case    0 hours              1
#>  6     6 2018-05-01 UTI            15     6 Case    0 hours              1
#>  7     7 2018-05-07 BSI            15     7 Case    0 hours              1
#>  8     8 2018-05-13 BSI            15     8 Case    0 hours              1
#>  9     9 2018-05-19 RTI            15     9 Case    0 hours              1
#> 10    10 2018-05-25 RTI            15    10 Case    0 hours              1
#> 11    11 2018-05-31 BSI            15    11 Case    0 hours              1
#> # ... with 1 more variable: epid_interval <S4: Interval>

# One rolling episode per strata. Initial case_length of 16 days (difference of 15 days) and
# one recurrence period lasting 31 days (difference of 30 days)
infections$recur <- 30
epids <- episode_group(infections, date=date, case_length = epi_len, episode_type = "rolling",
recurrence_length = recur, episodes_max = 1, rolls_max = 1, display = FALSE, group_stats = TRUE)

#> Episode grouping complete - 3 record(s) assinged a unique ID. 

bind_cols(infections, epids)

#> # A tibble: 11 x 11
#>    rd_id date       infection epi_len recur    sn  epid case_nm epid_length
#>    <int> <date>     <chr>       <dbl> <dbl> <int> <dbl> <chr>   <time>     
#>  1     1 2018-04-01 BSI            15    30     1     1 Case    42 days    
#>  2     2 2018-04-07 UTI            15    30     2     1 Duplic~ 42 days    
#>  3     3 2018-04-13 UTI            15    30     3     1 Duplic~ 42 days    
#>  4     4 2018-04-19 UTI            15    30     4     1 Recurr~ 42 days    
#>  5     5 2018-04-25 BSI            15    30     5     1 Duplic~ 42 days    
#>  6     6 2018-05-01 UTI            15    30     6     1 Duplic~ 42 days    
#>  7     7 2018-05-07 BSI            15    30     7     1 Duplic~ 42 days    
#>  8     8 2018-05-13 BSI            15    30     8     1 Duplic~ 42 days    
#>  9     9 2018-05-19 RTI            15    30     9     9 Case     0 days    
#> 10    10 2018-05-25 RTI            15    30    10    10 Case     0 days    
#> 11    11 2018-05-31 BSI            15    30    11    11 Case     0 days    
#> # ... with 2 more variables: epid_total <int>, epid_interval <S4: Interval>

# User defined case assignment
# Preference for case assignment - UTI > BSI > RTI
infections$infX <- factor(infections$infection, levels = c("UTI","BSI","RTI"))

# Different case and recurrence lengths for different sources of infection
infections <- mutate(infections,
                     epi_len = case_when(
                       infection == "BSI" ~ 14,
                       infection == "UTI" ~ 30,
                       infection == "RTI" ~ 60
                     )
)

# n-day episodes beginning with the earliest record with the specified preference; UTI > BSI > RTI
epids <- episode_group(infections, rd_id, date=date, case_length = epi_len,
                       custom_sort = infX, group_stats = TRUE,  display = FALSE)

#> Episode grouping complete - 2 record(s) assinged a unique ID. 

bind_cols(infections, epids)

#> # A tibble: 11 x 12
#>    rd_id date       infection epi_len recur infX     sn  epid case_nm
#>    <int> <date>     <chr>       <dbl> <dbl> <fct> <int> <dbl> <chr>  
#>  1     1 2018-04-01 BSI            14    30 BSI       1     1 Case   
#>  2     2 2018-04-07 UTI            30    30 UTI       2     2 Case   
#>  3     3 2018-04-13 UTI            30    30 UTI       3     2 Duplic~
#>  4     4 2018-04-19 UTI            30    30 UTI       4     2 Duplic~
#>  5     5 2018-04-25 BSI            14    30 BSI       5     2 Duplic~
#>  6     6 2018-05-01 UTI            30    30 UTI       6     2 Duplic~
#>  7     7 2018-05-07 BSI            14    30 BSI       7     2 Duplic~
#>  8     8 2018-05-13 BSI            14    30 BSI       8     8 Case   
#>  9     9 2018-05-19 RTI            60    30 RTI       9     8 Duplic~
#> 10    10 2018-05-25 RTI            60    30 RTI      10     8 Duplic~
#> 11    11 2018-05-31 BSI            14    30 BSI      11    11 Case   
#> # ... with 3 more variables: epid_length <time>, epid_total <int>,
#> #   epid_interval <S4: Interval>

# Another preference - RTI > UTI, or  RTI > BSI, or earliest record
infections$infX <- ifelse(infections$infection =="RTI",0,1)
epids <- episode_group(infections, rd_id, date=date, case_length = epi_len,
custom_sort = infX, from_last = TRUE, bi_direction = TRUE, display = FALSE, group_stats = TRUE)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

bind_cols(infections, epids)

#> # A tibble: 11 x 12
#>    rd_id date       infection epi_len recur  infX    sn  epid case_nm
#>    <int> <date>     <chr>       <dbl> <dbl> <dbl> <int> <int> <chr>  
#>  1     1 2018-04-01 BSI            14    30     1     1    10 Duplic~
#>  2     2 2018-04-07 UTI            30    30     1     2    10 Duplic~
#>  3     3 2018-04-13 UTI            30    30     1     3    10 Duplic~
#>  4     4 2018-04-19 UTI            30    30     1     4    10 Duplic~
#>  5     5 2018-04-25 BSI            14    30     1     5    10 Duplic~
#>  6     6 2018-05-01 UTI            30    30     1     6    10 Duplic~
#>  7     7 2018-05-07 BSI            14    30     1     7    10 Duplic~
#>  8     8 2018-05-13 BSI            14    30     1     8    10 Duplic~
#>  9     9 2018-05-19 RTI            60    30     0     9    10 Duplic~
#> 10    10 2018-05-25 RTI            60    30     0    10    10 Case   
#> 11    11 2018-05-31 BSI            14    30     1    11    10 Duplic~
#> # ... with 3 more variables: epid_length <time>, epid_total <int>,
#> #   epid_interval <S4: Interval>

# Stratified episode grouping
infections$patient_id <- c(rep("PID 1",8), rep("PID 2",3))

# Only three 9-day (difference of 8 days) rolling episodes per patient and infection.
infections$epi_len <- 8
epids <- episode_group(infections, rd_id, date=date, strata = c(patient_id, infection),
case_length = epi_len, episode_type = "rolling", recurrence_length = recur, episodes_max = 3,
data_source = c(patient_id, infection), display = FALSE)

#> Episode grouping complete - 3 record(s) assinged a unique ID. 

bind_cols(infections, epids)

#> # A tibble: 11 x 11
#>    rd_id date       infection epi_len recur  infX patient_id    sn  epid case_nm
#>    <int> <date>     <chr>       <dbl> <dbl> <dbl> <chr>      <int> <dbl> <chr>  
#>  1     1 2018-04-01 BSI             8    30     1 PID 1          1     1 Case   
#>  2     2 2018-04-07 UTI             8    30     1 PID 1          2     2 Case   
#>  3     3 2018-04-13 UTI             8    30     1 PID 1          3     2 Duplic~
#>  4     4 2018-04-19 UTI             8    30     1 PID 1          4     2 Recurr~
#>  5     5 2018-04-25 BSI             8    30     1 PID 1          5     5 Case   
#>  6     6 2018-05-01 UTI             8    30     1 PID 1          6     2 Duplic~
#>  7     7 2018-05-07 BSI             8    30     1 PID 1          7     7 Case   
#>  8     8 2018-05-13 BSI             8    30     1 PID 1          8     7 Duplic~
#>  9     9 2018-05-19 RTI             8    30     0 PID 2          9     9 Case   
#> 10    10 2018-05-25 RTI             8    30     0 PID 2         10     9 Duplic~
#> 11    11 2018-05-31 BSI             8    30     1 PID 2         11    11 Case   
#> # ... with 1 more variable: epid_dataset <chr>

# Interval grouping
data(hospital_admissions); hospital_admissions

#> # A tibble: 9 x 4
#>   rd_id admin_dt   discharge_dt epi_len
#>   <int> <date>     <date>         <dbl>
#> 1     1 2019-01-01 2019-01-01         0
#> 2     2 2019-01-01 2019-01-10         0
#> 3     3 2019-01-10 2019-01-13         0
#> 4     4 2019-01-05 2019-01-06         0
#> 5     5 2019-01-05 2019-01-15         0
#> 6     6 2019-01-07 2019-01-15         0
#> 7     7 2019-01-04 2019-01-13         0
#> 8     8 2019-01-20 2019-01-30         0
#> 9     9 2019-01-26 2019-01-31         0

hospital_admissions$admin_period <- interval(hospital_admissions$admin_dt,
hospital_admissions$discharge_dt)
hospital_admissions <- select(hospital_admissions, -c(discharge_dt, admin_dt))

# Episodes of overlaping intervals of admission
epids <- episode_group(hospital_admissions, date=admin_period,
sn=rd_id, case_length = epi_len, group_stats = TRUE)

#> Episode or recurrence window 1 .
#> 7 of 9 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 2 .
#> 2 of 2 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

bind_cols(hospital_admissions, epids)

#> # A tibble: 9 x 9
#>   rd_id epi_len admin_period                      sn  epid case_nm epid_length
#>   <int>   <dbl> <S4: Interval>                 <int> <dbl> <chr>   <time>     
#> 1     1       0 2019-01-01 UTC--2019-01-01 UTC     1     2 Duplic~ 14 days    
#> 2     2       0 2019-01-01 UTC--2019-01-10 UTC     2     2 Case    14 days    
#> 3     3       0 2019-01-10 UTC--2019-01-13 UTC     3     2 Duplic~ 14 days    
#> 4     4       0 2019-01-05 UTC--2019-01-06 UTC     4     2 Duplic~ 14 days    
#> 5     5       0 2019-01-05 UTC--2019-01-15 UTC     5     2 Duplic~ 14 days    
#> 6     6       0 2019-01-07 UTC--2019-01-15 UTC     6     2 Duplic~ 14 days    
#> 7     7       0 2019-01-04 UTC--2019-01-13 UTC     7     2 Duplic~ 14 days    
#> 8     8       0 2019-01-20 UTC--2019-01-30 UTC     8     8 Case    11 days    
#> 9     9       0 2019-01-26 UTC--2019-01-31 UTC     9     8 Duplic~ 11 days    
#> # ... with 2 more variables: epid_total <int>, epid_interval <S4: Interval>

# Overlaping intervals of admission seperated by 1 month
hospital_admissions$epi_len <- 1
epids <- episode_group(hospital_admissions, date=admin_period, sn=rd_id,
case_length = epi_len, episode_unit = "months")

#> Episode or recurrence window 1 .
#> 9 of 9 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

bind_cols(hospital_admissions, epids)

#> # A tibble: 9 x 6
#>   rd_id epi_len admin_period                      sn  epid case_nm  
#>   <int>   <dbl> <S4: Interval>                 <int> <int> <chr>    
#> 1     1       1 2019-01-01 UTC--2019-01-01 UTC     1     2 Duplicate
#> 2     2       1 2019-01-01 UTC--2019-01-10 UTC     2     2 Case     
#> 3     3       1 2019-01-10 UTC--2019-01-13 UTC     3     2 Duplicate
#> 4     4       1 2019-01-05 UTC--2019-01-06 UTC     4     2 Duplicate
#> 5     5       1 2019-01-05 UTC--2019-01-15 UTC     5     2 Duplicate
#> 6     6       1 2019-01-07 UTC--2019-01-15 UTC     6     2 Duplicate
#> 7     7       1 2019-01-04 UTC--2019-01-13 UTC     7     2 Duplicate
#> 8     8       1 2019-01-20 UTC--2019-01-30 UTC     8     2 Duplicate
#> 9     9       1 2019-01-26 UTC--2019-01-31 UTC     9     2 Duplicate

# Episodes of chained intervals, and those with aligned end periods
hospital_admissions$epi_len <- 0
epids <- episode_group(hospital_admissions, date=admin_period, sn=rd_id,
case_length = epi_len, overlap_method = c("chain","aligns_end"))

#> Episode or recurrence window 1 .
#> 3 of 9 record(s) grouped into episodes. 6 records not yet grouped.
#> Episode or recurrence window 2 .
#> 1 of 6 record(s) grouped into episodes. 5 records not yet grouped.
#> Episode or recurrence window 3 .
#> 2 of 5 record(s) grouped into episodes. 3 records not yet grouped.
#> Episode or recurrence window 4 .
#> 1 of 3 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 5 .
#> 1 of 2 record(s) grouped into episodes. 1 records not yet grouped.
#> Episode or recurrence window 6 .
#> 1 of 1 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 4 record(s) assinged a unique ID. 

bind_cols(hospital_admissions, epids)

#> # A tibble: 9 x 6
#>   rd_id epi_len admin_period                      sn  epid case_nm  
#>   <int>   <dbl> <S4: Interval>                 <int> <dbl> <chr>    
#> 1     1       0 2019-01-01 UTC--2019-01-01 UTC     1     2 Duplicate
#> 2     2       0 2019-01-01 UTC--2019-01-10 UTC     2     2 Case     
#> 3     3       0 2019-01-10 UTC--2019-01-13 UTC     3     2 Duplicate
#> 4     4       0 2019-01-05 UTC--2019-01-06 UTC     4     4 Case     
#> 5     5       0 2019-01-05 UTC--2019-01-15 UTC     5     5 Case     
#> 6     6       0 2019-01-07 UTC--2019-01-15 UTC     6     5 Duplicate
#> 7     7       0 2019-01-04 UTC--2019-01-13 UTC     7     7 Case     
#> 8     8       0 2019-01-20 UTC--2019-01-30 UTC     8     8 Case     
#> 9     9       0 2019-01-26 UTC--2019-01-31 UTC     9     9 Case