Episode grouping for case definitions and record deduplication

Group events into chronological episodes

episode_group_old(df, sn = NULL, strata = NULL, date, case_length,
  episode_type = "fixed", episode_unit = "days", episodes_max = Inf,
  recurrence_length = NULL, rolls_max = Inf, data_source = NULL,
  data_links = "ANY", custom_sort = NULL, skip_order = NULL,
  from_last = FALSE, overlap_method = c("exact", "across", "inbetween",
  "aligns_start", "aligns_end", "chain"), overlap_methods = NULL,
  bi_direction = FALSE, group_stats = FALSE, display = TRUE,
  deduplicate = FALSE, to_s4 = TRUE, recurrence_from_last = TRUE,
  case_for_recurrence = FALSE)

episode_group(df, sn = NULL, strata = NULL, date, case_length,
  episode_type = "fixed", episode_unit = "days", episodes_max = Inf,
  recurrence_length = NULL, rolls_max = Inf,
  skip_if_b4_lengths = TRUE, data_source = NULL, data_links = "ANY",
  custom_sort = NULL, skip_order = NULL, from_last = FALSE,
  overlap_method = c("exact", "across", "inbetween", "aligns_start",
  "aligns_end", "chain"), overlap_methods = NULL, bi_direction = FALSE,
  group_stats = FALSE, display = TRUE, deduplicate = FALSE,
  to_s4 = TRUE, recurrence_from_last = TRUE,
  case_for_recurrence = FALSE)

fixed_episodes(date, sn = NULL, strata = NULL, case_length,
  episode_unit = "days", episodes_max = Inf,
  skip_if_b4_lengths = TRUE, data_source = NULL, data_links = "ANY",
  custom_sort = NULL, skip_order = NULL, from_last = FALSE,
  overlap_method = c("exact", "across", "inbetween", "aligns_start",
  "aligns_end", "chain"),
  overlap_methods = "exact|across|chain|aligns_start|aligns_end|inbetween",
  bi_direction = FALSE, group_stats = FALSE, display = TRUE,
  deduplicate = FALSE, x, to_s4 = TRUE)

rolling_episodes(date, sn = NULL, strata = NULL, case_length,
  recurrence_length = NULL, episode_unit = "days",
  episodes_max = Inf, rolls_max = Inf, skip_if_b4_lengths = TRUE,
  data_source = NULL, data_links = "ANY", custom_sort = NULL,
  skip_order = NULL, from_last = FALSE, overlap_method = c("exact",
  "across", "inbetween", "aligns_start", "aligns_end", "chain"),
  overlap_methods = "exact|across|chain|aligns_start|aligns_end|inbetween",
  bi_direction = FALSE, group_stats = FALSE, display = TRUE,
  deduplicate = FALSE, x, to_s4 = TRUE, recurrence_from_last = TRUE,
  case_for_recurrence = FALSE)

Arguments

df
data.frame. One or more datasets appended together.

sn
Unique numerical record identifier. Optional.

strata
Subsets of the dataset. Episode grouping will be done separately within each subset of the dataset. In episode_group, you can use multiple columns. record_group can create useful strata e.g. patient identifiers.

date
Date (date, datetime or numeric) or period (number_line) of events.

case_length
Duration after a "case" within which subsequent events are considered "duplicate" events. This period is referred to as the the case window.

episode_type
"fixed" or "rolling".

episode_unit
Time units. Options are "seconds", "minutes", "hours", "days", "weeks", "months" or "years". See diyar::episode_unit.

episodes_max
Maximum number of episodes to have within each strata.

recurrence_length
Duration after the last or first event (see recurrence_from_last) of the previous window within which subsequent events are considered "recurrent" events. This period is referred to as the recurrence window. If recurrence_length is not supplied, case_length is used as the recurrence_length.

rolls_max
Maximum number of times an episode can reoccur. Only used if episode_type is "rolling".

data_source
Unique dataset identifier. Useful when the dataset contains data from multiple sources. In episode_group, you can use multiple columns supplied as column names.

data_links
Ungroup episodes that will not include records from certain data_sources. data_links should be a list with every element named 'l' (links) or 'g' (groups). Useful in skipping episodes that are not required and so minimises the overall processing time. Ignored if data_source is NULL.

custom_sort
Preferential order for "case" assignment. Useful in specifying that episode grouping begins at particular events regardless of chronological order. In episode_group, you can use multiple columns as sort levels.

skip_order
Skip episodes whose case events have custom_sort values that are less than or equal to the "nth" order of custom_sort. Useful in skipping episodes that are not required and so minimises the overall processing time. Ignored if custom_sort is NULL.

from_last
If TRUE, episode grouping will be backwards in time - starting at the most recent event and proceeding to the earliest. If FALSE, it'll be forward in time - starting at the earliest event and proceeding to the most recent one.

overlap_method
Methods of overlap considered when grouping event periods. Each pair of periods are checked with the same set of overlap_method. Deprecated use overlap_methods instead.

overlap_methods
Methods of overlap considered when grouping event periods. Different pairs of periods can be checked with different sets overlap_methods

bi_direction
If FALSE, "duplicate" events will be those within the case_length before or after the "case" as determined by from_last. If TRUE, "duplicate" events will be those within the same period before and after the "case".

group_stats
If TRUE, the output will include additional information with useful stats for each episode group.

display
If TRUE, a progress message is printed on screen.

deduplicate
if TRUE, "dupilcate" records are excluded from the output.

to_s4
if TRUE, changes the returned output to an epid object.

recurrence_from_last
if TRUE, the reference event for a recurrence window will be the last event from the previous window. If FALSE (default), it will be from the first event. Only used if episode_type is "rolling".

case_for_recurrence
if TRUE, both case and recurrence events will have a case window. If FALSE (default), only case events will have a case window. Only used if episode_type is "rolling".

skip_if_b4_lengths
if TRUE (default), records events before the case_length or the recurrence_length range are skipped.

x
Record date or interval. Deprecated. Please use date

Value

epid objects or data.frame if to_s4 is FALSE)

sn - unique record identifier as provided (or generated)

epid | .Data - unique episode identifier

wind_id - unique window identifier

wind_nm - type of window i.e. "Case" or "Recurrence"

case_nm - record type in regards to case assignment

epid_dataset - data sources in each episode

epid_interval - episode start and end dates. A number_line object.

epid_length - difference between episode start and end dates (difftime). If possible, it's the same unit as episode_unit otherwise, a difference in days is returned

epid_total - number of records in each episode

Details

Episode grouping begins at a reference event ("case") and proceeds forward or backward in time depending on from_last. If custom_sort is used, episode grouping can be forced to begin at certain events before proceeding forward or backwards in time. The maximum duration of a "fixed" episode is the case_length. This period is referred to as the case window. The maximum duration of a "rolling" episode is the case_length plus all recurrence periods. The recurrence periods are referred to as recurrence windows. This is a specified duration (recurrence_length) after the last or first (depending on recurrence_from_last) event from the previous window. Events within this period are considered "recurrent" events.

When a data_source identifier is provided, epid_dataset is included in the output. This lists the source of every event in each episode.

fixed_episodes() and rolling_episodes() are wrapper functions of episode_group(). They are convenient alternatives with the same functionalities.

See also

Examples

library(diyar)

#1. Fixed episodes
data(infections); infections

#> # A tibble: 11 x 4
#>    rd_id date       infection epi_len
#>    <int> <date>     <chr>       <dbl>
#>  1     1 2018-04-01 BSI            15
#>  2     2 2018-04-07 UTI            15
#>  3     3 2018-04-13 UTI            15
#>  4     4 2018-04-19 UTI            15
#>  5     5 2018-04-25 BSI            15
#>  6     6 2018-05-01 UTI            15
#>  7     7 2018-05-07 BSI            15
#>  8     8 2018-05-13 BSI            15
#>  9     9 2018-05-19 RTI            15
#> 10    10 2018-05-25 RTI            15
#> 11    11 2018-05-31 BSI            15

db_1 <- infections
# 16-day (difference of 15 days) episodes beginning from the earliest record
db_1$fd <- fixed_episodes(db_1$date, case_length = 15, display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

# 16-hour (difference of 15 hours) episodes beginning from the earliest record
db_1$fh <- fixed_episodes(db_1$date, case_length = 15,
episode_unit = "hours", display = FALSE)

#> Episode grouping complete - 11 record(s) with unique IDs. 

db_1

#> # A tibble: 11 x 6
#>    rd_id date       infection epi_len fd       fh      
#>    <int> <date>     <chr>       <dbl> <epid>   <epid>  
#>  1     1 2018-04-01 BSI            15 E.01 (C) E.01 (C)
#>  2     2 2018-04-07 UTI            15 E.01 (D) E.02 (C)
#>  3     3 2018-04-13 UTI            15 E.01 (D) E.03 (C)
#>  4     4 2018-04-19 UTI            15 E.04 (C) E.04 (C)
#>  5     5 2018-04-25 BSI            15 E.04 (D) E.05 (C)
#>  6     6 2018-05-01 UTI            15 E.04 (D) E.06 (C)
#>  7     7 2018-05-07 BSI            15 E.07 (C) E.07 (C)
#>  8     8 2018-05-13 BSI            15 E.07 (D) E.08 (C)
#>  9     9 2018-05-19 RTI            15 E.07 (D) E.09 (C)
#> 10    10 2018-05-25 RTI            15 E.10 (C) E.10 (C)
#> 11    11 2018-05-31 BSI            15 E.10 (D) E.11 (C)

#2. Rolling episodes
# Case length and recurrence periods of 16 days
db_1$rd_a <- rolling_episodes(db_1$date, case_length = 15, display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

# Case length of 16 days and recurrence periods of 11 days
db_1$rd_b <- rolling_episodes(db_1$date, case_length = 15,
recurrence_length = 10, display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

# Case length of 16 days and 2 recurrence periods of 11 days
db_1$rd_c <- rolling_episodes(db_1$date, case_length = 15,
recurrence_length = 10, rolls_max = 2, display = FALSE)

#> Episode grouping complete - 1 record(s) with unique IDs. 

db_1

#> # A tibble: 11 x 9
#>    rd_id date       infection epi_len fd       fh       rd_a    rd_b    rd_c    
#>    <int> <date>     <chr>       <dbl> <epid>   <epid>   <epid>  <epid>  <epid>  
#>  1     1 2018-04-01 BSI            15 E.01 (C) E.01 (C) E.1 (C) E.1 (C) E.01 (C)
#>  2     2 2018-04-07 UTI            15 E.01 (D) E.02 (C) E.1 (D) E.1 (D) E.01 (D)
#>  3     3 2018-04-13 UTI            15 E.01 (D) E.03 (C) E.1 (D) E.1 (D) E.01 (D)
#>  4     4 2018-04-19 UTI            15 E.04 (C) E.04 (C) E.1 (R) E.1 (R) E.01 (R)
#>  5     5 2018-04-25 BSI            15 E.04 (D) E.05 (C) E.1 (D) E.1 (R) E.01 (R)
#>  6     6 2018-05-01 UTI            15 E.04 (D) E.06 (C) E.1 (R) E.1 (R) E.06 (C)
#>  7     7 2018-05-07 BSI            15 E.07 (C) E.07 (C) E.1 (D) E.1 (R) E.06 (D)
#>  8     8 2018-05-13 BSI            15 E.07 (D) E.08 (C) E.1 (R) E.1 (R) E.06 (D)
#>  9     9 2018-05-19 RTI            15 E.07 (D) E.09 (C) E.1 (D) E.1 (R) E.06 (R)
#> 10    10 2018-05-25 RTI            15 E.10 (C) E.10 (C) E.1 (R) E.1 (R) E.06 (R)
#> 11    11 2018-05-31 BSI            15 E.10 (D) E.11 (C) E.1 (D) E.1 (R) E.11 (C)

# 3. Stratified episode grouping
db_3 <- infections

db_3$patient_id <- c(rep("PID 1",8), rep("PID 2",3))
# One 16-day episode per patient
db_3$epids_p <- fixed_episodes(date=db_3$date, strata = db_3$patient_id,
case_length = 15, episodes_max = 1, display = FALSE)

#> 5 record(s); 5 skipped
#> Episode grouping complete - 5 record(s) with unique IDs. 

db_3

#> # A tibble: 11 x 6
#>    rd_id date       infection epi_len patient_id epids_p
#>    <int> <date>     <chr>       <dbl> <chr>      <epid> 
#>  1     1 2018-04-01 BSI            15 PID 1      E.1 (C)
#>  2     2 2018-04-07 UTI            15 PID 1      E.1 (D)
#>  3     3 2018-04-13 UTI            15 PID 1      E.1 (D)
#>  4     4 2018-04-19 UTI            15 PID 1      E.4 (S)
#>  5     5 2018-04-25 BSI            15 PID 1      E.5 (S)
#>  6     6 2018-05-01 UTI            15 PID 1      E.6 (S)
#>  7     7 2018-05-07 BSI            15 PID 1      E.7 (S)
#>  8     8 2018-05-13 BSI            15 PID 1      E.8 (S)
#>  9     9 2018-05-19 RTI            15 PID 2      E.9 (C)
#> 10    10 2018-05-25 RTI            15 PID 2      E.9 (D)
#> 11    11 2018-05-31 BSI            15 PID 2      E.9 (D)

# 4. Case assignment
db_4 <- infections

## 4.1 Chronological order
db_4$forward_time <- fixed_episodes(db_4$date, case_length = 1,
episode_unit = "months", display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

db_4$backward_time <- fixed_episodes(db_4$date, case_length = 1,
episode_unit = "months", from_last = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

db_4

#> # A tibble: 11 x 6
#>    rd_id date       infection epi_len forward_time backward_time
#>    <int> <date>     <chr>       <dbl> <epid>       <epid>       
#>  1     1 2018-04-01 BSI            15 E.1 (C)      E.05 (D)     
#>  2     2 2018-04-07 UTI            15 E.1 (D)      E.05 (D)     
#>  3     3 2018-04-13 UTI            15 E.1 (D)      E.05 (D)     
#>  4     4 2018-04-19 UTI            15 E.1 (D)      E.05 (D)     
#>  5     5 2018-04-25 BSI            15 E.1 (D)      E.05 (C)     
#>  6     6 2018-05-01 UTI            15 E.1 (D)      E.11 (D)     
#>  7     7 2018-05-07 BSI            15 E.7 (C)      E.11 (D)     
#>  8     8 2018-05-13 BSI            15 E.7 (D)      E.11 (D)     
#>  9     9 2018-05-19 RTI            15 E.7 (D)      E.11 (D)     
#> 10    10 2018-05-25 RTI            15 E.7 (D)      E.11 (D)     
#> 11    11 2018-05-31 BSI            15 E.7 (D)      E.11 (C)     

## 4.2 User defined order
db_4b <- infections
db_4b

#> # A tibble: 11 x 4
#>    rd_id date       infection epi_len
#>    <int> <date>     <chr>       <dbl>
#>  1     1 2018-04-01 BSI            15
#>  2     2 2018-04-07 UTI            15
#>  3     3 2018-04-13 UTI            15
#>  4     4 2018-04-19 UTI            15
#>  5     5 2018-04-25 BSI            15
#>  6     6 2018-05-01 UTI            15
#>  7     7 2018-05-07 BSI            15
#>  8     8 2018-05-13 BSI            15
#>  9     9 2018-05-19 RTI            15
#> 10    10 2018-05-25 RTI            15
#> 11    11 2018-05-31 BSI            15

# RTI > UTI, or RTI > BSI
db_4b$ord1 <- ifelse(db_4b$infection =="RTI",0,1)
# UTI > BSI > RTI
db_4b$ord2 <- factor(db_4b$infection, levels = c("UTI","BSI","RTI"))

db_4b$epids_1 <- fixed_episodes(db_4b$date, case_length = 15,
custom_sort = db_4b$ord1, display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

db_4b$epids_2 <- fixed_episodes(db_4b$date, case_length = 15,
custom_sort = db_4b$ord2, display = FALSE)

#> Episode grouping complete - 3 record(s) with unique IDs. 

db_4b$epids_2b <- fixed_episodes(db_4b$date, case_length = 15,
custom_sort = db_4b$ord2, bi_direction = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

db_4b

#> # A tibble: 11 x 9
#>    rd_id date       infection epi_len  ord1 ord2  epids_1 epids_2  epids_2b
#>    <int> <date>     <chr>       <dbl> <dbl> <fct> <epid>  <epid>   <epid>  
#>  1     1 2018-04-01 BSI            15     1 BSI   E.1 (C) E.01 (C) E.02 (D)
#>  2     2 2018-04-07 UTI            15     1 UTI   E.1 (D) E.02 (C) E.02 (C)
#>  3     3 2018-04-13 UTI            15     1 UTI   E.1 (D) E.02 (D) E.02 (D)
#>  4     4 2018-04-19 UTI            15     1 UTI   E.4 (C) E.02 (D) E.02 (D)
#>  5     5 2018-04-25 BSI            15     1 BSI   E.4 (D) E.05 (C) E.06 (D)
#>  6     6 2018-05-01 UTI            15     1 UTI   E.4 (D) E.06 (C) E.06 (C)
#>  7     7 2018-05-07 BSI            15     1 BSI   E.7 (C) E.06 (D) E.06 (D)
#>  8     8 2018-05-13 BSI            15     1 BSI   E.7 (D) E.06 (D) E.06 (D)
#>  9     9 2018-05-19 RTI            15     0 RTI   E.9 (C) E.09 (C) E.11 (D)
#> 10    10 2018-05-25 RTI            15     0 RTI   E.9 (D) E.09 (D) E.11 (D)
#> 11    11 2018-05-31 BSI            15     1 BSI   E.9 (D) E.11 (C) E.11 (C)

#5. Interval grouping
data(hospital_admissions)

hospital_admissions$admin_period <- number_line(hospital_admissions$admin_dt,
hospital_admissions$discharge_dt)
admissions <- hospital_admissions[c("admin_period","epi_len")]
admissions

#> # A tibble: 11 x 2
#>    admin_period             epi_len
#>    <numbr_ln>                 <dbl>
#>  1 2019-01-01 == 2019-01-01       0
#>  2 2019-01-01 -> 2019-01-10       0
#>  3 2019-01-10 -> 2019-01-13       0
#>  4 2019-01-05 -> 2019-01-06       0
#>  5 2019-01-05 -> 2019-01-15       0
#>  6 2019-01-07 -> 2019-01-15       0
#>  7 2019-01-04 -> 2019-01-13       0
#>  8 2019-01-20 -> 2019-01-30       0
#>  9 2019-01-26 -> 2019-01-31       0
#> 10 2019-01-01 -> 2019-01-10       0
#> 11 2019-01-20 -> 2019-01-30       0

# Episodes of overlaping periods of admission
admissions$epi_0 <- fixed_episodes(date=admissions$admin_period, case_length = 0,
group_stats = TRUE, to_s4=TRUE)

#> Episode or recurrence window 1.
#> 11 record(s); 8 grouped to episodes and 3 left to group.
#> Episode or recurrence window 2.
#> 3 record(s); 3 grouped to episodes and 0 left to group.
#> 
#> Episode grouping complete - 0 record(s) with unique IDs. 

admissions

#> # A tibble: 11 x 3
#>    admin_period             epi_len epi_0                           
#>    <numbr_ln>                 <dbl> <epid>                          
#>  1 2019-01-01 == 2019-01-01       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  2 2019-01-01 -> 2019-01-10       0 E.2 2019-01-01 -> 2019-01-15 (C)
#>  3 2019-01-10 -> 2019-01-13       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  4 2019-01-05 -> 2019-01-06       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  5 2019-01-05 -> 2019-01-15       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  6 2019-01-07 -> 2019-01-15       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  7 2019-01-04 -> 2019-01-13       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  8 2019-01-20 -> 2019-01-30       0 E.8 2019-01-20 -> 2019-01-31 (C)
#>  9 2019-01-26 -> 2019-01-31       0 E.8 2019-01-20 -> 2019-01-31 (D)
#> 10 2019-01-01 -> 2019-01-10       0 E.2 2019-01-01 -> 2019-01-15 (D)
#> 11 2019-01-20 -> 2019-01-30       0 E.8 2019-01-20 -> 2019-01-31 (D)

# Overlaping periods of admission seperated by 1 month
admissions$epi_1 <- fixed_episodes(date=admissions$admin_period, case_length = 1,
episode_unit = "months", group_stats = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

admissions

#> # A tibble: 11 x 4
#>    admin_period             epi_len epi_0                           
#>    <numbr_ln>                 <dbl> <epid>                          
#>  1 2019-01-01 == 2019-01-01       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  2 2019-01-01 -> 2019-01-10       0 E.2 2019-01-01 -> 2019-01-15 (C)
#>  3 2019-01-10 -> 2019-01-13       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  4 2019-01-05 -> 2019-01-06       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  5 2019-01-05 -> 2019-01-15       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  6 2019-01-07 -> 2019-01-15       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  7 2019-01-04 -> 2019-01-13       0 E.2 2019-01-01 -> 2019-01-15 (D)
#>  8 2019-01-20 -> 2019-01-30       0 E.8 2019-01-20 -> 2019-01-31 (C)
#>  9 2019-01-26 -> 2019-01-31       0 E.8 2019-01-20 -> 2019-01-31 (D)
#> 10 2019-01-01 -> 2019-01-10       0 E.2 2019-01-01 -> 2019-01-15 (D)
#> 11 2019-01-20 -> 2019-01-30       0 E.8 2019-01-20 -> 2019-01-31 (D)
#> # ... with 1 more variable: epi_1 <epid>

# Episodes of chained admission periods, and those with aligned end periods
admissions$epi_0b <- fixed_episodes(date=admissions$admin_period, case_length = 0,
overlap_methods = c("chain|aligns_end"), group_stats = TRUE, display = FALSE)

#> Episode grouping complete - 7 record(s) with unique IDs. 

admissions["epi_0b"]

#> # A tibble: 11 x 1
#>    epi_0b                           
#>    <epid>                           
#>  1 E.01 2019-01-01 == 2019-01-01 (C)
#>  2 E.02 2019-01-01 -> 2019-01-13 (C)
#>  3 E.02 2019-01-01 -> 2019-01-13 (D)
#>  4 E.04 2019-01-05 -> 2019-01-06 (C)
#>  5 E.05 2019-01-05 -> 2019-01-15 (C)
#>  6 E.05 2019-01-05 -> 2019-01-15 (D)
#>  7 E.07 2019-01-04 -> 2019-01-13 (C)
#>  8 E.08 2019-01-20 -> 2019-01-30 (C)
#>  9 E.09 2019-01-26 -> 2019-01-31 (C)
#> 10 E.10 2019-01-01 -> 2019-01-10 (C)
#> 11 E.11 2019-01-20 -> 2019-01-30 (C)

# Note - episode_group() takes column names not actual values
db_5 <- infections

db_5$recur <- 20
db_5$epids_f <- episode_group(db_5, date=date, episode_type = "fixed",
case_length = epi_len, display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

db_5$epids_r <- episode_group(db_5, date=date, episode_type = "rolling",
case_length = epi_len, recurrence_length = recur, display = FALSE)

#> Episode grouping complete - 0 record(s) with unique IDs. 

db_5

#> # A tibble: 11 x 7
#>    rd_id date       infection epi_len recur epids_f  epids_r
#>    <int> <date>     <chr>       <dbl> <dbl> <epid>   <epid> 
#>  1     1 2018-04-01 BSI            15    20 E.01 (C) E.1 (C)
#>  2     2 2018-04-07 UTI            15    20 E.01 (D) E.1 (D)
#>  3     3 2018-04-13 UTI            15    20 E.01 (D) E.1 (D)
#>  4     4 2018-04-19 UTI            15    20 E.04 (C) E.1 (R)
#>  5     5 2018-04-25 BSI            15    20 E.04 (D) E.1 (D)
#>  6     6 2018-05-01 UTI            15    20 E.04 (D) E.1 (D)
#>  7     7 2018-05-07 BSI            15    20 E.07 (C) E.1 (R)
#>  8     8 2018-05-13 BSI            15    20 E.07 (D) E.1 (D)
#>  9     9 2018-05-19 RTI            15    20 E.07 (D) E.1 (D)
#> 10    10 2018-05-25 RTI            15    20 E.10 (C) E.1 (R)
#> 11    11 2018-05-31 BSI            15    20 E.10 (D) E.1 (D)