Episode grouping for case definitions

This function assigns records into unique chronological episodes

episode_group(df, sn = NULL, strata = NULL, date, case_length,
  episode_type = "fixed", episode_unit = "days", episodes_max = Inf,
  recurrence_length = NULL, rolls_max = Inf, data_source = NULL,
  custom_sort = NULL, from_last = FALSE,
  overlap_method = c("overlap", "within", "aligns_start", "aligns_end",
  "chain"), bi_direction = FALSE, group_stats = FALSE,
  display = TRUE)

Arguments

df
Dataframe. One or more datasets appended together.

sn
numeric unique record indentifier for the dataframe.

strata
Column names. Episodes will be unique to each strata. See record_group.

date
Record date. date, datetime or interval objects.

case_length
Duration from the "Case" within which another record of the same strata will be considered a "Duplicate".

episode_type
"fixed" or "rolling".

episode_unit
Time units not less than "seconds" and supported by duration

episodes_max
Maximum number times to group episodes in each strata.

recurrence_length
Duration from the last record of an episode within which another record of the same strata will be considered a "Recurrent" record.

rolls_max
Maximum number of recurrence periods permitted within each episode. Only used if episode_type is "rolling".

data_source
Unique dataset indentifier for the dataframe. Useful when dataframe contains multiple datsets.

custom_sort
If TRUE, "Case" assignment will be in preference to this sort order. Useful in specifying that episode grouping begins at a particular kind of record regardless of chronological order.

from_last
If TRUE, episode grouping will be backwards in time - starting at the most recent record and proceeding to the earliest. If FALSE, it'll be forward in time - starting at the earliest record and proceeding to the most recent.

overlap_method
A set of options by which grouped intervals should overlap. Options are; "none" (default), "overlap", "aligns_start", "aligns_end", "within" and "chain"

bi_direction
If FALSE, "Duplicate" records are those within the case_length and recurrence_length before or after the "Case" as determined by from_last. If TRUE, "Duplicate" records are those on either side of the "Case".

group_stats
If TRUE, output will include additional columns with useful stats for each episode.

display
If TRUE, status messages are not printed on screen.

Value

Dataframe

sn - the unique record identifier provided

epid - unique episode indentifier

case_nm - record type in regards to case assignment

epid_dataset - datasets contained in each episode

epid_interval - lubridate interval object. Episode start and end dates

epid_length - difftime object. Difference between episode start and end dates. Unit for will match that supplied as episode_unit

epid_total - number of records in each record group

Details

Episode grouping begins at a reference record ("Case") and proceeds forward or backward in time depending on from_last. If custom_sort is used, episode grouping can be forced to begin at certain record before proceeding forward or backwards in time. The maximun duration of a "fixed" episode is the case_length while, the maximum duration of a "rolling" episode is the case_length in addition to all recurrence periods. A recurrence period is the recurrence_length from the last record in an episode

See also

Examples

library(lubridate)

#> Warning: package 'lubridate' was built under R version 3.5.1

#> 
#> Attaching package: 'lubridate'

#> The following object is masked from 'package:base':
#> 
#>     date

library(dplyr)

#> Warning: package 'dplyr' was built under R version 3.5.2

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:lubridate':
#> 
#>     intersect, setdiff, union

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

data <- data.frame(date = seq.Date(dmy("01/04/2018"), dmy("31/05/2018"), by="6 days"))
data$epi_len <- 15
data <- mutate(data, rd_id = row_number())

# Begin episode grouping at the earliest record thereby making it the "Case"
cbind(data,
      episode_group(data, sn=rd_id, date = date, case_length = epi_len)
      )

#> Episode or recurrence window 1 .
#> 3 of 11 record(s) grouped into episodes. 8 records not yet grouped.
#> Episode or recurrence window 2 .
#> 3 of 8 record(s) grouped into episodes. 5 records not yet grouped.
#> Episode or recurrence window 3 .
#> 3 of 5 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 4 .
#> 2 of 2 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#>          date epi_len rd_id sn epid   case_nm
#> 1  2018-04-01      15     1  1    1      Case
#> 2  2018-04-07      15     2  2    1 Duplicate
#> 3  2018-04-13      15     3  3    1 Duplicate
#> 4  2018-04-19      15     4  4    4      Case
#> 5  2018-04-25      15     5  5    4 Duplicate
#> 6  2018-05-01      15     6  6    4 Duplicate
#> 7  2018-05-07      15     7  7    7      Case
#> 8  2018-05-13      15     8  8    7 Duplicate
#> 9  2018-05-19      15     9  9    7 Duplicate
#> 10 2018-05-25      15    10 10   10      Case
#> 11 2018-05-31      15    11 11   10 Duplicate

# Begin episode grouping at the most recent record thereby making it the "Case"
cbind(data,
      episode_group(data, rd_id, date=date, case_length = epi_len, from_last = TRUE, display = FALSE)
      )

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#>          date epi_len rd_id sn epid   case_nm
#> 1  2018-04-01      15     1  1    2 Duplicate
#> 2  2018-04-07      15     2  2    2      Case
#> 3  2018-04-13      15     3  3    5 Duplicate
#> 4  2018-04-19      15     4  4    5 Duplicate
#> 5  2018-04-25      15     5  5    5      Case
#> 6  2018-05-01      15     6  6    8 Duplicate
#> 7  2018-05-07      15     7  7    8 Duplicate
#> 8  2018-05-13      15     8  8    8      Case
#> 9  2018-05-19      15     9  9   11 Duplicate
#> 10 2018-05-25      15    10 10   11 Duplicate
#> 11 2018-05-31      15    11 11   11      Case

# Rolling episodes
data$recur <- 30
cbind(data,
      episode_group(data, rd_id, date=date, case_length = epi_len, episode_type = "rolling",
                    recurrence_length = recur, display = FALSE)
      )

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#>          date epi_len rd_id recur sn epid   case_nm
#> 1  2018-04-01      15     1    30  1    1      Case
#> 2  2018-04-07      15     2    30  2    1 Duplicate
#> 3  2018-04-13      15     3    30  3    1 Duplicate
#> 4  2018-04-19      15     4    30  4    1 Recurrent
#> 5  2018-04-25      15     5    30  5    1 Duplicate
#> 6  2018-05-01      15     6    30  6    1 Duplicate
#> 7  2018-05-07      15     7    30  7    1 Duplicate
#> 8  2018-05-13      15     8    30  8    1 Duplicate
#> 9  2018-05-19      15     9    30  9    1 Recurrent
#> 10 2018-05-25      15    10    30 10    1 Duplicate
#> 11 2018-05-31      15    11    30 11    1 Duplicate

# Limit the number of recurrence periods
cbind(data,
      episode_group(data, rd_id, date=date, case_length = epi_len, episode_type = "rolling",
                    recurrence_length = recur, rolls_max = 1, display = FALSE)
      )

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#>          date epi_len rd_id recur sn epid   case_nm
#> 1  2018-04-01      15     1    30  1    1      Case
#> 2  2018-04-07      15     2    30  2    1 Duplicate
#> 3  2018-04-13      15     3    30  3    1 Duplicate
#> 4  2018-04-19      15     4    30  4    1 Recurrent
#> 5  2018-04-25      15     5    30  5    1 Duplicate
#> 6  2018-05-01      15     6    30  6    1 Duplicate
#> 7  2018-05-07      15     7    30  7    1 Duplicate
#> 8  2018-05-13      15     8    30  8    1 Duplicate
#> 9  2018-05-19      15     9    30  9    9      Case
#> 10 2018-05-25      15    10    30 10    9 Duplicate
#> 11 2018-05-31      15    11    30 11    9 Duplicate

# Limit the number times episode grouping occurs
cbind(data,
      episode_group(data, rd_id, date=date, case_length = epi_len, episode_type = "fixed",
                    recurrence_length = recur, episodes_max = 1, from_last = FALSE, display = FALSE)
      )

#> Episode grouping complete - 8 record(s) assinged a unique ID. 

#>          date epi_len rd_id recur sn epid   case_nm
#> 1  2018-04-01      15     1    30  1    1      Case
#> 2  2018-04-07      15     2    30  2    1 Duplicate
#> 3  2018-04-13      15     3    30  3    1 Duplicate
#> 4  2018-04-19      15     4    30  4    4      Case
#> 5  2018-04-25      15     5    30  5    5      Case
#> 6  2018-05-01      15     6    30  6    6      Case
#> 7  2018-05-07      15     7    30  7    7      Case
#> 8  2018-05-13      15     8    30  8    8      Case
#> 9  2018-05-19      15     9    30  9    9      Case
#> 10 2018-05-25      15    10    30 10   10      Case
#> 11 2018-05-31      15    11    30 11   11      Case

# User defined case assignment
data$infection <- c("BSI", rep("UTI",2), "UTI", "BSI",  "UTI", rep("BSI",2), "RTI","RTI","BSI")

  # preference for case assignment - UTI > BSI > RTI
data$infection <- factor(data$infection, levels = c("UTI","BSI","RTI"))

  # Different case and recurrence lengths for different records
data <- mutate(data,
               epi_len = case_when(
                 infection == "BSI" ~ 14,
                 infection == "UTI" ~ 30,
                 infection == "RTI" ~ 60
                 )
               )

cbind(data,
      episode_group(data, rd_id, date=date, case_length = epi_len,
                    recurrence_length = recur, custom_sort = infection, from_last = FALSE, display = FALSE)
      )

#> Episode grouping complete - 2 record(s) assinged a unique ID. 

#>          date epi_len rd_id recur infection sn epid   case_nm
#> 1  2018-04-01      14     1    30       BSI  1    1      Case
#> 2  2018-04-07      30     2    30       UTI  2    2      Case
#> 3  2018-04-13      30     3    30       UTI  3    2 Duplicate
#> 4  2018-04-19      30     4    30       UTI  4    2 Duplicate
#> 5  2018-04-25      14     5    30       BSI  5    2 Duplicate
#> 6  2018-05-01      30     6    30       UTI  6    2 Duplicate
#> 7  2018-05-07      14     7    30       BSI  7    2 Duplicate
#> 8  2018-05-13      14     8    30       BSI  8    8      Case
#> 9  2018-05-19      60     9    30       RTI  9    8 Duplicate
#> 10 2018-05-25      60    10    30       RTI 10    8 Duplicate
#> 11 2018-05-31      14    11    30       BSI 11   11      Case

  # Can make it such that only combination with BSI records will have a specified preference for case assignment
  # When it's a combination without a "BSI" record, preference for case assignment reverts back to the earliest or most recent record depending on from_last
data$infection_ord <- ifelse(data$infection =="BTI",0,1)

cbind(data,
      episode_group(data, rd_id, date=date, case_length = epi_len,
                    custom_sort = infection_ord, group_stats = TRUE, from_last = FALSE, bi_direction = TRUE, display = FALSE)
)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#>          date epi_len rd_id recur infection infection_ord sn epid   case_nm
#> 1  2018-04-01      14     1    30       BSI             1  1    1      Case
#> 2  2018-04-07      30     2    30       UTI             1  2    1 Duplicate
#> 3  2018-04-13      30     3    30       UTI             1  3    1 Duplicate
#> 4  2018-04-19      30     4    30       UTI             1  4    4      Case
#> 5  2018-04-25      14     5    30       BSI             1  5    4 Duplicate
#> 6  2018-05-01      30     6    30       UTI             1  6    4 Duplicate
#> 7  2018-05-07      14     7    30       BSI             1  7    4 Duplicate
#> 8  2018-05-13      14     8    30       BSI             1  8    4 Duplicate
#> 9  2018-05-19      60     9    30       RTI             1  9    4 Duplicate
#> 10 2018-05-25      60    10    30       RTI             1 10   10      Case
#> 11 2018-05-31      14    11    30       BSI             1 11   10 Duplicate
#>                     epid_interval epid_length epid_total
#> 1  2018-04-01 UTC--2018-04-13 UTC     12 days          3
#> 2  2018-04-01 UTC--2018-04-13 UTC     12 days          3
#> 3  2018-04-01 UTC--2018-04-13 UTC     12 days          3
#> 4  2018-04-19 UTC--2018-05-19 UTC     30 days          6
#> 5  2018-04-19 UTC--2018-05-19 UTC     30 days          6
#> 6  2018-04-19 UTC--2018-05-19 UTC     30 days          6
#> 7  2018-04-19 UTC--2018-05-19 UTC     30 days          6
#> 8  2018-04-19 UTC--2018-05-19 UTC     30 days          6
#> 9  2018-04-19 UTC--2018-05-19 UTC     30 days          6
#> 10 2018-05-25 UTC--2018-05-31 UTC      6 days          2
#> 11 2018-05-25 UTC--2018-05-31 UTC      6 days          2

  # Can make it such that duplicates on either side of the case are grouped into the same episode
data$infection_ord <- ifelse(data$infection =="RTI",0,1)

cbind(data,
      episode_group(data, rd_id, date=date, case_length = epi_len,
                   custom_sort = infection_ord, group_stats = TRUE, from_last = FALSE, bi_direction = TRUE, display = FALSE)
      )

#> Episode grouping complete - 1 record(s) assinged a unique ID. 

#>          date epi_len rd_id recur infection infection_ord sn epid   case_nm
#> 1  2018-04-01      14     1    30       BSI             1  1    1      Case
#> 2  2018-04-07      30     2    30       UTI             1  2    9 Duplicate
#> 3  2018-04-13      30     3    30       UTI             1  3    9 Duplicate
#> 4  2018-04-19      30     4    30       UTI             1  4    9 Duplicate
#> 5  2018-04-25      14     5    30       BSI             1  5    9 Duplicate
#> 6  2018-05-01      30     6    30       UTI             1  6    9 Duplicate
#> 7  2018-05-07      14     7    30       BSI             1  7    9 Duplicate
#> 8  2018-05-13      14     8    30       BSI             1  8    9 Duplicate
#> 9  2018-05-19      60     9    30       RTI             0  9    9      Case
#> 10 2018-05-25      60    10    30       RTI             0 10    9 Duplicate
#> 11 2018-05-31      14    11    30       BSI             1 11    9 Duplicate
#>                     epid_interval epid_length epid_total
#> 1  2018-04-01 UTC--2018-04-01 UTC      0 days          1
#> 2  2018-04-07 UTC--2018-05-31 UTC     54 days         10
#> 3  2018-04-07 UTC--2018-05-31 UTC     54 days         10
#> 4  2018-04-07 UTC--2018-05-31 UTC     54 days         10
#> 5  2018-04-07 UTC--2018-05-31 UTC     54 days         10
#> 6  2018-04-07 UTC--2018-05-31 UTC     54 days         10
#> 7  2018-04-07 UTC--2018-05-31 UTC     54 days         10
#> 8  2018-04-07 UTC--2018-05-31 UTC     54 days         10
#> 9  2018-04-07 UTC--2018-05-31 UTC     54 days         10
#> 10 2018-04-07 UTC--2018-05-31 UTC     54 days         10
#> 11 2018-04-07 UTC--2018-05-31 UTC     54 days         10

# Stratified episode grouping
data$patient_id <- c(rep("PID 1",8), rep("PID 2",3))
cbind(data,
      episode_group(data, rd_id, date=date, strata = patient_id, case_length = epi_len,
                    recurrence_length = recur, episodes_max = 1, from_last = FALSE, display = FALSE)
      )

#> Episode grouping complete - 5 record(s) assinged a unique ID. 

#>          date epi_len rd_id recur infection infection_ord patient_id sn epid
#> 1  2018-04-01      14     1    30       BSI             1      PID 1  1    1
#> 2  2018-04-07      30     2    30       UTI             1      PID 1  2    1
#> 3  2018-04-13      30     3    30       UTI             1      PID 1  3    1
#> 4  2018-04-19      30     4    30       UTI             1      PID 1  4    4
#> 5  2018-04-25      14     5    30       BSI             1      PID 1  5    5
#> 6  2018-05-01      30     6    30       UTI             1      PID 1  6    6
#> 7  2018-05-07      14     7    30       BSI             1      PID 1  7    7
#> 8  2018-05-13      14     8    30       BSI             1      PID 1  8    8
#> 9  2018-05-19      60     9    30       RTI             0      PID 2  9    9
#> 10 2018-05-25      60    10    30       RTI             0      PID 2 10    9
#> 11 2018-05-31      14    11    30       BSI             1      PID 2 11    9
#>      case_nm
#> 1       Case
#> 2  Duplicate
#> 3  Duplicate
#> 4       Case
#> 5       Case
#> 6       Case
#> 7       Case
#> 8       Case
#> 9       Case
#> 10 Duplicate
#> 11 Duplicate

cbind(data,
      episode_group(data, rd_id, date=date, strata = c(patient_id, infection), case_length = epi_len,
                    recurrence_length = recur, episodes_max = 1, from_last = FALSE, display = FALSE)
      )

#> Episode grouping complete - 5 record(s) assinged a unique ID. 

#>          date epi_len rd_id recur infection infection_ord patient_id sn epid
#> 1  2018-04-01      14     1    30       BSI             1      PID 1  1    1
#> 2  2018-04-07      30     2    30       UTI             1      PID 1  2    2
#> 3  2018-04-13      30     3    30       UTI             1      PID 1  3    2
#> 4  2018-04-19      30     4    30       UTI             1      PID 1  4    2
#> 5  2018-04-25      14     5    30       BSI             1      PID 1  5    5
#> 6  2018-05-01      30     6    30       UTI             1      PID 1  6    2
#> 7  2018-05-07      14     7    30       BSI             1      PID 1  7    7
#> 8  2018-05-13      14     8    30       BSI             1      PID 1  8    8
#> 9  2018-05-19      60     9    30       RTI             0      PID 2  9    9
#> 10 2018-05-25      60    10    30       RTI             0      PID 2 10    9
#> 11 2018-05-31      14    11    30       BSI             1      PID 2 11   11
#>      case_nm
#> 1       Case
#> 2       Case
#> 3  Duplicate
#> 4  Duplicate
#> 5       Case
#> 6  Duplicate
#> 7       Case
#> 8       Case
#> 9       Case
#> 10 Duplicate
#> 11      Case

# data_source is useful when grouping episodes across multiple datasets or as diagnostic option
cbind(data,
      episode_group(data, rd_id, date=date, case_length = epi_len, episode_type = "fixed",
                    recurrence_length = recur, episodes_max = 1, data_source = c(patient_id, infection), display = FALSE)
)

#> Episode grouping complete - 8 record(s) assinged a unique ID. 

#>          date epi_len rd_id recur infection infection_ord patient_id sn epid
#> 1  2018-04-01      14     1    30       BSI             1      PID 1  1    1
#> 2  2018-04-07      30     2    30       UTI             1      PID 1  2    1
#> 3  2018-04-13      30     3    30       UTI             1      PID 1  3    1
#> 4  2018-04-19      30     4    30       UTI             1      PID 1  4    4
#> 5  2018-04-25      14     5    30       BSI             1      PID 1  5    5
#> 6  2018-05-01      30     6    30       UTI             1      PID 1  6    6
#> 7  2018-05-07      14     7    30       BSI             1      PID 1  7    7
#> 8  2018-05-13      14     8    30       BSI             1      PID 1  8    8
#> 9  2018-05-19      60     9    30       RTI             0      PID 2  9    9
#> 10 2018-05-25      60    10    30       RTI             0      PID 2 10   10
#> 11 2018-05-31      14    11    30       BSI             1      PID 2 11   11
#>      case_nm        epid_dataset
#> 1       Case PID 1-BSI,PID 1-UTI
#> 2  Duplicate PID 1-BSI,PID 1-UTI
#> 3  Duplicate PID 1-BSI,PID 1-UTI
#> 4       Case           PID 1-UTI
#> 5       Case           PID 1-BSI
#> 6       Case           PID 1-UTI
#> 7       Case           PID 1-BSI
#> 8       Case           PID 1-BSI
#> 9       Case           PID 2-RTI
#> 10      Case           PID 2-RTI
#> 11      Case           PID 2-BSI

hopital_record <- data.frame(
  rd_id = 1:8,
  admin_dt = c(dmy("01/01/2010"), dmy("01/01/2010"), dmy("05/01/2010"), dmy("05/01/2010"), dmy("07/01/2010"), dmy("04/01/2010"), dmy("20/01/2010"), dmy("26/01/2010")),
  discharge_dt = c(dmy("01/01/2010"), dmy("10/01/2010"), dmy("06/01/2010"), dmy("15/01/2010"), dmy("15/01/2010"), dmy("13/01/2010"), dmy("30/01/2010"), dmy("31/01/2010"))
)

hopital_record$epi_len <- 0
hopital_record$admin_period <- interval(hopital_record$admin_dt, hopital_record$discharge_dt)

# time point, fixed episode grouping
cbind(
  hopital_record,
  episode_group(hopital_record, date=admin_dt, sn=rd_id, case_length = epi_len, group_stats = TRUE)
  ) %>% select(-admin_period)

#> Episode or recurrence window 1 .
#> 2 of 8 record(s) grouped into episodes. 6 records not yet grouped.
#> Episode or recurrence window 2 .
#> 1 of 6 record(s) grouped into episodes. 5 records not yet grouped.
#> Episode or recurrence window 3 .
#> 2 of 5 record(s) grouped into episodes. 3 records not yet grouped.
#> Episode or recurrence window 4 .
#> 1 of 3 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 5 .
#> 1 of 2 record(s) grouped into episodes. 1 records not yet grouped.
#> Episode or recurrence window 6 .
#> 1 of 1 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 4 record(s) assinged a unique ID. 

#>   rd_id   admin_dt discharge_dt epi_len sn epid   case_nm
#> 1     1 2010-01-01   2010-01-01       0  1    1      Case
#> 2     2 2010-01-01   2010-01-10       0  2    1 Duplicate
#> 3     3 2010-01-05   2010-01-06       0  3    3      Case
#> 4     4 2010-01-05   2010-01-15       0  4    3 Duplicate
#> 5     5 2010-01-07   2010-01-15       0  5    5      Case
#> 6     6 2010-01-04   2010-01-13       0  6    6      Case
#> 7     7 2010-01-20   2010-01-30       0  7    7      Case
#> 8     8 2010-01-26   2010-01-31       0  8    8      Case
#>                    epid_interval epid_length epid_total
#> 1 2010-01-01 UTC--2010-01-01 UTC      0 days          2
#> 2 2010-01-01 UTC--2010-01-01 UTC      0 days          2
#> 3 2010-01-05 UTC--2010-01-05 UTC      0 days          2
#> 4 2010-01-05 UTC--2010-01-05 UTC      0 days          2
#> 5 2010-01-07 UTC--2010-01-07 UTC      0 days          1
#> 6 2010-01-04 UTC--2010-01-04 UTC      0 days          1
#> 7 2010-01-20 UTC--2010-01-20 UTC      0 days          1
#> 8 2010-01-26 UTC--2010-01-26 UTC      0 days          1

# time interval, fixed episode grouping
cbind(
  hopital_record,
  episode_group(hopital_record, date=admin_period, sn=rd_id, case_length = epi_len)
  ) %>% select(-c(admin_dt, discharge_dt))

#> Episode or recurrence window 1 .
#> 6 of 8 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 2 .
#> 2 of 2 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#>   rd_id epi_len                   admin_period sn epid   case_nm
#> 1     1       0 2010-01-01 UTC--2010-01-01 UTC  1    2 Duplicate
#> 2     2       0 2010-01-01 UTC--2010-01-10 UTC  2    2      Case
#> 3     3       0 2010-01-05 UTC--2010-01-06 UTC  3    2 Duplicate
#> 4     4       0 2010-01-05 UTC--2010-01-15 UTC  4    2 Duplicate
#> 5     5       0 2010-01-07 UTC--2010-01-15 UTC  5    2 Duplicate
#> 6     6       0 2010-01-04 UTC--2010-01-13 UTC  6    2 Duplicate
#> 7     7       0 2010-01-20 UTC--2010-01-30 UTC  7    7      Case
#> 8     8       0 2010-01-26 UTC--2010-01-31 UTC  8    7 Duplicate

#time interval, rolling episode.
cbind(hopital_record,
      episode_group(hopital_record, date=admin_period, sn=rd_id, case_length = epi_len, episode_type = "rolling")
      ) %>% select(-c(admin_dt, discharge_dt))

#> Episode or recurrence window 1 .
#> 6 of 8 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 2 .
#> 0 of 2 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 3 .
#> 2 of 2 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#>   rd_id epi_len                   admin_period sn epid   case_nm
#> 1     1       0 2010-01-01 UTC--2010-01-01 UTC  1    2 Duplicate
#> 2     2       0 2010-01-01 UTC--2010-01-10 UTC  2    2      Case
#> 3     3       0 2010-01-05 UTC--2010-01-06 UTC  3    2 Duplicate
#> 4     4       0 2010-01-05 UTC--2010-01-15 UTC  4    2 Duplicate
#> 5     5       0 2010-01-07 UTC--2010-01-15 UTC  5    2 Duplicate
#> 6     6       0 2010-01-04 UTC--2010-01-13 UTC  6    2 Duplicate
#> 7     7       0 2010-01-20 UTC--2010-01-30 UTC  7    7      Case
#> 8     8       0 2010-01-26 UTC--2010-01-31 UTC  8    7 Duplicate

#Episodes of overlapping intervals
cbind(
  hopital_record,
  episode_group(hopital_record, date=admin_period, sn=rd_id, case_length = epi_len, overlap_method = "overlap")
  ) %>% select(-c(admin_dt, discharge_dt))

#> Episode or recurrence window 1 .
#> 4 of 8 record(s) grouped into episodes. 4 records not yet grouped.
#> Episode or recurrence window 2 .
#> 1 of 4 record(s) grouped into episodes. 3 records not yet grouped.
#> Episode or recurrence window 3 .
#> 1 of 3 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 4 .
#> 2 of 2 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 2 record(s) assinged a unique ID. 

#>   rd_id epi_len                   admin_period sn epid   case_nm
#> 1     1       0 2010-01-01 UTC--2010-01-01 UTC  1    1      Case
#> 2     2       0 2010-01-01 UTC--2010-01-10 UTC  2    2      Case
#> 3     3       0 2010-01-05 UTC--2010-01-06 UTC  3    3      Case
#> 4     4       0 2010-01-05 UTC--2010-01-15 UTC  4    2 Duplicate
#> 5     5       0 2010-01-07 UTC--2010-01-15 UTC  5    2 Duplicate
#> 6     6       0 2010-01-04 UTC--2010-01-13 UTC  6    2 Duplicate
#> 7     7       0 2010-01-20 UTC--2010-01-30 UTC  7    7      Case
#> 8     8       0 2010-01-26 UTC--2010-01-31 UTC  8    7 Duplicate

#Episodes of intervals within other intervals
cbind(
  hopital_record,
  episode_group(hopital_record, date=admin_period, sn=rd_id, case_length = epi_len, overlap_method = "within")
  ) %>% select(-c(admin_dt, discharge_dt))

#> Episode or recurrence window 1 .
#> 2 of 8 record(s) grouped into episodes. 6 records not yet grouped.
#> Episode or recurrence window 2 .
#> 1 of 6 record(s) grouped into episodes. 5 records not yet grouped.
#> Episode or recurrence window 3 .
#> 1 of 5 record(s) grouped into episodes. 4 records not yet grouped.
#> Episode or recurrence window 4 .
#> 1 of 4 record(s) grouped into episodes. 3 records not yet grouped.
#> Episode or recurrence window 5 .
#> 1 of 3 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 6 .
#> 1 of 2 record(s) grouped into episodes. 1 records not yet grouped.
#> Episode or recurrence window 7 .
#> 1 of 1 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 6 record(s) assinged a unique ID. 

#>   rd_id epi_len                   admin_period sn epid   case_nm
#> 1     1       0 2010-01-01 UTC--2010-01-01 UTC  1    1      Case
#> 2     2       0 2010-01-01 UTC--2010-01-10 UTC  2    2      Case
#> 3     3       0 2010-01-05 UTC--2010-01-06 UTC  3    2 Duplicate
#> 4     4       0 2010-01-05 UTC--2010-01-15 UTC  4    4      Case
#> 5     5       0 2010-01-07 UTC--2010-01-15 UTC  5    5      Case
#> 6     6       0 2010-01-04 UTC--2010-01-13 UTC  6    6      Case
#> 7     7       0 2010-01-20 UTC--2010-01-30 UTC  7    7      Case
#> 8     8       0 2010-01-26 UTC--2010-01-31 UTC  8    8      Case

#Episodes of chained intervals episodes, and those with aligned end periods
cbind(
  hopital_record,
  episode_group(hopital_record, date=admin_period, sn=rd_id, case_length = epi_len, overlap_method = c("chain","aligns_end"))
  ) %>% select(-c(admin_dt, discharge_dt))

#> Episode or recurrence window 1 .
#> 2 of 8 record(s) grouped into episodes. 6 records not yet grouped.
#> Episode or recurrence window 2 .
#> 1 of 6 record(s) grouped into episodes. 5 records not yet grouped.
#> Episode or recurrence window 3 .
#> 2 of 5 record(s) grouped into episodes. 3 records not yet grouped.
#> Episode or recurrence window 4 .
#> 1 of 3 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 5 .
#> 1 of 2 record(s) grouped into episodes. 1 records not yet grouped.
#> Episode or recurrence window 6 .
#> 1 of 1 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 4 record(s) assinged a unique ID. 

#>   rd_id epi_len                   admin_period sn epid   case_nm
#> 1     1       0 2010-01-01 UTC--2010-01-01 UTC  1    2 Duplicate
#> 2     2       0 2010-01-01 UTC--2010-01-10 UTC  2    2      Case
#> 3     3       0 2010-01-05 UTC--2010-01-06 UTC  3    3      Case
#> 4     4       0 2010-01-05 UTC--2010-01-15 UTC  4    4      Case
#> 5     5       0 2010-01-07 UTC--2010-01-15 UTC  5    4 Duplicate
#> 6     6       0 2010-01-04 UTC--2010-01-13 UTC  6    6      Case
#> 7     7       0 2010-01-20 UTC--2010-01-30 UTC  7    7      Case
#> 8     8       0 2010-01-26 UTC--2010-01-31 UTC  8    8      Case