Episode grouping for record deduplication and case assignment

This function assigns unique episode indentifiers to chronological episodes

episode_group(df, sn = NULL, strata = NULL, date, case_length,
  episode_type = "fixed", episode_unit = "days", episodes_max = Inf,
  recurrence_length = NULL, rolls_max = Inf, data_source = NULL,
  custom_sort = NULL, from_last = FALSE,
  overlap_method = c("overlap", "within", "aligns_start", "aligns_end",
  "chain"), bi_direction = FALSE, group_stats = FALSE,
  display = TRUE)

Arguments

df
Dataframe. One or more datasets appended together.

sn
Unique numeric record indentifier.

strata
Column names. Episodes will be unique to each strata. See record_group.

date
Record date or interval. date, datetime or interval objects.

case_length
Duration from the "Case" within which another record of the same strata will be considered a "Duplicate" record.

episode_type
"fixed" or "rolling".

episode_unit
Time units as supported by duration.

episodes_max
Maximum number times to group episodes in each strata.

recurrence_length
Duration from the last record of an episode within which another record of the same strata will be considered a "Recurrent" record.

rolls_max
Maximum number of recurrence periods permitted within each episode. Only used if episode_type is "rolling".

data_source
Unique dataset indentifier for the dataframe. Useful when dataframe contains multiple datsets.

custom_sort
If TRUE, "Case" assignment will be in preference to this sort order. Useful in specifying that episode grouping begins at a particular kind of record regardless of chronological order.

from_last
If TRUE, episode grouping will be backwards in time - starting at the most recent record and proceeding to the earliest. If FALSE, it'll be forward in time - starting at the earliest record and proceeding to the most recent.

overlap_method
A set methods for grouped intervals to overlap. Options are; "overlap", "aligns_start", "aligns_end", "within", "chain" or all (default),

bi_direction
If FALSE, "Duplicate" records are those within the case_length and recurrence_length, before or after the "Case" as determined by from_last. If TRUE, "Duplicate" records are those on either side of the "Case".

group_stats
If TRUE, output will include additional columns with useful stats for each episode.

display
If TRUE, status messages are not printed on screen.

Value

Dataframe

sn - unique record identifier as provided

epid - unique episode indentifier

case_nm - record type in regards to case assignment

epid_dataset - datasets contained in each episode

epid_interval - lubridate interval object. Episode start and end dates

epid_length - difftime object. Difference between episode start and end dates. Unit will match that supplied as episode_unit if possible othewrise, a difference in days is returned

epid_total - number of records in each record group

Details

Episode grouping begins at a reference record ("Case") and proceeds forward or backward in time depending on from_last. If custom_sort is used, episode grouping can be forced to begin at certain record before proceeding forward or backwards in time. The maximun duration of a "fixed" episode is the case_length while, the maximum duration of a "rolling" episode is the case_length in addition to all recurrence periods. A recurrence period is the recurrence_length from the last record in an episode

See also

Examples

library(lubridate)

#> 
#> Attaching package: 'lubridate'

#> The following object is masked from 'package:base':
#> 
#>     date

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:lubridate':
#> 
#>     intersect, setdiff, union

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

hospital_infections <- tibble(
  rd_id = c(1:11),
  date = seq.Date(dmy("01/04/2018"), dmy("31/05/2018"), by="6 days"),
  infection = c("BSI", rep("UTI",2), "UTI", "BSI",  "UTI", rep("BSI",2), "RTI","RTI","BSI"),
  epi_len = 15
)

hospital_infections

#> # A tibble: 11 x 4
#>    rd_id date       infection epi_len
#>    <int> <date>     <chr>       <dbl>
#>  1     1 2018-04-01 BSI            15
#>  2     2 2018-04-07 UTI            15
#>  3     3 2018-04-13 UTI            15
#>  4     4 2018-04-19 UTI            15
#>  5     5 2018-04-25 BSI            15
#>  6     6 2018-05-01 UTI            15
#>  7     7 2018-05-07 BSI            15
#>  8     8 2018-05-13 BSI            15
#>  9     9 2018-05-19 RTI            15
#> 10    10 2018-05-25 RTI            15
#> 11    11 2018-05-31 BSI            15

# 16-day (difference of 15 days) episodes, and the earliest record defined as the "Case"
bind_cols(hospital_infections,
          episode_group(hospital_infections, sn=rd_id, date = date, case_length = epi_len)
          ) %>% select(-sn)

#> Episode or recurrence window 1 .
#> 3 of 11 record(s) grouped into episodes. 8 records not yet grouped.
#> Episode or recurrence window 2 .
#> 3 of 8 record(s) grouped into episodes. 5 records not yet grouped.
#> Episode or recurrence window 3 .
#> 3 of 5 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 4 .
#> 2 of 2 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#> # A tibble: 11 x 6
#>    rd_id date       infection epi_len  epid case_nm  
#>    <int> <date>     <chr>       <dbl> <dbl> <chr>    
#>  1     1 2018-04-01 BSI            15     1 Case     
#>  2     2 2018-04-07 UTI            15     1 Duplicate
#>  3     3 2018-04-13 UTI            15     1 Duplicate
#>  4     4 2018-04-19 UTI            15     4 Case     
#>  5     5 2018-04-25 BSI            15     4 Duplicate
#>  6     6 2018-05-01 UTI            15     4 Duplicate
#>  7     7 2018-05-07 BSI            15     7 Case     
#>  8     8 2018-05-13 BSI            15     7 Duplicate
#>  9     9 2018-05-19 RTI            15     7 Duplicate
#> 10    10 2018-05-25 RTI            15    10 Case     
#> 11    11 2018-05-31 BSI            15    10 Duplicate

# 16-hour (difference of 15 hours) episodes, and the most recent record defined as the "Case"
bind_cols(hospital_infections,
          episode_group(hospital_infections, sn=rd_id, date = date, case_length = epi_len,
                        from_last = TRUE, episode_unit = "hours", display = FALSE)
          ) %>% select(-sn)

#> Episode grouping complete - 11 record(s) assinged a unique ID. 

#> # A tibble: 11 x 6
#>    rd_id date       infection epi_len  epid case_nm
#>    <int> <date>     <chr>       <dbl> <dbl> <chr>  
#>  1     1 2018-04-01 BSI            15     1 Case   
#>  2     2 2018-04-07 UTI            15     2 Case   
#>  3     3 2018-04-13 UTI            15     3 Case   
#>  4     4 2018-04-19 UTI            15     4 Case   
#>  5     5 2018-04-25 BSI            15     5 Case   
#>  6     6 2018-05-01 UTI            15     6 Case   
#>  7     7 2018-05-07 BSI            15     7 Case   
#>  8     8 2018-05-13 BSI            15     8 Case   
#>  9     9 2018-05-19 RTI            15     9 Case   
#> 10    10 2018-05-25 RTI            15    10 Case   
#> 11    11 2018-05-31 BSI            15    11 Case   

# 15-week (difference of 9072000 seconds) episodes , and the most recent record defined as the "Case"
bind_cols(hospital_infections,
          episode_group(hospital_infections, sn=rd_id, date = date, case_length = epi_len,
                        from_last = TRUE, episode_unit = "weeks", display = FALSE)
          ) %>% select(-sn)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#> # A tibble: 11 x 6
#>    rd_id date       infection epi_len  epid case_nm  
#>    <int> <date>     <chr>       <dbl> <int> <chr>    
#>  1     1 2018-04-01 BSI            15    11 Duplicate
#>  2     2 2018-04-07 UTI            15    11 Duplicate
#>  3     3 2018-04-13 UTI            15    11 Duplicate
#>  4     4 2018-04-19 UTI            15    11 Duplicate
#>  5     5 2018-04-25 BSI            15    11 Duplicate
#>  6     6 2018-05-01 UTI            15    11 Duplicate
#>  7     7 2018-05-07 BSI            15    11 Duplicate
#>  8     8 2018-05-13 BSI            15    11 Duplicate
#>  9     9 2018-05-19 RTI            15    11 Duplicate
#> 10    10 2018-05-25 RTI            15    11 Duplicate
#> 11    11 2018-05-31 BSI            15    11 Case     

# 16-day (difference of 15 days) rolling episodes with a periods of recurrence, each lasting 31 days (difference of 30 days)
hospital_infections$recur <- 30
bind_cols(hospital_infections,
          episode_group(hospital_infections, rd_id, date=date, case_length = epi_len, episode_type = "rolling",
                        recurrence_length = recur, display = FALSE)
          ) %>% select(-sn)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#> # A tibble: 11 x 7
#>    rd_id date       infection epi_len recur  epid case_nm  
#>    <int> <date>     <chr>       <dbl> <dbl> <dbl> <chr>    
#>  1     1 2018-04-01 BSI            15    30     1 Case     
#>  2     2 2018-04-07 UTI            15    30     1 Duplicate
#>  3     3 2018-04-13 UTI            15    30     1 Duplicate
#>  4     4 2018-04-19 UTI            15    30     1 Recurrent
#>  5     5 2018-04-25 BSI            15    30     1 Duplicate
#>  6     6 2018-05-01 UTI            15    30     1 Duplicate
#>  7     7 2018-05-07 BSI            15    30     1 Duplicate
#>  8     8 2018-05-13 BSI            15    30     1 Duplicate
#>  9     9 2018-05-19 RTI            15    30     1 Recurrent
#> 10    10 2018-05-25 RTI            15    30     1 Duplicate
#> 11    11 2018-05-31 BSI            15    30     1 Duplicate

# 16-day (difference of 15 days) rolling episodes with only one period of recurrence lasting 31 days (difference of 30 days)
bind_cols(hospital_infections,
          episode_group(hospital_infections, rd_id, date=date, case_length = epi_len, episode_type = "rolling",
                        recurrence_length = recur, rolls_max = 1, display = FALSE)
          ) %>% select(-sn)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#> # A tibble: 11 x 7
#>    rd_id date       infection epi_len recur  epid case_nm  
#>    <int> <date>     <chr>       <dbl> <dbl> <dbl> <chr>    
#>  1     1 2018-04-01 BSI            15    30     1 Case     
#>  2     2 2018-04-07 UTI            15    30     1 Duplicate
#>  3     3 2018-04-13 UTI            15    30     1 Duplicate
#>  4     4 2018-04-19 UTI            15    30     1 Recurrent
#>  5     5 2018-04-25 BSI            15    30     1 Duplicate
#>  6     6 2018-05-01 UTI            15    30     1 Duplicate
#>  7     7 2018-05-07 BSI            15    30     1 Duplicate
#>  8     8 2018-05-13 BSI            15    30     1 Duplicate
#>  9     9 2018-05-19 RTI            15    30     9 Case     
#> 10    10 2018-05-25 RTI            15    30     9 Duplicate
#> 11    11 2018-05-31 BSI            15    30     9 Duplicate

# Only one 16-day (difference of 15 days) episode
bind_cols(hospital_infections,
          episode_group(hospital_infections, rd_id, date=date, case_length = epi_len, episode_type = "fixed",
                        recurrence_length = recur, episodes_max = 1, from_last = FALSE, display = FALSE)
          ) %>% select(-sn)

#> Episode grouping complete - 8 record(s) assinged a unique ID. 

#> # A tibble: 11 x 7
#>    rd_id date       infection epi_len recur  epid case_nm  
#>    <int> <date>     <chr>       <dbl> <dbl> <dbl> <chr>    
#>  1     1 2018-04-01 BSI            15    30     1 Case     
#>  2     2 2018-04-07 UTI            15    30     1 Duplicate
#>  3     3 2018-04-13 UTI            15    30     1 Duplicate
#>  4     4 2018-04-19 UTI            15    30     4 Case     
#>  5     5 2018-04-25 BSI            15    30     5 Case     
#>  6     6 2018-05-01 UTI            15    30     6 Case     
#>  7     7 2018-05-07 BSI            15    30     7 Case     
#>  8     8 2018-05-13 BSI            15    30     8 Case     
#>  9     9 2018-05-19 RTI            15    30     9 Case     
#> 10    10 2018-05-25 RTI            15    30    10 Case     
#> 11    11 2018-05-31 BSI            15    30    11 Case     

# User defined case assignment
# preference for case assignment - UTI > BSI > RTI
hospital_infections$infection <- factor(hospital_infections$infection, levels = c("UTI","BSI","RTI"))

# Different case and recurrence lengths for different source of infection
hospital_infections <- mutate(
  hospital_infections,
  epi_len = case_when(
    infection == "BSI" ~ 14,
    infection == "UTI" ~ 30,
    infection == "RTI" ~ 60
  )
)

# n-day episodes beginning with the earliest record with the specified preference; UTI > BSI > RTI
bind_cols(hospital_infections,
          episode_group(hospital_infections, rd_id, date=date, case_length = epi_len,
                        custom_sort = infection,  display = FALSE)
          ) %>% select(-sn)

#> Episode grouping complete - 2 record(s) assinged a unique ID. 

#> # A tibble: 11 x 7
#>    rd_id date       infection epi_len recur  epid case_nm  
#>    <int> <date>     <fct>       <dbl> <dbl> <dbl> <chr>    
#>  1     1 2018-04-01 BSI            14    30     1 Case     
#>  2     2 2018-04-07 UTI            30    30     2 Case     
#>  3     3 2018-04-13 UTI            30    30     2 Duplicate
#>  4     4 2018-04-19 UTI            30    30     2 Duplicate
#>  5     5 2018-04-25 BSI            14    30     2 Duplicate
#>  6     6 2018-05-01 UTI            30    30     2 Duplicate
#>  7     7 2018-05-07 BSI            14    30     2 Duplicate
#>  8     8 2018-05-13 BSI            14    30     8 Case     
#>  9     9 2018-05-19 RTI            60    30     8 Duplicate
#> 10    10 2018-05-25 RTI            60    30     8 Duplicate
#> 11    11 2018-05-31 BSI            14    30    11 Case     

# preference for case assignment - BSI > UTI, or  BSI > RTI, or earliest record
hospital_infections$infection_ord <- ifelse(hospital_infections$infection =="UTI",0,1)

# n-day episodes beginning with the earliest "BSI" record, otherwise begin at the earliest record
bind_cols(hospital_infections,
          episode_group(hospital_infections, rd_id, date=date, case_length = epi_len,
                        custom_sort = infection_ord, group_stats = TRUE, display = FALSE)
          ) %>% select(-sn)

#> Episode grouping complete - 2 record(s) assinged a unique ID. 

#> # A tibble: 11 x 11
#>    rd_id date       infection epi_len recur infection_ord  epid case_nm
#>    <int> <date>     <fct>       <dbl> <dbl>         <dbl> <dbl> <chr>  
#>  1     1 2018-04-01 BSI            14    30             1     1 Case   
#>  2     2 2018-04-07 UTI            30    30             0     2 Case   
#>  3     3 2018-04-13 UTI            30    30             0     2 Duplic~
#>  4     4 2018-04-19 UTI            30    30             0     2 Duplic~
#>  5     5 2018-04-25 BSI            14    30             1     2 Duplic~
#>  6     6 2018-05-01 UTI            30    30             0     2 Duplic~
#>  7     7 2018-05-07 BSI            14    30             1     2 Duplic~
#>  8     8 2018-05-13 BSI            14    30             1     8 Case   
#>  9     9 2018-05-19 RTI            60    30             1     8 Duplic~
#> 10    10 2018-05-25 RTI            60    30             1     8 Duplic~
#> 11    11 2018-05-31 BSI            14    30             1    11 Case   
#> # ... with 3 more variables: epid_interval <S4: Interval>, epid_length <time>,
#> #   epid_total <int>

# preference for case assignment - RTI > UTI, or  RTI > BSI, or earliest record
hospital_infections$infection_ord <- ifelse(hospital_infections$infection =="RTI",0,1)

# n-day episodes with duplicates before and after the most recent "RTI" record, otherwise begin at the most recent record
bind_cols(hospital_infections,
          episode_group(hospital_infections, rd_id, date=date, case_length = epi_len,
                        custom_sort = infection_ord, from_last = TRUE, bi_direction = TRUE, display = FALSE)
          ) %>% select(-sn)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#> # A tibble: 11 x 8
#>    rd_id date       infection epi_len recur infection_ord  epid case_nm  
#>    <int> <date>     <fct>       <dbl> <dbl>         <dbl> <int> <chr>    
#>  1     1 2018-04-01 BSI            14    30             1    10 Duplicate
#>  2     2 2018-04-07 UTI            30    30             1    10 Duplicate
#>  3     3 2018-04-13 UTI            30    30             1    10 Duplicate
#>  4     4 2018-04-19 UTI            30    30             1    10 Duplicate
#>  5     5 2018-04-25 BSI            14    30             1    10 Duplicate
#>  6     6 2018-05-01 UTI            30    30             1    10 Duplicate
#>  7     7 2018-05-07 BSI            14    30             1    10 Duplicate
#>  8     8 2018-05-13 BSI            14    30             1    10 Duplicate
#>  9     9 2018-05-19 RTI            60    30             0    10 Duplicate
#> 10    10 2018-05-25 RTI            60    30             0    10 Case     
#> 11    11 2018-05-31 BSI            14    30             1    10 Duplicate

# Stratified episode grouping
hospital_infections$patient_id <- c(rep("PID 1",8), rep("PID 2",3))

# Only one n-day episode per patient_id
bind_cols(hospital_infections,
          episode_group(hospital_infections, rd_id, date=date, strata = patient_id, case_length = epi_len,
                        episodes_max = 1, from_last = FALSE, display = FALSE, data_source = infection)
          ) %>% select(-sn)

#> Episode grouping complete - 5 record(s) assinged a unique ID. 

#> # A tibble: 11 x 10
#>    rd_id date       infection epi_len recur infection_ord patient_id  epid
#>    <int> <date>     <fct>       <dbl> <dbl>         <dbl> <chr>      <dbl>
#>  1     1 2018-04-01 BSI            14    30             1 PID 1          1
#>  2     2 2018-04-07 UTI            30    30             1 PID 1          1
#>  3     3 2018-04-13 UTI            30    30             1 PID 1          1
#>  4     4 2018-04-19 UTI            30    30             1 PID 1          4
#>  5     5 2018-04-25 BSI            14    30             1 PID 1          5
#>  6     6 2018-05-01 UTI            30    30             1 PID 1          6
#>  7     7 2018-05-07 BSI            14    30             1 PID 1          7
#>  8     8 2018-05-13 BSI            14    30             1 PID 1          8
#>  9     9 2018-05-19 RTI            60    30             0 PID 2          9
#> 10    10 2018-05-25 RTI            60    30             0 PID 2          9
#> 11    11 2018-05-31 BSI            14    30             1 PID 2          9
#> # ... with 2 more variables: case_nm <chr>, epid_dataset <chr>

# Only three 9-day (difference of 8 days) rolling episode per patient and infection.
hospital_infections$epi_len <- 8
bind_cols(hospital_infections,
          episode_group(hospital_infections, rd_id, date=date, strata = c(patient_id, infection), case_length = epi_len,
                        episode_type = "rolling", recurrence_length = recur, episodes_max = 3, data_source = c(patient_id, infection),
                        display = FALSE)
          ) %>% select(-sn)

#> Episode grouping complete - 3 record(s) assinged a unique ID. 

#> # A tibble: 11 x 10
#>    rd_id date       infection epi_len recur infection_ord patient_id  epid
#>    <int> <date>     <fct>       <dbl> <dbl>         <dbl> <chr>      <dbl>
#>  1     1 2018-04-01 BSI             8    30             1 PID 1          1
#>  2     2 2018-04-07 UTI             8    30             1 PID 1          2
#>  3     3 2018-04-13 UTI             8    30             1 PID 1          2
#>  4     4 2018-04-19 UTI             8    30             1 PID 1          2
#>  5     5 2018-04-25 BSI             8    30             1 PID 1          5
#>  6     6 2018-05-01 UTI             8    30             1 PID 1          2
#>  7     7 2018-05-07 BSI             8    30             1 PID 1          7
#>  8     8 2018-05-13 BSI             8    30             1 PID 1          7
#>  9     9 2018-05-19 RTI             8    30             0 PID 2          9
#> 10    10 2018-05-25 RTI             8    30             0 PID 2          9
#> 11    11 2018-05-31 BSI             8    30             1 PID 2         11
#> # ... with 2 more variables: case_nm <chr>, epid_dataset <chr>

hospital_admissions <- tibble(
  rd_id = 1:9,
  admin_dt = c(dmy("01/01/2010"), dmy("01/01/2010"), dmy("10/01/2010"), dmy("05/01/2010"),
               dmy("05/01/2010"), dmy("07/01/2010"), dmy("04/01/2010"),
               dmy("20/01/2010"), dmy("26/01/2010")),
  discharge_dt = c(dmy("01/01/2010"), dmy("10/01/2010"), dmy("13/01/2010"), dmy("06/01/2010"),
                   dmy("15/01/2010"), dmy("15/01/2010"), dmy("13/01/2010"),
                   dmy("30/01/2010"), dmy("31/01/2010"))
)

hospital_admissions$epi_len <- 0
hospital_admissions$admin_period <- interval(hospital_admissions$admin_dt, hospital_admissions$discharge_dt)
hospital_admissions

#> # A tibble: 9 x 5
#>   rd_id admin_dt   discharge_dt epi_len admin_period                  
#>   <int> <date>     <date>         <dbl> <S4: Interval>                
#> 1     1 2010-01-01 2010-01-01         0 2010-01-01 UTC--2010-01-01 UTC
#> 2     2 2010-01-01 2010-01-10         0 2010-01-01 UTC--2010-01-10 UTC
#> 3     3 2010-01-10 2010-01-13         0 2010-01-10 UTC--2010-01-13 UTC
#> 4     4 2010-01-05 2010-01-06         0 2010-01-05 UTC--2010-01-06 UTC
#> 5     5 2010-01-05 2010-01-15         0 2010-01-05 UTC--2010-01-15 UTC
#> 6     6 2010-01-07 2010-01-15         0 2010-01-07 UTC--2010-01-15 UTC
#> 7     7 2010-01-04 2010-01-13         0 2010-01-04 UTC--2010-01-13 UTC
#> 8     8 2010-01-20 2010-01-30         0 2010-01-20 UTC--2010-01-30 UTC
#> 9     9 2010-01-26 2010-01-31         0 2010-01-26 UTC--2010-01-31 UTC

# 1-day (difference of 0 days) episodes of hospital admissions
bind_cols(
  hospital_admissions,
  episode_group(hospital_admissions, date=admin_dt, sn=rd_id, case_length = epi_len, group_stats = TRUE)
  ) %>% select(-admin_period, sn)

#> Episode or recurrence window 1 .
#> 2 of 9 record(s) grouped into episodes. 7 records not yet grouped.
#> Episode or recurrence window 2 .
#> 1 of 7 record(s) grouped into episodes. 6 records not yet grouped.
#> Episode or recurrence window 3 .
#> 2 of 6 record(s) grouped into episodes. 4 records not yet grouped.
#> Episode or recurrence window 4 .
#> 1 of 4 record(s) grouped into episodes. 3 records not yet grouped.
#> Episode or recurrence window 5 .
#> 1 of 3 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 6 .
#> 1 of 2 record(s) grouped into episodes. 1 records not yet grouped.
#> Episode or recurrence window 7 .
#> 1 of 1 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 5 record(s) assinged a unique ID. 

#> # A tibble: 9 x 10
#>   rd_id admin_dt   discharge_dt epi_len    sn  epid case_nm
#>   <int> <date>     <date>         <dbl> <int> <dbl> <chr>  
#> 1     1 2010-01-01 2010-01-01         0     1     1 Case   
#> 2     2 2010-01-01 2010-01-10         0     2     1 Duplic~
#> 3     3 2010-01-10 2010-01-13         0     3     3 Case   
#> 4     4 2010-01-05 2010-01-06         0     4     4 Case   
#> 5     5 2010-01-05 2010-01-15         0     5     4 Duplic~
#> 6     6 2010-01-07 2010-01-15         0     6     6 Case   
#> 7     7 2010-01-04 2010-01-13         0     7     7 Case   
#> 8     8 2010-01-20 2010-01-30         0     8     8 Case   
#> 9     9 2010-01-26 2010-01-31         0     9     9 Case   
#> # ... with 3 more variables: epid_interval <S4: Interval>, epid_length <time>,
#> #   epid_total <int>

# episodes of overlaping intervals of admission
bind_cols(
  hospital_admissions,
  episode_group(hospital_admissions, date=admin_period, sn=rd_id, case_length = epi_len, group_stats = TRUE)
  ) %>% select(-c(admin_dt, discharge_dt, sn))

#> Episode or recurrence window 1 .
#> 7 of 9 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 2 .
#> 2 of 2 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#> # A tibble: 9 x 8
#>   rd_id epi_len admin_period                    epid case_nm
#>   <int>   <dbl> <S4: Interval>                 <dbl> <chr>  
#> 1     1       0 2010-01-01 UTC--2010-01-01 UTC     2 Duplic~
#> 2     2       0 2010-01-01 UTC--2010-01-10 UTC     2 Case   
#> 3     3       0 2010-01-10 UTC--2010-01-13 UTC     2 Duplic~
#> 4     4       0 2010-01-05 UTC--2010-01-06 UTC     2 Duplic~
#> 5     5       0 2010-01-05 UTC--2010-01-15 UTC     2 Duplic~
#> 6     6       0 2010-01-07 UTC--2010-01-15 UTC     2 Duplic~
#> 7     7       0 2010-01-04 UTC--2010-01-13 UTC     2 Duplic~
#> 8     8       0 2010-01-20 UTC--2010-01-30 UTC     8 Case   
#> 9     9       0 2010-01-26 UTC--2010-01-31 UTC     8 Duplic~
#> # ... with 3 more variables: epid_interval <S4: Interval>, epid_length <time>,
#> #   epid_total <int>

# rolling episodes of overlaping intervals of admission, and those within 10 days of the last interval
hospital_admissions$epi_len <- 0
hospital_admissions$recur <- 1
bind_cols(
  hospital_admissions,
  episode_group(hospital_admissions, date=admin_period, sn=rd_id, case_length = epi_len,
                episode_type = "rolling", recurrence_length = recur, episode_unit = "months")
  ) %>% select(-c(admin_dt, discharge_dt, sn))

#> Episode or recurrence window 1 .
#> 7 of 9 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 2 .
#> 2 of 2 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#> # A tibble: 9 x 6
#>   rd_id epi_len admin_period                   recur  epid case_nm  
#>   <int>   <dbl> <S4: Interval>                 <dbl> <dbl> <chr>    
#> 1     1       0 2010-01-01 UTC--2010-01-01 UTC     1     2 Duplicate
#> 2     2       0 2010-01-01 UTC--2010-01-10 UTC     1     2 Case     
#> 3     3       0 2010-01-10 UTC--2010-01-13 UTC     1     2 Duplicate
#> 4     4       0 2010-01-05 UTC--2010-01-06 UTC     1     2 Duplicate
#> 5     5       0 2010-01-05 UTC--2010-01-15 UTC     1     2 Duplicate
#> 6     6       0 2010-01-07 UTC--2010-01-15 UTC     1     2 Duplicate
#> 7     7       0 2010-01-04 UTC--2010-01-13 UTC     1     2 Duplicate
#> 8     8       0 2010-01-20 UTC--2010-01-30 UTC     1     2 Recurrent
#> 9     9       0 2010-01-26 UTC--2010-01-31 UTC     1     2 Duplicate

# fixed episodes of overlaping intervals of admission seperated by 1 month
hospital_admissions$epi_len <- 1
bind_cols(hospital_admissions,
          episode_group(hospital_admissions, date=admin_period, sn=rd_id, case_length = epi_len, episode_unit = "months")
          ) %>% select(-c(admin_dt, discharge_dt, sn))

#> Episode or recurrence window 1 .
#> 9 of 9 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

#> # A tibble: 9 x 6
#>   rd_id epi_len admin_period                   recur  epid case_nm  
#>   <int>   <dbl> <S4: Interval>                 <dbl> <int> <chr>    
#> 1     1       1 2010-01-01 UTC--2010-01-01 UTC     1     2 Duplicate
#> 2     2       1 2010-01-01 UTC--2010-01-10 UTC     1     2 Case     
#> 3     3       1 2010-01-10 UTC--2010-01-13 UTC     1     2 Duplicate
#> 4     4       1 2010-01-05 UTC--2010-01-06 UTC     1     2 Duplicate
#> 5     5       1 2010-01-05 UTC--2010-01-15 UTC     1     2 Duplicate
#> 6     6       1 2010-01-07 UTC--2010-01-15 UTC     1     2 Duplicate
#> 7     7       1 2010-01-04 UTC--2010-01-13 UTC     1     2 Duplicate
#> 8     8       1 2010-01-20 UTC--2010-01-30 UTC     1     2 Duplicate
#> 9     9       1 2010-01-26 UTC--2010-01-31 UTC     1     2 Duplicate

# Episodes of intervals within other intervals
hospital_admissions$epi_len <- 0
bind_cols(
  hospital_admissions,
  episode_group(hospital_admissions, date=admin_period, sn=rd_id, case_length = epi_len, overlap_method = "within")
  ) %>% select(-c(admin_dt, discharge_dt, sn))

#> Episode or recurrence window 1 .
#> 2 of 9 record(s) grouped into episodes. 7 records not yet grouped.
#> Episode or recurrence window 2 .
#> 1 of 7 record(s) grouped into episodes. 6 records not yet grouped.
#> Episode or recurrence window 3 .
#> 1 of 6 record(s) grouped into episodes. 5 records not yet grouped.
#> Episode or recurrence window 4 .
#> 2 of 5 record(s) grouped into episodes. 3 records not yet grouped.
#> Episode or recurrence window 5 .
#> 1 of 3 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 6 .
#> 1 of 2 record(s) grouped into episodes. 1 records not yet grouped.
#> Episode or recurrence window 7 .
#> 1 of 1 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 5 record(s) assinged a unique ID. 

#> # A tibble: 9 x 6
#>   rd_id epi_len admin_period                   recur  epid case_nm  
#>   <int>   <dbl> <S4: Interval>                 <dbl> <dbl> <chr>    
#> 1     1       0 2010-01-01 UTC--2010-01-01 UTC     1     1 Case     
#> 2     2       0 2010-01-01 UTC--2010-01-10 UTC     1     2 Case     
#> 3     3       0 2010-01-10 UTC--2010-01-13 UTC     1     5 Duplicate
#> 4     4       0 2010-01-05 UTC--2010-01-06 UTC     1     2 Duplicate
#> 5     5       0 2010-01-05 UTC--2010-01-15 UTC     1     5 Case     
#> 6     6       0 2010-01-07 UTC--2010-01-15 UTC     1     6 Case     
#> 7     7       0 2010-01-04 UTC--2010-01-13 UTC     1     7 Case     
#> 8     8       0 2010-01-20 UTC--2010-01-30 UTC     1     8 Case     
#> 9     9       0 2010-01-26 UTC--2010-01-31 UTC     1     9 Case     

#Episodes of chained intervals episodes, and those with aligned end periods
hospital_admissions$epi_len <- 0
bind_cols(
  hospital_admissions,
  episode_group(hospital_admissions, date=admin_period, sn=rd_id, case_length = epi_len, overlap_method = c("chain","aligns_end"))
  ) %>% select(-c(admin_dt, discharge_dt, sn))

#> Episode or recurrence window 1 .
#> 3 of 9 record(s) grouped into episodes. 6 records not yet grouped.
#> Episode or recurrence window 2 .
#> 1 of 6 record(s) grouped into episodes. 5 records not yet grouped.
#> Episode or recurrence window 3 .
#> 2 of 5 record(s) grouped into episodes. 3 records not yet grouped.
#> Episode or recurrence window 4 .
#> 1 of 3 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 5 .
#> 1 of 2 record(s) grouped into episodes. 1 records not yet grouped.
#> Episode or recurrence window 6 .
#> 1 of 1 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 4 record(s) assinged a unique ID. 

#> # A tibble: 9 x 6
#>   rd_id epi_len admin_period                   recur  epid case_nm  
#>   <int>   <dbl> <S4: Interval>                 <dbl> <dbl> <chr>    
#> 1     1       0 2010-01-01 UTC--2010-01-01 UTC     1     2 Duplicate
#> 2     2       0 2010-01-01 UTC--2010-01-10 UTC     1     2 Case     
#> 3     3       0 2010-01-10 UTC--2010-01-13 UTC     1     2 Duplicate
#> 4     4       0 2010-01-05 UTC--2010-01-06 UTC     1     4 Case     
#> 5     5       0 2010-01-05 UTC--2010-01-15 UTC     1     5 Case     
#> 6     6       0 2010-01-07 UTC--2010-01-15 UTC     1     5 Duplicate
#> 7     7       0 2010-01-04 UTC--2010-01-13 UTC     1     7 Case     
#> 8     8       0 2010-01-20 UTC--2010-01-30 UTC     1     8 Case     
#> 9     9       0 2010-01-26 UTC--2010-01-31 UTC     1     9 Case