Episode grouping for record deduplication and case assignment

Group records into chronological episodes

episode_group(df, sn = NULL, strata = NULL, date, case_length,
  episode_type = "fixed", episode_unit = "days", episodes_max = Inf,
  recurrence_length = NULL, rolls_max = Inf, data_source = NULL,
  custom_sort = NULL, from_last = FALSE, overlap_method = c("across",
  "inbetween", "aligns_start", "aligns_end", "chain"),
  bi_direction = FALSE, group_stats = FALSE, display = TRUE,
  deduplicate = FALSE, to_s4 = FALSE)

fixed_episodes(date, sn = NULL, strata = NULL, case_length,
  episode_unit = "days", episodes_max = Inf, data_source = NULL,
  custom_sort = NULL, from_last = FALSE, overlap_method = c("across",
  "inbetween", "aligns_start", "aligns_end", "chain"),
  bi_direction = FALSE, group_stats = FALSE, display = TRUE,
  deduplicate = FALSE, x, to_s4 = FALSE)

rolling_episodes(date, sn = NULL, strata = NULL, case_length,
  recurrence_length = NULL, episode_unit = "days",
  episodes_max = Inf, rolls_max = Inf, data_source = NULL,
  custom_sort = NULL, from_last = FALSE, overlap_method = c("across",
  "inbetween", "aligns_start", "aligns_end", "chain"),
  bi_direction = FALSE, group_stats = FALSE, display = TRUE,
  deduplicate = FALSE, x, to_s4 = FALSE)

Arguments

df
data.frame. One or more datasets appended together.

sn
Unique numeric record identifier. Optional.

strata
Episode grouping will be done separately for these subsets (strata) of the dataset. episode_group supports the use of multiple columns supplied as column names. record_group can be used to create strata.

date
Record date (date or datetime) or period (number_line) objects.

case_length
Period after a "Case (C)" within which another record from the same strata is considered a "Duplicate (D)" record.

episode_type
"fixed" or "rolling".

episode_unit
Time units as supported by lubridate's duration function.

episodes_max
Maximum number of times to group episodes within each strata.

recurrence_length
Period after the last record ("Case (C)" , "Duplicate (D)" or "Recurrent (R)") in an episode within which another record from the same strata is considered a "Recurrent (R)" record. If recurrence_length is not supplied, case_length is used as recurrence_length.

rolls_max
Maximum number of times an event an reoccur within an episode. Only used if episode_type is "rolling".

data_source
Unique dataset identifier. Useful when the dataset contains data from multiple sources. episode_group support the use of multiple columns supplied as column names.

custom_sort
If TRUE, "Case (C)" assignment will be in preference to this sort order. Useful in specifying that episode grouping begins at particular records regardless of chronological order. episode_group supports the use of multiple columns supplied as column names.

from_last
If TRUE, episode grouping will be backwards in time - starting at the most recent record and proceeding to the earliest. If FALSE, it'll be forward in time - starting at the earliest record and proceeding to the most recent one.

overlap_method
A set of ways for grouped intervals to overlap. Options are; "across", "aligns_start", "aligns_end", "inbetween", "chain". See overlap functions.

bi_direction
If FALSE, "Duplicate (D)" records will be those within the case_length period, before or after the "Case (C)" as determined by from_last. If TRUE, "Duplicate (D)" records will be those within the same period before and after the "Case (C)".

group_stats
If TRUE, the output will include additional columns with useful stats for each episode group.

display
If TRUE, status messages are printed on screen.

deduplicate
if TRUE, retains only one the "Case" (C) record in an episode.

to_s4
if TRUE, changes the returned value to an epid object.

x
Record date or interval. Deprecated. Please use date

Value

data.frame (epid objects if to_s4 is TRUE)

sn - unique record identifier as provided

epid - unique episode identifier

case_nm - record type in regards to case assignment

epid_dataset - datasets contained in each episode

epid_interval - Episode start and end dates. number_line object.

epid_length - Difference between episode start and end dates (difftime). If possible, it's the same unit as episode_unit otherwise, a difference in days is returned

epid_total - number of records in each episode

epid objects will be the default output from the next release onwards

Details

Episode grouping begins at a reference record ("Case (C)") and proceeds forward or backward in time depending on from_last. If custom_sort is used, episode grouping can be forced to begin at certain records before proceeding forward or backwards in time. The maximum duration of a "fixed" episode is the case_length while, the maximum duration of a "rolling" episode is the case_length plus all recurrence periods. A recurrence period is a fixed period (recurrence_length) after the last record in an episode.

fixed_episodes() and rolling_episodes() are wrapper functions of episode_group. They can be more convenient options and have the same functionalities as episode_group().

See also

Examples

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

library(lubridate)

#> 
#> Attaching package: 'lubridate'

#> The following object is masked from 'package:base':
#> 
#>     date

#1. Fixed episodes
data(infections); infections

#> # A tibble: 11 x 4
#>    rd_id date       infection epi_len
#>    <int> <date>     <chr>       <dbl>
#>  1     1 2018-04-01 BSI            15
#>  2     2 2018-04-07 UTI            15
#>  3     3 2018-04-13 UTI            15
#>  4     4 2018-04-19 UTI            15
#>  5     5 2018-04-25 BSI            15
#>  6     6 2018-05-01 UTI            15
#>  7     7 2018-05-07 BSI            15
#>  8     8 2018-05-13 BSI            15
#>  9     9 2018-05-19 RTI            15
#> 10    10 2018-05-25 RTI            15
#> 11    11 2018-05-31 BSI            15

db_1 <- infections
# 16-day (difference of 15 days) episodes beginning from the earliest record
db_1$fd <- fixed_episodes(db_1$date, case_length = 15, to_s4 = TRUE, display = FALSE)

#> The default output of fixed_episodes() will be changed to epid objects in the next release.
#> Please consider switching earlier by using 'to_s4=TRUE' or to_s4()
#> This message is displayed once per session.

#> The default output of episode_group() will be changed to epid objects in the next release.
#> Please consider switching earlier by using 'to_s4=TRUE' or to_s4()
#> This message is displayed once per session.

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

# 16-hour (difference of 15 hours) episodes beginning from the earliest record
db_1$fh <- fixed_episodes(db_1$date, case_length = 15,
episode_unit = "hours", to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 11 record(s) assinged a unique ID. 

db_1

#> # A tibble: 11 x 6
#>    rd_id date       infection epi_len fd       fh      
#>    <int> <date>     <chr>       <dbl> <epid>   <epid>  
#>  1     1 2018-04-01 BSI            15 E-01 (C) E-01 (C)
#>  2     2 2018-04-07 UTI            15 E-01 (D) E-02 (C)
#>  3     3 2018-04-13 UTI            15 E-01 (D) E-03 (C)
#>  4     4 2018-04-19 UTI            15 E-04 (C) E-04 (C)
#>  5     5 2018-04-25 BSI            15 E-04 (D) E-05 (C)
#>  6     6 2018-05-01 UTI            15 E-04 (D) E-06 (C)
#>  7     7 2018-05-07 BSI            15 E-07 (C) E-07 (C)
#>  8     8 2018-05-13 BSI            15 E-07 (D) E-08 (C)
#>  9     9 2018-05-19 RTI            15 E-07 (D) E-09 (C)
#> 10    10 2018-05-25 RTI            15 E-10 (C) E-10 (C)
#> 11    11 2018-05-31 BSI            15 E-10 (D) E-11 (C)

#2. Rolling episodes
# Case length and recurrence periods of 15 days
db_1$rd_a <- rolling_episodes(db_1$date, case_length = 15, to_s4 = TRUE, display = FALSE)

#> The default output of rolling_episodes() will be changed to epid objects in the next release.
#> Please consider switching earlier by using 'to_s4=TRUE' or to_s4()
#> This message is displayed once per session.

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

# Case length of 15 days and recurrence periods of 10 days
db_1$rd_b <- rolling_episodes(db_1$date, case_length = 15,
recurrence_length = 10, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

# Case length of 15 days and 2 recurrence periods of 10 days
db_1$rd_c <- rolling_episodes(db_1$date, case_length = 15,
recurrence_length = 10, rolls_max = 2, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 1 record(s) assinged a unique ID. 

db_1

#> # A tibble: 11 x 9
#>    rd_id date       infection epi_len fd       fh       rd_a    rd_b    rd_c    
#>    <int> <date>     <chr>       <dbl> <epid>   <epid>   <epid>  <epid>  <epid>  
#>  1     1 2018-04-01 BSI            15 E-01 (C) E-01 (C) E-1 (C) E-1 (C) E-01 (C)
#>  2     2 2018-04-07 UTI            15 E-01 (D) E-02 (C) E-1 (D) E-1 (D) E-01 (D)
#>  3     3 2018-04-13 UTI            15 E-01 (D) E-03 (C) E-1 (D) E-1 (D) E-01 (D)
#>  4     4 2018-04-19 UTI            15 E-04 (C) E-04 (C) E-1 (R) E-1 (R) E-01 (R)
#>  5     5 2018-04-25 BSI            15 E-04 (D) E-05 (C) E-1 (D) E-1 (R) E-01 (R)
#>  6     6 2018-05-01 UTI            15 E-04 (D) E-06 (C) E-1 (R) E-1 (R) E-06 (C)
#>  7     7 2018-05-07 BSI            15 E-07 (C) E-07 (C) E-1 (D) E-1 (R) E-06 (D)
#>  8     8 2018-05-13 BSI            15 E-07 (D) E-08 (C) E-1 (R) E-1 (R) E-06 (D)
#>  9     9 2018-05-19 RTI            15 E-07 (D) E-09 (C) E-1 (D) E-1 (R) E-06 (R)
#> 10    10 2018-05-25 RTI            15 E-10 (C) E-10 (C) E-1 (R) E-1 (R) E-06 (R)
#> 11    11 2018-05-31 BSI            15 E-10 (D) E-11 (C) E-1 (D) E-1 (R) E-11 (C)

# 3. Stratified episode grouping
db_3 <- infections

db_3$patient_id <- c(rep("PID 1",8), rep("PID 2",3))
# One 16-day episode per patient
db_3$epids_p <- fixed_episodes(date=db_3$date, strata = db_3$patient_id,
case_length = 15, episodes_max = 1, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 5 record(s) assinged a unique ID. 

db_3

#> # A tibble: 11 x 6
#>    rd_id date       infection epi_len patient_id epids_p
#>    <int> <date>     <chr>       <dbl> <chr>      <epid> 
#>  1     1 2018-04-01 BSI            15 PID 1      E-1 (C)
#>  2     2 2018-04-07 UTI            15 PID 1      E-1 (D)
#>  3     3 2018-04-13 UTI            15 PID 1      E-1 (D)
#>  4     4 2018-04-19 UTI            15 PID 1      E-4 (C)
#>  5     5 2018-04-25 BSI            15 PID 1      E-5 (C)
#>  6     6 2018-05-01 UTI            15 PID 1      E-6 (C)
#>  7     7 2018-05-07 BSI            15 PID 1      E-7 (C)
#>  8     8 2018-05-13 BSI            15 PID 1      E-8 (C)
#>  9     9 2018-05-19 RTI            15 PID 2      E-9 (C)
#> 10    10 2018-05-25 RTI            15 PID 2      E-9 (D)
#> 11    11 2018-05-31 BSI            15 PID 2      E-9 (D)

# 4. Case assignment
db_4 <- infections

## 4.1 Chronological order
db_4$forward_time <- fixed_episodes(db_4$date, case_length = 1,
episode_unit = "month", to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

db_4$backward_time <- fixed_episodes(db_4$date, case_length = 1,
episode_unit = "month", from_last = TRUE, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

db_4

#> # A tibble: 11 x 6
#>    rd_id date       infection epi_len forward_time backward_time
#>    <int> <date>     <chr>       <dbl> <epid>       <epid>       
#>  1     1 2018-04-01 BSI            15 E-1 (C)      E-05 (D)     
#>  2     2 2018-04-07 UTI            15 E-1 (D)      E-05 (D)     
#>  3     3 2018-04-13 UTI            15 E-1 (D)      E-05 (D)     
#>  4     4 2018-04-19 UTI            15 E-1 (D)      E-05 (D)     
#>  5     5 2018-04-25 BSI            15 E-1 (D)      E-05 (C)     
#>  6     6 2018-05-01 UTI            15 E-1 (D)      E-11 (D)     
#>  7     7 2018-05-07 BSI            15 E-7 (C)      E-11 (D)     
#>  8     8 2018-05-13 BSI            15 E-7 (D)      E-11 (D)     
#>  9     9 2018-05-19 RTI            15 E-7 (D)      E-11 (D)     
#> 10    10 2018-05-25 RTI            15 E-7 (D)      E-11 (D)     
#> 11    11 2018-05-31 BSI            15 E-7 (D)      E-11 (C)     

## 4.2 User defined order
db_4b <- infections
# RTI > UTI, or RTI > BSI
db_4b$ord2 <- ifelse(db_4b$infection =="RTI",0,1)
# UTI > BSI > RTI
db_4b$ord1 <- factor(db_4b$infection, levels = c("UTI","BSI","RTI"))

db_4b$epids_1 <- fixed_episodes(db_4b$date, case_length = 15,
custom_sort = db_4b$ord2, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

db_4b$epids_2 <- fixed_episodes(db_4b$date, case_length = 15,
custom_sort = db_4b$ord1, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 3 record(s) assinged a unique ID. 

db_4b$epids_2b <- fixed_episodes(db_4b$date, case_length = 15,
custom_sort = db_4b$ord1, bi_direction = TRUE, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

db_4b

#> # A tibble: 11 x 9
#>    rd_id date       infection epi_len  ord2 ord1  epids_1 epids_2  epids_2b
#>    <int> <date>     <chr>       <dbl> <dbl> <fct> <epid>  <epid>   <epid>  
#>  1     1 2018-04-01 BSI            15     1 BSI   E-1 (C) E-01 (C) E-02 (D)
#>  2     2 2018-04-07 UTI            15     1 UTI   E-1 (D) E-02 (C) E-02 (C)
#>  3     3 2018-04-13 UTI            15     1 UTI   E-1 (D) E-02 (D) E-02 (D)
#>  4     4 2018-04-19 UTI            15     1 UTI   E-4 (C) E-02 (D) E-02 (D)
#>  5     5 2018-04-25 BSI            15     1 BSI   E-4 (D) E-05 (C) E-06 (D)
#>  6     6 2018-05-01 UTI            15     1 UTI   E-4 (D) E-06 (C) E-06 (C)
#>  7     7 2018-05-07 BSI            15     1 BSI   E-7 (C) E-06 (D) E-06 (D)
#>  8     8 2018-05-13 BSI            15     1 BSI   E-7 (D) E-06 (D) E-06 (D)
#>  9     9 2018-05-19 RTI            15     0 RTI   E-9 (C) E-09 (C) E-11 (D)
#> 10    10 2018-05-25 RTI            15     0 RTI   E-9 (D) E-09 (D) E-11 (D)
#> 11    11 2018-05-31 BSI            15     1 BSI   E-9 (D) E-11 (C) E-11 (C)

#5. Interval grouping
data(hospital_admissions)

hospital_admissions$admin_period <- number_line(hospital_admissions$admin_dt,
hospital_admissions$discharge_dt)
admissions <- hospital_admissions[c("admin_period","epi_len")]
admissions

#> # A tibble: 9 x 2
#>   admin_period             epi_len
#>   <numbr_ln>                 <dbl>
#> 1 2019-01-01 == 2019-01-01       0
#> 2 2019-01-01 -> 2019-01-10       0
#> 3 2019-01-10 -> 2019-01-13       0
#> 4 2019-01-05 -> 2019-01-06       0
#> 5 2019-01-05 -> 2019-01-15       0
#> 6 2019-01-07 -> 2019-01-15       0
#> 7 2019-01-04 -> 2019-01-13       0
#> 8 2019-01-20 -> 2019-01-30       0
#> 9 2019-01-26 -> 2019-01-31       0

# Episodes of overlaping periods of admission
admissions$epi_0 <- fixed_episodes(date=admissions$admin_period, case_length = 0,
group_stats = TRUE, to_s4=TRUE)

#> Episode or recurrence window 1.
#> 7 of 9 record(s) grouped into episodes. 2 records not yet grouped.
#> Episode or recurrence window 2.
#> 2 of 2 record(s) grouped into episodes. 0 records not yet grouped.
#> 
#> Episode grouping complete - 0 record(s) assinged a unique ID. 

admissions

#> # A tibble: 9 x 3
#>   admin_period             epi_len epi_0                           
#>   <numbr_ln>                 <dbl> <epid>                          
#> 1 2019-01-01 == 2019-01-01       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 2 2019-01-01 -> 2019-01-10       0 E-2 2019-01-01 -> 2019-01-15 (C)
#> 3 2019-01-10 -> 2019-01-13       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 4 2019-01-05 -> 2019-01-06       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 5 2019-01-05 -> 2019-01-15       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 6 2019-01-07 -> 2019-01-15       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 7 2019-01-04 -> 2019-01-13       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 8 2019-01-20 -> 2019-01-30       0 E-8 2019-01-20 -> 2019-01-31 (C)
#> 9 2019-01-26 -> 2019-01-31       0 E-8 2019-01-20 -> 2019-01-31 (D)

# Overlaping periods of admission seperated by 1 month
admissions$epi_1 <- fixed_episodes(date=admissions$admin_period, case_length = 1,
episode_unit = "months", group_stats = TRUE, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

admissions

#> # A tibble: 9 x 4
#>   admin_period             epi_len epi_0                           
#>   <numbr_ln>                 <dbl> <epid>                          
#> 1 2019-01-01 == 2019-01-01       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 2 2019-01-01 -> 2019-01-10       0 E-2 2019-01-01 -> 2019-01-15 (C)
#> 3 2019-01-10 -> 2019-01-13       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 4 2019-01-05 -> 2019-01-06       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 5 2019-01-05 -> 2019-01-15       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 6 2019-01-07 -> 2019-01-15       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 7 2019-01-04 -> 2019-01-13       0 E-2 2019-01-01 -> 2019-01-15 (D)
#> 8 2019-01-20 -> 2019-01-30       0 E-8 2019-01-20 -> 2019-01-31 (C)
#> 9 2019-01-26 -> 2019-01-31       0 E-8 2019-01-20 -> 2019-01-31 (D)
#> # ... with 1 more variable: epi_1 <epid>

# Episodes of chained admission periods, and those with aligned end periods
admissions$epi_0b <- fixed_episodes(date=admissions$admin_period, case_length = 0,
overlap_method = c("chain","aligns_end"), group_stats = TRUE, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 4 record(s) assinged a unique ID. 

admissions["epi_0b"]

#> # A tibble: 9 x 1
#>   epi_0b                          
#>   <epid>                          
#> 1 E-2 2019-01-01 -> 2019-01-13 (D)
#> 2 E-2 2019-01-01 -> 2019-01-13 (C)
#> 3 E-2 2019-01-01 -> 2019-01-13 (D)
#> 4 E-4 2019-01-05 -> 2019-01-06 (C)
#> 5 E-5 2019-01-05 -> 2019-01-15 (C)
#> 6 E-5 2019-01-05 -> 2019-01-15 (D)
#> 7 E-7 2019-01-04 -> 2019-01-13 (C)
#> 8 E-8 2019-01-20 -> 2019-01-30 (C)
#> 9 E-9 2019-01-26 -> 2019-01-31 (C)

# Note - episode_group() takes column names not actual values
db_5 <- infections

db_5$recur <- 20
db_5$epids_f <- episode_group(db_5, date=date, episode_type = "fixed",
case_length = epi_len, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

db_5$epids_r <- episode_group(db_5, date=date, episode_type = "rolling",
case_length = epi_len, recurrence_length = recur, to_s4 = TRUE, display = FALSE)

#> Episode grouping complete - 0 record(s) assinged a unique ID. 

db_5

#> # A tibble: 11 x 7
#>    rd_id date       infection epi_len recur epids_f  epids_r
#>    <int> <date>     <chr>       <dbl> <dbl> <epid>   <epid> 
#>  1     1 2018-04-01 BSI            15    20 E-01 (C) E-1 (C)
#>  2     2 2018-04-07 UTI            15    20 E-01 (D) E-1 (D)
#>  3     3 2018-04-13 UTI            15    20 E-01 (D) E-1 (D)
#>  4     4 2018-04-19 UTI            15    20 E-04 (C) E-1 (R)
#>  5     5 2018-04-25 BSI            15    20 E-04 (D) E-1 (D)
#>  6     6 2018-05-01 UTI            15    20 E-04 (D) E-1 (D)
#>  7     7 2018-05-07 BSI            15    20 E-07 (C) E-1 (R)
#>  8     8 2018-05-13 BSI            15    20 E-07 (D) E-1 (D)
#>  9     9 2018-05-19 RTI            15    20 E-07 (D) E-1 (D)
#> 10    10 2018-05-25 RTI            15    20 E-10 (C) E-1 (R)
#> 11    11 2018-05-31 BSI            15    20 E-10 (D) E-1 (D)