Episode grouping

This function assigns records into unique chronological episodes

episode_group(df, sn = NULL, strata = NULL, date, episode_length,
  episode_type = "static", episodes_max = Inf,
  rc_episode_length = NULL, rolls_max = Inf, data_source = NULL,
  source_sort = FALSE, from_last = FALSE, display = TRUE)

Arguments

df
Dataframe. One or more datasets appened together.

sn
Unique record indentifier for the dataframe

strata
Column names. Episode grouping will be straified by these columns and will be unique to each strata

date
Record date. A column of date type

episode_length
Episode duration/window in calendar days. Integer values between 0 and Inf, and should be unique to each strata.

episode_type
"static" or "rolling". If "static", each episode will only include records within a fixed period episode_length from the inital record (case). If "rolling", will include recurrent records. See rc_episode_length

episodes_max
Maximum number of episodes permitted in each strata. An integer between 0 and Inf.

rc_episode_length
Period of recurrence in calendar days. Recurrence here refers to records ocurring after episode_length of the first record but within the rc_episode_length of the last duplicate record. Only used if episode_type`is "rolling". Should also be unique to each strata. Integer value between 0 and Inf

rolls_max
Maximum number of recurrence within each episode. Integer between 0 and Inf. Only used if episode_type is "rolling".

data_source
Unique dataset indentifier for the dataframe. Usefull when dataframe contains multiple datsets.

source_sort
If TRUE, "Case" assignment is done in preference to the data source. Usefull in specifying that cases come from a particular dataset within the dataframe regardless of chronological order.

from_last
If TRUE, episode grouping will be backwards in time - starting from the most recent record to the earliest. If FALSE, it'll be forward in time - starting from the earliest record to most recent.

display
If TRUE, progress status at each stage of episode grouping is displayed on screen

Value

Dataframe with a unique episode identifier, type of record based on the episode length and type, and if selected, a list of datasets where each episode was identified

Examples

library(lubridate)

#> Warning: package 'lubridate' was built under R version 3.5.1

#> 
#> Attaching package: 'lubridate'

#> The following object is masked from 'package:base':
#> 
#>     date

library(dplyr)

data <- data.frame(date = seq.Date(dmy("01/04/2018"), dmy("31/05/2018"), by="3 days"))
data$pid <- "Patient 1"
data$episode_len <- 7
data <- mutate(data, rd_id = row_number())

cbind(head(data,10),
      episode_group(head(data,10), sn=rd_id, strata = pid, date = date, episode_length = episode_len)
      )

#> [1] "Episode window 1"
#> [1] "3 of 10 record(s) grouped into episodes. 7 records not yet grouped."
#> [1] "Episode window 2"
#> [1] "3 of 7 record(s) grouped into episodes. 4 records not yet grouped."
#> [1] "Episode window 3"
#> [1] "3 of 4 record(s) grouped into episodes. 1 records not yet grouped."
#> [1] "Episode window 4"
#> [1] "1 of 1 record(s) grouped into episodes. 0 records not yet grouped."
#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id sn epid   case_nm
#> 1  2018-04-01 Patient 1           7     1  1    1      Case
#> 2  2018-04-04 Patient 1           7     2  2    1 Duplicate
#> 3  2018-04-07 Patient 1           7     3  3    1 Duplicate
#> 4  2018-04-10 Patient 1           7     4  4    4      Case
#> 5  2018-04-13 Patient 1           7     5  5    4 Duplicate
#> 6  2018-04-16 Patient 1           7     6  6    4 Duplicate
#> 7  2018-04-19 Patient 1           7     7  7    7      Case
#> 8  2018-04-22 Patient 1           7     8  8    7 Duplicate
#> 9  2018-04-25 Patient 1           7     9  9    7 Duplicate
#> 10 2018-04-28 Patient 1           7    10 10   10      Case

# The progress updates are displayed and can be disabled by setting "display" to FALSE
cbind(head(data,10),
      episode_group(head(data,10), sn=rd_id, strata = pid, date = date, episode_length = episode_len, display = FALSE)
      )

#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id sn epid   case_nm
#> 1  2018-04-01 Patient 1           7     1  1    1      Case
#> 2  2018-04-04 Patient 1           7     2  2    1 Duplicate
#> 3  2018-04-07 Patient 1           7     3  3    1 Duplicate
#> 4  2018-04-10 Patient 1           7     4  4    4      Case
#> 5  2018-04-13 Patient 1           7     5  5    4 Duplicate
#> 6  2018-04-16 Patient 1           7     6  6    4 Duplicate
#> 7  2018-04-19 Patient 1           7     7  7    7      Case
#> 8  2018-04-22 Patient 1           7     8  8    7 Duplicate
#> 9  2018-04-25 Patient 1           7     9  9    7 Duplicate
#> 10 2018-04-28 Patient 1           7    10 10   10      Case

# By default, episode grouping starts from the earliest record of within each strata and proceeds forward in time. The reverse can be achieved by setting "from_last" to TRUE
# Below is an example of episode grouping from both directions with an episode length of 13 calendar days
data_2 <- mutate(head(data,10), episode_len_s=13)

cbind(data_2,
      select(episode_group(data_2, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, display = FALSE, from_last = FALSE), -sn, epid.1=epid, case.1=case_nm),
      select(episode_group(data_2, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, display = FALSE, from_last = TRUE), -sn, epid.2=epid, case.2=case_nm)
      )

#> [1] "Episode grouping complete"
#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id episode_len_s epid.1    case.1 epid.2
#> 1  2018-04-01 Patient 1           7     1            13      1      Case      5
#> 2  2018-04-04 Patient 1           7     2            13      1 Duplicate      5
#> 3  2018-04-07 Patient 1           7     3            13      1 Duplicate      5
#> 4  2018-04-10 Patient 1           7     4            13      1 Duplicate      5
#> 5  2018-04-13 Patient 1           7     5            13      1 Duplicate      5
#> 6  2018-04-16 Patient 1           7     6            13      6      Case     10
#> 7  2018-04-19 Patient 1           7     7            13      6 Duplicate     10
#> 8  2018-04-22 Patient 1           7     8            13      6 Duplicate     10
#> 9  2018-04-25 Patient 1           7     9            13      6 Duplicate     10
#> 10 2018-04-28 Patient 1           7    10            13      6 Duplicate     10
#>       case.2
#> 1  Duplicate
#> 2  Duplicate
#> 3  Duplicate
#> 4  Duplicate
#> 5       Case
#> 6  Duplicate
#> 7  Duplicate
#> 8  Duplicate
#> 9  Duplicate
#> 10      Case

# By deafult, "episode_type" is set to "static". This assumes an episode does not re-occur after a fixed number of days from the inital record (episode_length), refered to as the "Case".
# Conversely, chosing an "episode_type" of "rolling" will assume a record can re-occur if it's within a fixed number of days after the most recent report within the episode ("rc_episode_length"), this could be a "Case" or "Duplicate" record.
# When "episode_type" is "rolling", The "Case", "Duplicates" of the "Case", "Recurrent" record, and "Duplicates" of the "Recurrent" record are assigned to the same episode.
# The example below demonstrates this

cbind(data_2,
      select(episode_group(data_2, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, episode_type ="rolling", display = FALSE, from_last = FALSE), -sn, epid.1=epid, case.1=case_nm),
      select(episode_group(data_2, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, episode_type ="rolling", display = FALSE, from_last = TRUE), -sn, epid.2=epid, case.2=case_nm)
      )

#> [1] "Episode grouping complete"
#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id episode_len_s epid.1    case.1 epid.2
#> 1  2018-04-01 Patient 1           7     1            13      1      Case     10
#> 2  2018-04-04 Patient 1           7     2            13      1 Duplicate     10
#> 3  2018-04-07 Patient 1           7     3            13      1 Duplicate     10
#> 4  2018-04-10 Patient 1           7     4            13      1 Duplicate     10
#> 5  2018-04-13 Patient 1           7     5            13      1 Duplicate     10
#> 6  2018-04-16 Patient 1           7     6            13      1 Recurrent     10
#> 7  2018-04-19 Patient 1           7     7            13      1 Duplicate     10
#> 8  2018-04-22 Patient 1           7     8            13      1 Duplicate     10
#> 9  2018-04-25 Patient 1           7     9            13      1 Duplicate     10
#> 10 2018-04-28 Patient 1           7    10            13      1 Recurrent     10
#>       case.2
#> 1  Recurrent
#> 2  Duplicate
#> 3  Duplicate
#> 4  Duplicate
#> 5  Recurrent
#> 6  Duplicate
#> 7  Duplicate
#> 8  Duplicate
#> 9  Duplicate
#> 10      Case

# By default, the window period for the "Recurrent" record is the same as that for the "Case" (episode_length).
# However, a different "rc_episode_length" can be used.
# The example below demonsrates the difference between using the same "episode_length" and "rc_episode_length" (13 calendar days) compared to an "episode_length" of 13 days and an "rc_episode_length" of 30 days.

data_4 <- mutate(data_2, recurrence=30)
cbind(data_4,
      select(episode_group(data_4, sn=rd_id, strata = pid, date = date, episode_length = episode_len, episode_type ="rolling", display = FALSE), -sn, epid.1=epid, case.1=case_nm),
      select(episode_group(data_4, sn=rd_id, strata = pid, date = date, episode_length = episode_len, episode_type ="rolling", rc_episode_length = recurrence, display = FALSE), -sn, epid.2=epid, case.2=case_nm)
      )

#> [1] "Episode grouping complete"
#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id episode_len_s recurrence epid.1
#> 1  2018-04-01 Patient 1           7     1            13         30      1
#> 2  2018-04-04 Patient 1           7     2            13         30      1
#> 3  2018-04-07 Patient 1           7     3            13         30      1
#> 4  2018-04-10 Patient 1           7     4            13         30      1
#> 5  2018-04-13 Patient 1           7     5            13         30      1
#> 6  2018-04-16 Patient 1           7     6            13         30      1
#> 7  2018-04-19 Patient 1           7     7            13         30      1
#> 8  2018-04-22 Patient 1           7     8            13         30      1
#> 9  2018-04-25 Patient 1           7     9            13         30      1
#> 10 2018-04-28 Patient 1           7    10            13         30      1
#>       case.1 epid.2    case.2
#> 1       Case      1      Case
#> 2  Duplicate      1 Duplicate
#> 3  Duplicate      1 Duplicate
#> 4  Recurrent      1 Recurrent
#> 5  Duplicate      1 Duplicate
#> 6  Recurrent      1 Duplicate
#> 7  Duplicate      1 Duplicate
#> 8  Recurrent      1 Duplicate
#> 9  Duplicate      1 Duplicate
#> 10 Recurrent      1 Duplicate

# Both scenerio yield one episode and therefore one "Case" however, the first 4 "Recurrent" records while the second has only 2 "Recurrent" records.

# By default, the function will assume a record can continue to re-occur indefinitely until there is no chronological record within the "rc_episode_length" of the most recent record in the epsiode.
# The "rolls_max" argument can be used to set the number of times an episode can re-occur as defined above
# Once "rolls_max" is reached the next chronological records is assigned to a new episode even if it's within the "rc_episode_length" of most recent record.
# The example below demonsrates the difference between an Infinite "roll_max" (default) and a "rolls_max" of 1.

data_4 <- mutate(data_4, recurrence=4)
cbind(data_4,
      select(episode_group(data_4, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, episode_type ="rolling", rc_episode_length = recurrence, display = FALSE), -sn, epid.1=epid, case.1=case_nm),
      select(episode_group(data_4, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, episode_type ="rolling", rc_episode_length = recurrence, rolls_max = 1,  display = FALSE), -sn, epid.2=epid, case.2=case_nm)
      )

#> [1] "Episode grouping complete"
#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id episode_len_s recurrence epid.1
#> 1  2018-04-01 Patient 1           7     1            13          4      1
#> 2  2018-04-04 Patient 1           7     2            13          4      1
#> 3  2018-04-07 Patient 1           7     3            13          4      1
#> 4  2018-04-10 Patient 1           7     4            13          4      1
#> 5  2018-04-13 Patient 1           7     5            13          4      1
#> 6  2018-04-16 Patient 1           7     6            13          4      1
#> 7  2018-04-19 Patient 1           7     7            13          4      1
#> 8  2018-04-22 Patient 1           7     8            13          4      1
#> 9  2018-04-25 Patient 1           7     9            13          4      1
#> 10 2018-04-28 Patient 1           7    10            13          4      1
#>       case.1 epid.2    case.2
#> 1       Case      1      Case
#> 2  Duplicate      1 Duplicate
#> 3  Duplicate      1 Duplicate
#> 4  Duplicate      1 Duplicate
#> 5  Duplicate      1 Duplicate
#> 6  Recurrent      1 Recurrent
#> 7  Recurrent      7      Case
#> 8  Recurrent      7 Duplicate
#> 9  Recurrent      7 Duplicate
#> 10 Recurrent      7 Duplicate

# Note a "rolls_max" of "0" will have the result as as when "episode_type" is "static". See below.

cbind(data_4,
      select(episode_group(data_4, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, episode_type ="static", rc_episode_length = recurrence, display = FALSE), -sn, epid.1=epid, case.1=case_nm),
      select(episode_group(data_4, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, episode_type ="rolling", rc_episode_length = recurrence, rolls_max = 0,  display = FALSE), -sn, epid.2=epid, case.2=case_nm)
      )

#> [1] "Episode grouping complete"
#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id episode_len_s recurrence epid.1
#> 1  2018-04-01 Patient 1           7     1            13          4      1
#> 2  2018-04-04 Patient 1           7     2            13          4      1
#> 3  2018-04-07 Patient 1           7     3            13          4      1
#> 4  2018-04-10 Patient 1           7     4            13          4      1
#> 5  2018-04-13 Patient 1           7     5            13          4      1
#> 6  2018-04-16 Patient 1           7     6            13          4      6
#> 7  2018-04-19 Patient 1           7     7            13          4      6
#> 8  2018-04-22 Patient 1           7     8            13          4      6
#> 9  2018-04-25 Patient 1           7     9            13          4      6
#> 10 2018-04-28 Patient 1           7    10            13          4      6
#>       case.1 epid.2    case.2
#> 1       Case      1      Case
#> 2  Duplicate      1 Duplicate
#> 3  Duplicate      1 Duplicate
#> 4  Duplicate      1 Duplicate
#> 5  Duplicate      1 Duplicate
#> 6       Case      6      Case
#> 7  Duplicate      6 Duplicate
#> 8  Duplicate      6 Duplicate
#> 9  Duplicate      6 Duplicate
#> 10 Duplicate      6 Duplicate

# Similarly, by default, the function will group every record within a strata into epsiodes.
# The "episodes_max" argument can be used to set the number of times records are time records are grouped into episodes.
# Once "episodes_max" is reached, subsequent records are assigned to new distinct episodes ids.
# The example below demonsrates the difference between an Infinite "episode_max" (default) and an "episode_max" of 1.

cbind(data_4,
      select(episode_group(data_4, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, episode_type ="static", rc_episode_length = recurrence, episodes_max = 1, display = FALSE), -sn, epid_l=epid, case_l=case_nm),
      select(episode_group(data_4, sn=rd_id, strat = pid, date = date, episode_length = episode_len_s, episode_type ="static", rc_episode_length = recurrence, episodes_max = 2,  display = FALSE), -sn, epid_s=epid, case_s=case_nm)
      )

#> [1] "Episode grouping complete"
#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id episode_len_s recurrence epid_l
#> 1  2018-04-01 Patient 1           7     1            13          4      1
#> 2  2018-04-04 Patient 1           7     2            13          4      1
#> 3  2018-04-07 Patient 1           7     3            13          4      1
#> 4  2018-04-10 Patient 1           7     4            13          4      1
#> 5  2018-04-13 Patient 1           7     5            13          4      1
#> 6  2018-04-16 Patient 1           7     6            13          4      6
#> 7  2018-04-19 Patient 1           7     7            13          4      7
#> 8  2018-04-22 Patient 1           7     8            13          4      8
#> 9  2018-04-25 Patient 1           7     9            13          4      9
#> 10 2018-04-28 Patient 1           7    10            13          4     10
#>       case_l epid_s    case_s
#> 1       Case      1      Case
#> 2  Duplicate      1 Duplicate
#> 3  Duplicate      1 Duplicate
#> 4  Duplicate      1 Duplicate
#> 5  Duplicate      1 Duplicate
#> 6       Case      6      Case
#> 7       Case      6 Duplicate
#> 8       Case      6 Duplicate
#> 9       Case      6 Duplicate
#> 10      Case      6 Duplicate

# Note that "rolls_max" and "episodes_max" controls different aspects of episode grouping hence their behaviours are not always interchangeble.
# For instance, the example below will gives different result when a "episodes_max" and "rolls_max" of 2

data_5 <- mutate(data_4, recurrence=3)
# 13 day episode length
cbind(data_5,
      select(episode_group(data_5, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, episode_type ="rolling", rc_episode_length = recurrence, episodes_max = 2, display = FALSE), -sn, epid_l=epid, case_l=case_nm),
      select(episode_group(data_5, sn=rd_id, strata = pid, date = date, episode_length = episode_len_s, episode_type ="rolling", rc_episode_length = recurrence, rolls_max = 2,  display = FALSE), -sn, epid_s=epid, case_s=case_nm)
      )

#> [1] "Episode grouping complete"
#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id episode_len_s recurrence epid_l
#> 1  2018-04-01 Patient 1           7     1            13          3      1
#> 2  2018-04-04 Patient 1           7     2            13          3      1
#> 3  2018-04-07 Patient 1           7     3            13          3      1
#> 4  2018-04-10 Patient 1           7     4            13          3      1
#> 5  2018-04-13 Patient 1           7     5            13          3      1
#> 6  2018-04-16 Patient 1           7     6            13          3      6
#> 7  2018-04-19 Patient 1           7     7            13          3      6
#> 8  2018-04-22 Patient 1           7     8            13          3      6
#> 9  2018-04-25 Patient 1           7     9            13          3      6
#> 10 2018-04-28 Patient 1           7    10            13          3      6
#>       case_l epid_s    case_s
#> 1       Case      1      Case
#> 2  Duplicate      1 Duplicate
#> 3  Duplicate      1 Duplicate
#> 4  Duplicate      1 Duplicate
#> 5  Duplicate      1 Duplicate
#> 6       Case      6      Case
#> 7  Duplicate      6 Duplicate
#> 8  Duplicate      6 Duplicate
#> 9  Duplicate      6 Duplicate
#> 10 Duplicate      6 Duplicate

# While the one below will not

# 7 day episode length
cbind(data_5,
      select(episode_group(data_5, sn=rd_id, strata = pid, date = date, episode_length = episode_len, episode_type ="rolling", rc_episode_length = recurrence, episodes_max = 2, display = FALSE), -sn, epid_l=epid, case_l=case_nm),
      select(episode_group(data_5, sn=rd_id, strata = pid, date = date, episode_length = episode_len, episode_type ="rolling", rc_episode_length = recurrence, rolls_max = 2,  display = FALSE), -sn, epid_s=epid, case_s=case_nm)
      )

#> [1] "Episode grouping complete"
#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id episode_len_s recurrence epid_l
#> 1  2018-04-01 Patient 1           7     1            13          3      1
#> 2  2018-04-04 Patient 1           7     2            13          3      1
#> 3  2018-04-07 Patient 1           7     3            13          3      1
#> 4  2018-04-10 Patient 1           7     4            13          3      4
#> 5  2018-04-13 Patient 1           7     5            13          3      4
#> 6  2018-04-16 Patient 1           7     6            13          3      4
#> 7  2018-04-19 Patient 1           7     7            13          3      7
#> 8  2018-04-22 Patient 1           7     8            13          3      8
#> 9  2018-04-25 Patient 1           7     9            13          3      9
#> 10 2018-04-28 Patient 1           7    10            13          3     10
#>       case_l epid_s    case_s
#> 1       Case      1      Case
#> 2  Duplicate      1 Duplicate
#> 3  Duplicate      1 Duplicate
#> 4       Case      4      Case
#> 5  Duplicate      4 Duplicate
#> 6  Duplicate      4 Duplicate
#> 7       Case      7      Case
#> 8       Case      7 Duplicate
#> 9       Case      7 Duplicate
#> 10      Case     10      Case

# If the dataframe contains multiple datasets use the "data_source" argument to show which datasets have been grouped into particular epsiodes
data_6 <- data_5
data_6$dataset <- paste("DS",c(1:3, rep(c(1:2),2), rep(3,3)), sep="")

cbind(data_6,
      select(episode_group(data_6, sn=rd_id, strata = pid, date = date, episode_length = episode_len, episode_type ="rolling", rc_episode_length = recurrence, data_source = dataset, display = FALSE), -sn, epid_l=epid, case_l=case_nm, epid_grp_l=epid_grp)
)

#> [1] "Episode grouping complete"

#>          date       pid episode_len rd_id episode_len_s recurrence dataset
#> 1  2018-04-01 Patient 1           7     1            13          3     DS1
#> 2  2018-04-04 Patient 1           7     2            13          3     DS2
#> 3  2018-04-07 Patient 1           7     3            13          3     DS3
#> 4  2018-04-10 Patient 1           7     4            13          3     DS1
#> 5  2018-04-13 Patient 1           7     5            13          3     DS2
#> 6  2018-04-16 Patient 1           7     6            13          3     DS1
#> 7  2018-04-19 Patient 1           7     7            13          3     DS2
#> 8  2018-04-22 Patient 1           7     8            13          3     DS3
#> 9  2018-04-25 Patient 1           7     9            13          3     DS3
#> 10 2018-04-28 Patient 1           7    10            13          3     DS3
#>    epid_l    case_l  epid_grp_l
#> 1       1      Case DS1,DS2,DS3
#> 2       1 Duplicate DS1,DS2,DS3
#> 3       1 Duplicate DS1,DS2,DS3
#> 4       4      Case     DS1,DS2
#> 5       4 Duplicate     DS1,DS2
#> 6       4 Duplicate     DS1,DS2
#> 7       7      Case     DS2,DS3
#> 8       7 Duplicate     DS2,DS3
#> 9       7 Duplicate     DS2,DS3
#> 10     10      Case         DS3