Link events to chronological episodes.

Link events with matching attributes and within specified durations of each other. Each set of linked records are assigned a unique identifier with relevant group-level information.

episodes(
  date,
  case_length = Inf,
  episode_type = "fixed",
  recurrence_length = case_length,
  episode_unit = "days",
  strata = NULL,
  sn = NULL,
  episodes_max = Inf,
  rolls_max = Inf,
  case_overlap_methods = 8,
  recurrence_overlap_methods = case_overlap_methods,
  skip_if_b4_lengths = FALSE,
  data_source = NULL,
  data_links = "ANY",
  custom_sort = NULL,
  skip_order = Inf,
  reference_event = "last_record",
  case_for_recurrence = FALSE,
  from_last = FALSE,
  group_stats = FALSE,
  display = "none",
  case_sub_criteria = NULL,
  recurrence_sub_criteria = case_sub_criteria,
  case_length_total = 1,
  recurrence_length_total = case_length_total
)

episodes_wf_splits(
  date,
  case_length = Inf,
  episode_type = "fixed",
  recurrence_length = case_length,
  episode_unit = "days",
  strata = NULL,
  sn = NULL,
  episodes_max = Inf,
  rolls_max = Inf,
  case_overlap_methods = 8,
  recurrence_overlap_methods = case_overlap_methods,
  skip_if_b4_lengths = FALSE,
  data_source = NULL,
  data_links = "ANY",
  custom_sort = NULL,
  skip_order = Inf,
  reference_event = "last_record",
  case_for_recurrence = FALSE,
  from_last = FALSE,
  group_stats = FALSE,
  display = "none",
  case_sub_criteria = NULL,
  recurrence_sub_criteria = case_sub_criteria,
  case_length_total = 1,
  recurrence_length_total = case_length_total
)

fixed_episodes(
  date,
  case_length = Inf,
  episode_unit = "days",
  to_s4 = TRUE,
  case_overlap_methods = 8,
  deduplicate = FALSE,
  display = "none",
  bi_direction = FALSE,
  recurrence_length = case_length,
  recurrence_overlap_methods = case_overlap_methods,
  include_index_period = TRUE,
  ...,
  overlap_methods = 8,
  overlap_method = 8,
  x
)

rolling_episodes(
  date,
  case_length = Inf,
  recurrence_length = case_length,
  episode_unit = "days",
  to_s4 = TRUE,
  case_overlap_methods = 8,
  recurrence_overlap_methods = case_overlap_methods,
  deduplicate = FALSE,
  display = "none",
  bi_direction = FALSE,
  include_index_period = TRUE,
  ...,
  overlap_methods = 8,
  overlap_method = 8,
  x
)

episode_group(df, ..., episode_type = "fixed")

Arguments

date
[date|datetime|integer|number_line]. Event date or period.

case_length
[integer|number_line]. Duration from index event distinguishing one "case" from another. This is the case window.

episode_type
[character]. Options are "fixed" (default), "rolling" or "recursive". See Details.

recurrence_length
[integer|number_line]. Duration from the last "duplicate" event distinguishing a "recurrent" event from its index event. This is the recurrence window.

episode_unit
[character]. Time units for case_length and recurrence_length. Options are "seconds", "minutes", "hours", "days" (default), "weeks", "months" or "years". See diyar::episode_unit.

strata
[atomic]. Subsets of the dataset. Episodes are created separately for each strata.

sn
[integer]. Unique record identifier. Useful for creating familiar epid identifiers.

episodes_max
[integer]. The maximum number of episodes permitted within each strata.

rolls_max
[integer]. Maximum number of times the index "case" can recur. Only used if episode_type is "rolling".

case_overlap_methods
[character|integer]. Methods of overlap considered when tracking duplicates of "case" events. See (overlaps)

recurrence_overlap_methods
[character|integer]. Methods of overlap considered when tracking duplicates of "recurrent" events. See (overlaps)

skip_if_b4_lengths
[logical]. If TRUE (default), when using lagged case_length or recurrence_length, events before the cut-off point or period are skipped.

data_source
[character]. Unique data source identifier. Adds the list of datasets in each episode to the epid. Useful when the dataset has data from multiple sources.

data_links
[list|character]. A set of data_sources required in each epid. An episode without records from these data_sources will be unlinked. See Details.

custom_sort
[atomic]. Preferential order for selecting index or reference events.

skip_order
[integer]. "nth" level of custom_sort. Episodes with index events beyond this level of preference are skipped.

reference_event
[character]. Specifies which of the duplicates are used as reference events for subsequent windows. Options are "last_record" (default), "last_event", "first_record" or ""firs_event".

case_for_recurrence
[logical]. If TRUE, both "case" and "recurrent" events will have a case window. If FALSE (default), only case events will have a case window. Only used if episode_type is "rolling".

from_last
[logical]. Chronological order of episode tracking i.e. ascending (TRUE) or descending (FALSE).

group_stats
[logical]. If TRUE (default), episode-specific information like episode start and end dates are returned.

display
[character]. The progress messages printed on screen. Options are; "none" (default), "progress" and "stats".

case_sub_criteria
[sub_criteria]. Matching conditions for "case" windows in addition to temporal links.

recurrence_sub_criteria
[sub_criteria]. Matching conditions for "recurrence" windows in addition to temporal links.

case_length_total
[integer|number_line]. Minimum number of matched case windows required for an episode.

recurrence_length_total
[integer|number_line]. Minimum number of matched recurrence windows required for an episode.

to_s4
[logical]. Deprecated. Output type - epid (TRUE) or data.frame (FALSE).

deduplicate
[logical]. Deprecated. If TRUE, "duplicate" events are excluded from the epid.

bi_direction
[logical]. Deprecated. If TRUE, "duplicate" events before and after the index event are tracked.

include_index_period
[logical]. Deprecated. If TRUE, events overlapping with the index event or period are linked even if they are outside the cut-off period.

...
Arguments passed to episodes.

overlap_methods
[character]. Deprecated. Please use case_overlap_methods or recurrence_overlap_methods. Methods of overlap considered when tracking duplicate event. See (overlaps)

overlap_method
[character]. Deprecated. Please use case_overlap_methods or recurrence_overlap_methods. Methods of overlap considered when tracking event. All event are checked by the same set of overlap_method.

x
[date|datetime|integer|number_line]. Deprecated. Record date or period. Please use date.

df
[data.frame]. Deprecated. One or more datasets appended together. See Details.

Value

epid

Details

All dated records within a specified duration of an index record are linked together as an episode. By default, this process occurs in ascending order, beginning with the earliest event. This can be changed to a descending (from_last) or custom order (custom_sort). Ties are always broken by the chronological order of events.

A "fixed" episode has a fixed maximum duration determined by case_length, while a "rolling" episode can continue to recur. A "rolling" episode will persist as long as is specified by rolls_max.

episodes() will categorise records into 5 type of events;

"Case" - Index event of the episode.

"Duplicate_C" - Duplicate of the index event.

"Recurrent" - Recurrence of the index event.

"Duplicate_R" - Duplicate of the recurrent event.

"Skipped" - Records excluded from the episode tracking process.

Every element in data_links must be named "l" (links) or "g" (groups). Unnamed elements of data_links will be assumed to be "l".

If named "l", only groups with records from every listed data_source will be unlinked.

If named "g", only groups with records from any listed data_source will be unlinked.

NA values in strata excludes records from the episode tracking process

episodes_wf_splits() is a wrapper function of episode_group() which reframes the data before the episode tracking process. This aims to reduce the overall processing time for the tracking process. See vignette("episodes") for further details.

episode_group() has been retired. It now only exists to support previous code with minimal input from users. Moving forward, please use episodes().

rolling_episodes() and rolling_episodes() are convenience functions to support previous code with minimal input from users. Moving forward, please use episodes().

See vignette("episodes") for more information.

See also

Examples

data(infections)
data(hospital_admissions)

db_1 <- infections
db_1$patient_id <- c(rep("PID 1",8), rep("PID 2",3))

# Fixed episodes
# One 16-day (15-day difference) episode per patient
db_1$epids_p <- episodes(date = db_1$date,
                         strata = db_1$patient_id,
                         case_length = 15,
                         episodes_max = 1)
# Rolling episodes
# 16-day episodes with recurrence periods of 11 days
db_1$rd_b <- episodes(date = db_1$date,
                     case_length = 15,
                     recurrence_length = 10,
                     episode_type = "rolling")

# Interval grouping
hospital_admissions$admin_period <- number_line(hospital_admissions$admin_dt,
                                                hospital_admissions$discharge_dt)
admissions <- hospital_admissions[c("admin_period","epi_len")]

# Episodes of overlapping periods of admission
hospital_admissions$epids_i <- episodes(date = hospital_admissions$admin_period,
                                       case_length = 0,
                                       case_overlap_methods = "inbetween")