Multistage deterministic record linkage

Link records with matching criteria in ordered stages of relevance. Each set of linked records are assigned a unique identifier with relevant group-level information.

links(
  criteria,
  sub_criteria = NULL,
  sn = NULL,
  strata = NULL,
  data_source = NULL,
  data_links = "ANY",
  display = "none",
  group_stats = FALSE,
  expand = TRUE,
  shrink = FALSE,
  recursive = TRUE,
  check_duplicates = FALSE
)

record_group(df, ..., to_s4 = TRUE)

Arguments

criteria
[list|atomic]. Attributes to compare. Each element of the list is a stage in the linkage process. NA values in criteria excludes records from the corresponding stage of the linkage process. See Details.

sub_criteria
[list|sub_criteria]. Additional matching condition for each stage of the linkage process. See sub_criteria

sn
[integer]. Unique record identifier. Useful for creating familiar pid identifiers.

strata
[atomic]. Subsets of the dataset. Record groups are created separately for each strata. NA values in strata excludes records from the entire linkage process.

data_source
[character]. Unique data source identifier. Adds the list of data sources in each record group to the pid. Useful when the dataset has data from multiple sources.

data_links
[list|character]. A set of data_sources required in each pid. A pid without records from these data_sources will be unlinked. See Details.

display
[character]. Progress messages printed on screen. Options are; "none" (default) or, "progress" and "stats" for a progress update or a more detailed breakdown of the linkage process.

group_stats
[logical]. If TRUE (default), returns group-specific information like record counts for each pid.

expand
[logical]. If TRUE, allows a record group to expand with subsequent stages of the linkage process. Not interchangeable with shrink.

shrink
[logical]. If TRUE, forces a record group to shrink with each subsequent stage of the linkage process. Not interchangeable with expand.

df
[data.frame]. One or more datasets appended together. See Details.

to_s4
[logical]. Output type - pid (TRUE) or data.frame (FALSE).

Value

pid

Details

Match priority decreases with each subsequent stage of linkage i.e. earlier stages (criteria) are considered superior. Therefore, it's important for each criteria to be listed in an order of decreasing relevance.

Records with missing criteria (NA) are skipped at each stage of the linkage process, while records with missing strata (NA) are skipped from the entire linkage process.

If a record is skipped, another attempt will be made to match the record at the next stage. If a record does not match any other record by the end of the linkage process (or it has a missing strata), it is assigned a unique pid.

A sub_criteria can be used to request additional matching conditions for each stage of the linkage process. When used, only records with matching criteria and sub_criteria are linked.

In links, each sub_criteria must be linked to a criteria. This is done by naming creating a list of the sub_criteria. Each name must correspond to a stage. See below for an example of 3 sub_criteria linked to criteria 1, 5 and 13.

For example;

$$list("cr1" = sub_criteria, "cr5" = sub_criteria, "cr13" = sub_criteria).$$

Multiple matching criteria can be supplied as a nested sub_criteria. The same sub_criteria or nested sub_criteria can be linked to different criteria. Any unlinked sub_criteria will be ignored.

By default, attributes in a sub_criteria are compared for an exact_match. However, user-defined logical tests (function) are also permitted. Such tests must meet 3 requirements:

It must be able to compare two atomic vectors

It must have two arguments named `x` and `y`, where `y` is the value for one observation being compared against all other observations (`x`).

It must return either TRUE or FALSE

Every element in data_links must be named "l" (links) or "g" (groups). Unnamed elements of data_links will be assumed to be "l".

If named "l", only groups with records from every listed data_source will be retained.

If named "g", only groups with records from any listed data_source will be retained.

record_group() as it existed before v0.2.0 has been retired. It now exists to support previous code and arguments with minimal input from users. Moving forward, please use links().

See vignette("links") for more information.

See also

Examples

# Exact match
links(criteria = c("Obinna","James","Ojay","James","Obinna"))

#> [1] "P.1 (CRI 001)" "P.2 (CRI 001)" "P.3 (No hits)" "P.2 (CRI 001)"
#> [5] "P.1 (CRI 001)"

# User-defined tests using `sub_criteria()`
# Matching `sex` and a 20-year age range
age <- c(30, 28, 40, 25, 25, 29, 27)
sex <- c("M", "M", "M", "F", "M", "M", "F")
f1 <- function(x, y) (y - x) %in% 0:20
links(criteria = sex,
      sub_criteria = list(cr1 = sub_criteria(age, funcs = f1)))

#> Error: Invalid object type for `...` in `sub_criteria()`:
#> i - `...` must be an `atomic` vector OR,
#> i - `...` must be a `list` of `atomic`  vectors OR,
#> i - `...` must be a `sub_criteria` object.
#> X - `...` contains is a `function` object.

# Multistage linkage
# Relevance of matches: `forename` > `surname`
data(staff_records); staff_records

#>   r_id forename  surname sex    dataset
#> 1    1    James    Green   M Staff list
#> 2    2     <NA> Anderson   M Staff list
#> 3    3    Jamey    Green   M  Pay slips
#> 4    4              <NA>   F  Pay slips
#> 5    5  Derrick Anderson   M Staff list
#> 6    6  Darrack Anderson   M  Pay slips
#> 7    7 Christie    Green   F Staff list

links(criteria = list(staff_records$forename, staff_records$surname),
      data_source = staff_records$sex)

#> [1] "P.1 (CRI 002)" "P.2 (CRI 002)" "P.1 (CRI 002)" "P.4 (No hits)"
#> [5] "P.2 (CRI 002)" "P.2 (CRI 002)" "P.1 (CRI 002)"

# Relevance of matches:
# `staff_id` > `age` (AND (`initials`, `hair_colour` OR `branch_office`))
data(missing_staff_id); missing_staff_id

#>   r_id staff_id age initials hair_colour     branch_office source_1 source_2
#> 1    1       NA  30     G.D.       Brown Republic of Ghana        A        3
#> 2    2       NA  30     B.G.        Teal            France        A        1
#> 3    3       NA  30     X.P.        <NA>              <NA>        A        1
#> 4    4       NA  30     X.P.       Green              <NA>        B        1
#> 5    5       NA  30     <NA>       Green            France        A        1
#> 6    6        2  30     G.D.  Dark brown             Ghana        A        1
#> 7    7        2  30     G.D.       Brown Republic of Ghana        B        2

links(criteria = list(missing_staff_id$staff_id, missing_staff_id$age),
      sub_criteria = list(cr2 = sub_criteria(missing_staff_id$initials,
                                          missing_staff_id$hair_colour,
                                          missing_staff_id$branch_office)),
      data_source = missing_staff_id$source_1)

#> [1] "P.6 (CRI 002)" "P.2 (CRI 002)" "P.2 (CRI 002)" "P.2 (CRI 002)"
#> [5] "P.2 (CRI 002)" "P.6 (CRI 001)" "P.6 (CRI 001)"

# Group expansion or shrinkage
match_cri <- list(c(1,NA,NA,1,NA,NA),
                  c(1,1,1,2,2,2),
                  c(3,3,3,2,2,2))
links(criteria = match_cri, expand = TRUE)

#> [1] "P.1 (CRI 001)" "P.1 (CRI 002)" "P.1 (CRI 002)" "P.1 (CRI 001)"
#> [5] "P.1 (CRI 002)" "P.1 (CRI 002)"

links(criteria = match_cri, expand = FALSE)

#> [1] "P.1 (CRI 001)" "P.2 (CRI 002)" "P.3 (CRI 002)" "P.1 (CRI 001)"
#> [5] "P.5 (CRI 002)" "P.6 (CRI 002)"

links(criteria = match_cri, shrink = TRUE)

#> [1] "P.1 (CRI 001)" "P.2 (CRI 003)" "P.2 (CRI 003)" "P.1 (CRI 001)"
#> [5] "P.5 (CRI 003)" "P.5 (CRI 003)"