number_line

A range of numeric values.

number_line(l, r, id = NULL, gid = NULL)

as.number_line(x)

is.number_line(x)

left_point(x)

left_point(x) <- value

right_point(x)

right_point(x) <- value

start_point(x)

start_point(x) <- value

end_point(x)

end_point(x) <- value

number_line_width(x)

reverse_number_line(x, direction = "both")

shift_number_line(x, by = 1)

expand_number_line(x, by = 1, point = "both")

invert_number_line(x, point = "both")

compress_number_line(
  x,
  methods = "overlap",
  collapse = FALSE,
  deduplicate = TRUE,
  method = "overlap"
)

number_line_sequence(
  x,
  by = NULL,
  length.out = 1,
  fill = TRUE,
  simplify = FALSE
)

Arguments

l
[numeric based]. Left point of the number_line. Must be able to be coerced to a numeric object

r
[numeric based]. Right point of the number_line. Must be able to be coerced to a numeric object

id
[integer]. Unique element identifier. Optional

gid
[integer]. Unique group identifier. Optional

x
[number_line]

value
[numeric based]

direction
[character]. Type of "number_line" objects to be reversed. Options are; "increasing", "decreasing" or "both" (default).

by
[integer]. Increment or decrement. Passed to seq() in number_line_sequence()

point
[character]. "start", "end", "left" or "right" point.

methods
[character]. Methods of overlap. Check different pairs of number_line by different methods

collapse
[lgocial]. If TRUE, collapse the compressed results yet again.

deduplicate
[lgocial]. If TRUE, retains only one number_line per overlapping set

method
[character]. Method of overlap. Check every pair of number_line by the same method. Deprecated. Please use methods instead.

length.out
[integer]. Number of splits. For example, 1 for two parts or 2 for three parts. Passed to seq()

fill
[logical]. Retain (TRUE) or drop (FALSE) the remainder of an uneven split

simplify
[logical]. Split into number_line or sequence finite numbers

Value

number_line

Details

A number_line represents a range of numbers on a number line.

It's made up of a start and end point which are the beginning and end of the range respectively. The location of the start point - left or right, determines whether it's an "increasing" or "decreasing" range. This is the direction of the number_line.

reverse_number_line() - reverses the direction of a number_line. A reversed number_line has its l and r points swapped. The direction argument specifies which type of number_line will be reversed. number_line with non-finite starts or end points i.e. (NA, NaN and Inf) can't be reversed.

shift_number_line() - Shift a number_line towards the positive or negative end of the number line.

expand_number_line() - Increase or decrease the width or length of a number_line.

invert_number_line() - Invert the left and/or right points from a negative to positive value or vice versa.

compress_number_line() - "compress" or "collapse" overlapping number_line into a new number_line that covers the start and end points of the originals. This results in duplicate number_line with the start and end points of the newly expanded number_line See overlaps for further details on overlapping number_line. compress_number_line can be an alternative for simple implementations of links or episodes

number_line_sequence() - Split a number_line into equal parts (length.out) or by a fixed recurring width (by).

See also

Examples

date <- function(x) as.Date(x, "%d/%m/%Y")
dttm <- function(x) as.POSIXct(x, "UTC", format = "%d/%m/%Y %H:%M:%S")

number_line(-100, 100)

#> [1] "-100 -> 100"

# Also compatible with other numeric based object classes
number_line(dttm("15/05/2019 13:15:07"), dttm("15/05/2019 15:17:10"))

#> [1] "2019-05-15 13:15:07 -> 2019-05-15 15:17:10"

# Coerce applicable object classes to `number_line` objects
as.number_line(5.1); as.number_line(date("21/10/2019"))

#> [1] "5.1 == 5.1"

#> [1] "2019-10-21 == 2019-10-21"

# A test for number_line objects
a <- number_line(date("25/04/2019"), date("01/01/2019"))
is.number_line(a)

#> [1] TRUE

# Structure of a number_line object
left_point(a); right_point(a); start_point(a); end_point(a)

#> [1] "2019-04-25"

#> [1] "2019-01-01"

#> [1] "2019-01-01"

#> [1] "2019-04-25"

# Reverse number_line objects
reverse_number_line(number_line(date("25/04/2019"), date("01/01/2019")))

#> [1] "2019-01-01 -> 2019-04-25"

reverse_number_line(number_line(200, -100), "increasing")

#> [1] "200 <- -100"

reverse_number_line(number_line(200, -100), "decreasing")

#> [1] "-100 -> 200"

c <- number_line(5, 6)
# Shift number_line objects towards the positive end of the number line
shift_number_line(x = c(c, c), by = c(2, 3))

#> [1] "7 -> 8" "8 -> 9"

# Shift number_line objects towards the negative end of the number line
shift_number_line(x = c(c, c), by = c(-2, -3))

#> [1] "3 -> 4" "2 -> 3"

# Change the duration, width or length of a number_line object
d <- c(number_line(3, 6), number_line(6, 3))

expand_number_line(d, 2)

#> [1] "1 -> 8" "8 <- 1"

expand_number_line(d, -2)

#> [1] "5 <- 4" "4 -> 5"

expand_number_line(d, c(2,-1))

#> [1] "1 -> 8" "5 <- 4"

expand_number_line(d, 2, "start")

#> [1] "1 -> 6" "6 <- 5"

expand_number_line(d, 2, "end")

#> [1] "3 -> 8" "8 <- 3"

# Invert `number_line` objects
e <- c(number_line(3, 6), number_line(-3, -6), number_line(-3, 6))
e

#> [1] "3 -> 6"   "-3 <- -6" "-3 -> 6" 

invert_number_line(e)

#> [1] "-3 <- -6" "3 -> 6"   "3 <- -6" 

invert_number_line(e, "start")

#> [1] "-3 -> 6" "-3 -> 6" "3 -> 6" 

invert_number_line(e, "end")

#> [1] "3 <- -6"  "3 <- -6"  "-3 <- -6"

# Collapse `number_line` objects
x <- c(number_line(10,10), number_line(10,20), number_line(5,30),  number_line(30,40))
compress_number_line(x, deduplicate = FALSE)

#> [1] "5 -> 30"  "5 -> 30"  "5 -> 30"  "30 -> 40"

compress_number_line(x)

#> [1] "5 -> 30"  "30 -> 40"

compress_number_line(x, collapse=TRUE)

#> [1] "5 -> 40"

compress_number_line(x, collapse=TRUE, methods = "inbetween")

#> [1] "5 -> 30"  "30 -> 40"

# Split number line objects
x <- number_line(Sys.Date() - 5, Sys.Date())
x

#> [1] "2021-04-02 -> 2021-04-07"

number_line_sequence(x, by = 2)

#> [1] "2021-04-02 -> 2021-04-04" "2021-04-04 -> 2021-04-06"
#> [3] "2021-04-06 -> 2021-04-07"

number_line_sequence(x, by = 4)

#> [1] "2021-04-02 -> 2021-04-06" "2021-04-06 -> 2021-04-07"

number_line_sequence(x, by = 4, fill = FALSE)

#> [1] "2021-04-02 -> 2021-04-06"

number_line_sequence(x, length.out = 2)

#> [1] "2021-04-02 -> 2021-04-04" "2021-04-04 -> 2021-04-07"