Number line objects

A set of functions to create and manipulate number_line objects.

number_line(l, r, id = NULL, gid = NULL)

as.number_line(x)

is.number_line(x)

left_point(x)

right_point(x)

start_point(x)

end_point(x)

number_line_width(x)

reverse_number_line(x, direction = "both")

shift_number_line(x, by = 1)

expand_number_line(x, by = 1, point = "both")

compress_number_line(x, method = c("across", "chain", "aligns_start",
  "aligns_end", "inbetween"), collapse = FALSE, deduplicate = TRUE)

number_line_sequence(x, by = 1)

Arguments

l
Left point of the number_line object. Should be, or can be coerced to a numeric object

r
Right point of the number_line object. Should be, or can be coerced to a numeric object

id
Unique numeric ID. Providing this is optional

gid
Unique numeric Group ID. Providing this is optional

x
number_line object

direction
Type of "number_line" objects whose direction are to be reversed. Options are; "increasing", "decreasing" or "both".

by
increment or decrement

point
"start" or "start" point

method
Method of overlap

collapse
If TRUE, collapse the compressed results based on method of overlaps

deduplicate
if TRUE, retains only one number_line object among duplicates

Value

number_line object

Details

A number_line object represents a series of real numbers on a number line.

Visually, it's presented as the left (l) and right (r) points of the series. This may differ from start and end points. The start point is the lowest number in the series, regardless of whether it's at the left or right point..

The location of the start point - left or right, indicate if it's an "increasing" or "decreasing" series. This is referred to as the direction of the number_line object.

reverse_number_line() - reverses the direction of a number_line object. A reversed number_line object has its l and r points swapped but maintains the same width or length. The direction argument determines which type of number_line objects will be reversed. number_line objects with non-finite numeric starts or end points i.e. (NA, NaN and Inf) can't be reversed.

shift_number_line() - a convenience function to shift a number_line object towards the positive or negative end of the number line.

expand_number_line() - a convenience function to increase or decrease the width or length of a number_line object.

compress_number_line() - Collapses overlapping number_line objects into a new number_line objects that covers the start and end points of the originals. This results in duplicate number_line objects with start and end points of the new expanded number_line object. See overlap for further details on overlapping number_line objects. If a familiar (but unique) id is used when creating the number_line objects, compress_number_line() can be a simple alternative to record_group or episode_group.

number_line_sequence() - a convenience function to convert a number_line object into a sequence of finite numbers. The sequence will also include the start and end points. The direction of the sequence will correspond to that of the number_line object.

Examples

library(lubridate)

number_line(-100, 100); number_line(10, 11.2)

#> [1] "-100 -> 100"

#> [1] "10 -> 11.2"

# Other numeric based object classes are also compatible for numeric_line objects
number_line(dmy_hms("15/05/2019 13:15:07"), dmy_hms("15/05/2019 15:17:10"))

#> [1] "2019-05-15 13:15:07 -> 2019-05-15 15:17:10"

# A warning is given if 'l' and 'r' have different classes. Consider if these need to be corrected
number_line(2, dmy("05/01/2019"))

#> Warning: 'l' and 'r' have different classes. It may need to be reconciled

#> [1] "2 -> 17901"

# Convert numeric based objects to number_line objects
as.number_line(5.1); as.number_line(dmy("21/10/2019"))

#> [1] "5.1 == 5.1"

#> [1] "2019-10-21 == 2019-10-21"

# Test for number_line objects
a <- number_line(0, -100)
b <- number_line(dmy("25/04/2019"), dmy("01/01/2019"))
is.number_line(a); is.number_line(b)

#> [1] TRUE

#> [1] TRUE

# Structure of a number_line object
left_point(a); right_point(a); start_point(a); end_point(a)

#> [1] 0

#> [1] -100

#> [1] -100

#> [1] 0

# Reverse number_line objects
reverse_number_line(number_line(dmy("25/04/2019"), dmy("01/01/2019")))

#> [1] "2019-01-01 -> 2019-04-25"

reverse_number_line(number_line(200,-100), "increasing")

#> [1] "200 <- -100"

reverse_number_line(number_line(200,-100), "decreasing")

#> [1] "-100 -> 200"

# Shift number_line objects
number_line(5,6)

#> [1] "5 -> 6"

# Towards the positive end of the number line
shift_number_line(number_line(5,6), 2)

#> [1] "7 -> 8"

# Towards the negative end of the number line
shift_number_line(number_line(6,1), -2)

#> [1] "4 <- -1"

# Increase or reduce the width or length of a \code{number_line} object
c(number_line(3,6), number_line(6,3))

#> [1] "3 -> 6" "6 <- 3"

expand_number_line(c(number_line(3,6), number_line(6,3)), 2)

#> [1] "1 -> 8" "8 <- 1"

expand_number_line(c(number_line(3,6), number_line(6,3)), -1)

#> [1] "4 -> 5" "5 <- 4"

expand_number_line(c(number_line(3,6), number_line(6,3)), 2, "start")

#> [1] "1 -> 6" "8 <- 3"

expand_number_line(c(number_line(3,6), number_line(6,3)), -2, "end")

#> [1] "3 -> 4" "6 <- 5"

# Collapse number line objects
x <- c(number_line(10,10), number_line(10,20), number_line(5,30),  number_line(30,40))
compress_number_line(x, deduplicate = FALSE)

#> [1] "5 -> 30"  "5 -> 30"  "5 -> 30"  "30 -> 40"

compress_number_line(x)

#> [1] "5 -> 30"  "30 -> 40"

compress_number_line(x, collapse=TRUE)

#> [1] "5 -> 40"

compress_number_line(x, collapse=TRUE, method = "inbetween")

#> [1] "5 -> 30"  "30 -> 40"

# Convert a number line object to its series of real numbers
number_line_sequence(number_line(1, 5))

#> [1] 1 2 3 4 5

number_line_sequence(number_line(5, 1), .5)

#> [1] 5.0 4.5 4.0 3.5 3.0 2.5 2.0 1.5 1.0

number_line_sequence(number_line(dmy("01/04/2019"), dmy("10/04/2019")), 1)

#>  [1] "2019-04-01" "2019-04-02" "2019-04-03" "2019-04-04" "2019-04-05"
#>  [6] "2019-04-06" "2019-04-07" "2019-04-08" "2019-04-09" "2019-04-10"

# The length of the vector depends on the object class
number_line_sequence(number_line(dmy("01/04/2019"), dmy("04/04/2019")), 1.5)

#> [1] "2019-04-01" "2019-04-02" "2019-04-04"

nl <- number_line(dmy_hms("01/04/2019 00:00:00"), dmy_hms("04/04/2019 00:00:00"))
head(number_line_sequence(nl, 1.5), 15)

#>  [1] "2019-04-01 00:00:00 UTC" "2019-04-01 00:00:01 UTC"
#>  [3] "2019-04-01 00:00:03 UTC" "2019-04-01 00:00:04 UTC"
#>  [5] "2019-04-01 00:00:06 UTC" "2019-04-01 00:00:07 UTC"
#>  [7] "2019-04-01 00:00:09 UTC" "2019-04-01 00:00:10 UTC"
#>  [9] "2019-04-01 00:00:12 UTC" "2019-04-01 00:00:13 UTC"
#> [11] "2019-04-01 00:00:15 UTC" "2019-04-01 00:00:16 UTC"
#> [13] "2019-04-01 00:00:18 UTC" "2019-04-01 00:00:19 UTC"
#> [15] "2019-04-01 00:00:21 UTC"

d <- duration(1.5,"days")
number_line_sequence(nl, d)

#> [1] "2019-04-01 00:00:00 UTC" "2019-04-02 12:00:00 UTC"
#> [3] "2019-04-04 00:00:00 UTC"