Multi-staged deterministic record linkage

This function assigns unique group identifiers to matching records from one or more datasets.

record_group(df, sn, criteria, sub_criteria = NULL, data_source = NULL,
  group_stats = FALSE, display = TRUE)

Arguments

df
Dataframe. One or more datasets appened together.

sn
Unique numeric record indentifier.

criteria
Column names of the attributes to match. Records with matching values in these columns are grouped together.

sub_criteria
Matching sub-criteria. Additional matching conditions for each stage (criteria).

data_source
Unique dataset indentifier. Useful when dataframe contains data from multiple datasets.

group_stats
If TRUE, output will include additional columns with useful stats for each record group.

display
If TRUE, status messages are not printed on screen.

Value

Dataframe

sn - unique record identifier as provided

pid - unique group indentifier

pid_cri - matching criteria for each record in the group

pid_dataset - list of datasets in each group

pid_total - number of records in each group

Details

Record grouping occurs in stages as listed in criteria. A match at each stage is considered more certain than those in subsequent stages. Therefore, criteria should be listed in order of decreasing certainty. sub_criteria can be used to force additonal matching conditions at each stage. If sub_criteria is not NULL, only records with matching criteria and sub_criteria values are grouped together. If a record has a missing values for any criteria, it's skipped at that stage, and another attempt is made at the next stage. If all criteria values are missing, that record is assigned a unique group ID. When a data_source identifier is included, pid_dataset is included in the output. This shows the datasets included in each group.

See also

Examples

 library(dplyr)
 library(tidyr)

# One stage record grouping
 df <- tibble(
   r_id = c(1:5),
   forename = c("Obinna","James","Ojay","James","Obinna"),
   stringsAsFactors = FALSE
 )

 bind_cols(df, record_group(df, r_id, forename))

#> 
#> Group criteria 1 - `forename`
#> 5 of 5 record(s) have been assigned a group ID. 0 records not yet grouped.
#> 1 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 1.
#> 
#> Record grouping complete - 1 record(s) assigned a unique ID. 

#> # A tibble: 5 x 6
#>    r_id forename stringsAsFactors    sn   pid pid_cri   
#>   <int> <chr>    <lgl>            <int> <dbl> <chr>     
#> 1     1 Obinna   FALSE                1     1 Criteria 1
#> 2     2 James    FALSE                2     2 Criteria 1
#> 3     3 Ojay     FALSE                3     3 None      
#> 4     4 James    FALSE                4     2 Criteria 1
#> 5     5 Obinna   FALSE                5     1 Criteria 1

 # Handling missing or unknown data - Recode missing or unknown values to NA or "".
 # These are excluded from record grouping at the relevant stage

 df %>%
   mutate(forename = ifelse(r_id %in% c(1,4), NA, forename)) %>%
   bind_cols(record_group(., r_id, forename))

#> 
#> Group criteria 1 - `forename`
#> 3 of 5 record(s) have been assigned a group ID. 2 records not yet grouped.
#> 3 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Record grouping complete - 5 record(s) assigned a unique ID. 

#> # A tibble: 5 x 6
#>    r_id forename stringsAsFactors    sn   pid pid_cri
#>   <int> <chr>    <lgl>            <int> <int> <chr>  
#> 1     1 NA       FALSE                1     1 None   
#> 2     2 James    FALSE                2     2 None   
#> 3     3 Ojay     FALSE                3     3 None   
#> 4     4 NA       FALSE                4     4 None   
#> 5     5 Obinna   FALSE                5     5 None   

 # Two or more stages of record grouping
 # Criteria should be listed in decreasing order of certainty
data_3 <- tibble(
  r_id = c(1:7),
  forename = c("James",NA,"Jamey","","Derrick","Darrack","Christie"),
  surname = c("Green","Anderson","Green",NA,"Anderson","Anderson","Green"),
  sex = c("M","M","M","F","M","M","F"),
  dataset = c("Staff list","Staff list", "Pay slips","Pay slips", "Staff list","Pay slips","Staff list"),
  stringsAsFactors = FALSE
)

bind_cols(data_3, record_group(data_3,r_id, c(forename, surname), data_source = sex, display = FALSE))

#> Record grouping complete - 1 record(s) assigned a unique ID. 

#> # A tibble: 7 x 10
#>    r_id forename surname sex   dataset stringsAsFactors    sn   pid pid_cri
#>   <int> <chr>    <chr>   <chr> <chr>   <lgl>            <int> <dbl> <chr>  
#> 1     1 James    Green   M     Staff ~ FALSE                1     1 Criter~
#> 2     2 NA       Anders~ M     Staff ~ FALSE                2     2 Criter~
#> 3     3 Jamey    Green   M     Pay sl~ FALSE                3     1 Criter~
#> 4     4 ""       NA      F     Pay sl~ FALSE                4     4 None   
#> 5     5 Derrick  Anders~ M     Staff ~ FALSE                5     2 Criter~
#> 6     6 Darrack  Anders~ M     Pay sl~ FALSE                6     2 Criter~
#> 7     7 Christie Green   F     Staff ~ FALSE                7     1 Criter~
#> # ... with 1 more variable: pid_dataset <chr>

# Add sex to the second stage to be more certain
data_3 %>%
  unite(cri_2, c(surname, sex)) %>%
  bind_cols(record_group(., r_id, c(forename, cri_2), data_source = dataset, display = FALSE))

#> Record grouping complete - 2 record(s) assigned a unique ID. 

#> # A tibble: 7 x 9
#>    r_id forename cri_2  dataset stringsAsFactors    sn   pid pid_cri pid_dataset
#>   <int> <chr>    <chr>  <chr>   <lgl>            <int> <dbl> <chr>   <chr>      
#> 1     1 James    Green~ Staff ~ FALSE                1     1 Criter~ Pay slips,~
#> 2     2 NA       Ander~ Staff ~ FALSE                2     2 Criter~ Pay slips,~
#> 3     3 Jamey    Green~ Pay sl~ FALSE                3     1 Criter~ Pay slips,~
#> 4     4 ""       NA_F   Pay sl~ FALSE                4     4 None    Pay slips  
#> 5     5 Derrick  Ander~ Staff ~ FALSE                5     2 Criter~ Pay slips,~
#> 6     6 Darrack  Ander~ Pay sl~ FALSE                6     2 Criter~ Pay slips,~
#> 7     7 Christie Green~ Staff ~ FALSE                7     7 None    Staff list 

# Using sub-criteria
data_4 <- tibble(
  r_id = c(1:7),
  staff_id = c(NA,NA,NA,NA,NA,2,2),
  age = rep(30,7),
  initials = c("G.D.","B.G.","X.P.","X.P.",NA,"G.D.","G.D."),
  hair_colour = c("Brown","Teal",NA,"Green","Green","Dark brown","Brown"),
  branch_office = c("Republic of Ghana","France",NA,NA,"France","Ghana","Republic of Ghana"),
  ds_1 = c("A","A","A","B","A","A","B"),
  ds_2 = c(3,1,1,1,1,1,2),
  stringsAsFactors = FALSE
)

bind_cols(data_4, record_group(data_4,r_id, c(staff_id, age), list(s2a=c("initials","hair_colour","branch_office")), data_source = ds_1))

#> 
#> Group criteria 1 - `staff_id`
#> 2 of 7 record(s) have been assigned a group ID. 5 records not yet grouped.
#> 0 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Group criteria 2 - `age`
#> Matching criteria 2: iteration 2
#> Matching criteria 2: iteration 3
#> 5 of 5 record(s) have been assigned a group ID. 0 records not yet grouped.
#> 0 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a unique ID. 

#> # A tibble: 7 x 13
#>    r_id staff_id   age initials hair_colour branch_office ds_1   ds_2
#>   <int>    <dbl> <dbl> <chr>    <chr>       <chr>         <chr> <dbl>
#> 1     1       NA    30 G.D.     Brown       Republic of ~ A         3
#> 2     2       NA    30 B.G.     Teal        France        A         1
#> 3     3       NA    30 X.P.     NA          NA            A         1
#> 4     4       NA    30 X.P.     Green       NA            B         1
#> 5     5       NA    30 NA       Green       France        A         1
#> 6     6        2    30 G.D.     Dark brown  Ghana         A         1
#> 7     7        2    30 G.D.     Brown       Republic of ~ B         2
#> # ... with 5 more variables: stringsAsFactors <lgl>, sn <int>, pid <dbl>,
#> #   pid_cri <chr>, pid_dataset <chr>

bind_cols(data_4, record_group(data_4,r_id, c(staff_id, age), list(s2a=c("initials","hair_colour","branch_office")), data_source = c(ds_1, ds_2)))

#> 
#> Group criteria 1 - `staff_id`
#> 2 of 7 record(s) have been assigned a group ID. 5 records not yet grouped.
#> 0 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Group criteria 2 - `age`
#> Matching criteria 2: iteration 2
#> Matching criteria 2: iteration 3
#> 5 of 5 record(s) have been assigned a group ID. 0 records not yet grouped.
#> 0 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a unique ID. 

#> # A tibble: 7 x 13
#>    r_id staff_id   age initials hair_colour branch_office ds_1   ds_2
#>   <int>    <dbl> <dbl> <chr>    <chr>       <chr>         <chr> <dbl>
#> 1     1       NA    30 G.D.     Brown       Republic of ~ A         3
#> 2     2       NA    30 B.G.     Teal        France        A         1
#> 3     3       NA    30 X.P.     NA          NA            A         1
#> 4     4       NA    30 X.P.     Green       NA            B         1
#> 5     5       NA    30 NA       Green       France        A         1
#> 6     6        2    30 G.D.     Dark brown  Ghana         A         1
#> 7     7        2    30 G.D.     Brown       Republic of ~ B         2
#> # ... with 5 more variables: stringsAsFactors <lgl>, sn <int>, pid <dbl>,
#> #   pid_cri <chr>, pid_dataset <chr>