Multistage deterministic record linkage

Group matching records from one or more datasets.

record_group(df, sn = NULL, criteria, sub_criteria = NULL,
  data_source = NULL, group_stats = FALSE, display = TRUE)

Arguments

df
data.frame. One or more datasets appended together.

sn
Unique numeric record identifier. Optional.

criteria
Column names of the attributes to match. Records with matching values in these columns are grouped together.

sub_criteria
Matching sub-criteria. Additional matching conditions for each stage (criteria).

data_source
Unique dataset identifier. Useful when data.frame contains data from multiple datasets.

group_stats
If TRUE, output will include additional columns with useful stats for each record group.

display
If TRUE, status messages are printed on screen.

Value

data.frame

sn - unique record identifier as provided

pid - unique group identifier

pid_cri - matching criteria for each record in the group

pid_dataset - list of datasets in each group

pid_total - number of records in each group

Details

Record grouping occurs in stages of matching criteria.

Records are matched in two ways; an exact match - the equivalent of (==), or matching a range of values. An example of range matching is matching on a date give or take 5 days, or matching on age give or take 2 years. To do this, create a number_line object with the range of values, and assign the actual value to the gid argument. Then use the number_line as a sub_criteria.

A match at each stage is considered more certain than those at subsequent stages. Therefore, criteria should be listed in order of decreasing certainty.

sub_criteria can be used to force additional matching conditions at each stage. If sub_criteria is not NULL, only records with matching criteria and sub_criteria values are grouped together. If a record has missing values for any criteria, it's skipped at that stage, and another attempt is made at the next stage. If all criteria values are missing, that record is assigned a unique group ID.

When a data_source identifier is included, pid_dataset is included in the output. This shows the datasets included in each group.

See also

Examples

library(dplyr)
library(tidyr)

three_people <- data.frame(forename=c("Obinna","James","Ojay","James","Obinna"),
stringsAsFactors = FALSE)
bind_cols(three_people, record_group(three_people, criteria= forename))

#> 
#> Group criteria 1 - `forename`
#> 5 of 5 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 1 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 1.
#> 
#> Record grouping complete - 1 record(s) assigned a group unique ID. 

#>   forename sn pid    pid_cri
#> 1   Obinna  1   1 Criteria 1
#> 2    James  2   2 Criteria 1
#> 3     Ojay  3   3       None
#> 4    James  4   2 Criteria 1
#> 5   Obinna  5   1 Criteria 1

# To handle missing or unknown data, recode missing or unknown values to NA or "".
three_people$r_id <- 1:5
three_people$forename <- ifelse(three_people$r_id %in% c(1,4), NA, three_people$forename)
bind_cols(three_people, record_group(three_people, criteria= forename))

#> 
#> Group criteria 1 - `forename`
#> 3 of 5 record(s) have been assigned a group ID. 2 record(s) not yet grouped.
#> 3 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Record grouping complete - 5 record(s) assigned a group unique ID. 

#>   forename r_id sn pid pid_cri
#> 1     <NA>    1  1   1    None
#> 2    James    2  2   2    None
#> 3     Ojay    3  3   3    None
#> 4     <NA>    4  4   4    None
#> 5   Obinna    5  5   5    None

data(staff_records); staff_records

#> # A tibble: 7 x 5
#>    r_id forename surname  sex   dataset   
#>   <int> <chr>    <chr>    <chr> <chr>     
#> 1     1 James    Green    M     Staff list
#> 2     2 NA       Anderson M     Staff list
#> 3     3 Jamey    Green    M     Pay slips 
#> 4     4 ""       NA       F     Pay slips 
#> 5     5 Derrick  Anderson M     Staff list
#> 6     6 Darrack  Anderson M     Pay slips 
#> 7     7 Christie Green    F     Staff list

# Range matching
dob <- select(staff_records, sex)
dob$age <- c(10,8,20,5,5,9,7)

# age range - age + 20 years
dob$range <- number_line(dob$age, dob$age+20, gid=dob$age)
bind_cols(dob, record_group(dob, criteria = sex, sub_criteria = list(s1a="range"), display = FALSE))

#> Record grouping complete - 1 record(s) assigned a group unique ID. 

#> # A tibble: 7 x 6
#>   sex     age range         sn   pid pid_cri   
#>   <chr> <dbl> <numbr_ln> <int> <dbl> <chr>     
#> 1 M        10 10 -> 30       1     2 Criteria 1
#> 2 M         8 8 -> 28        2     2 Criteria 1
#> 3 M        20 20 -> 40       3     2 Criteria 1
#> 4 F         5 5 -> 25        4     4 Criteria 1
#> 5 M         5 5 -> 25        5     5 None      
#> 6 M         9 9 -> 29        6     2 Criteria 1
#> 7 F         7 7 -> 27        7     4 Criteria 1

# age range - age +- 20 years
dob$range <- number_line(dob$age-20, dob$age+20, gid=dob$age)
bind_cols(dob, record_group(dob, criteria = sex, sub_criteria = list(s1a="range")))

#> 
#> Group criteria 1 - `sex`
#> 7 of 7 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a group unique ID. 

#> # A tibble: 7 x 6
#>   sex     age range         sn   pid pid_cri   
#>   <chr> <dbl> <numbr_ln> <int> <int> <chr>     
#> 1 M        10 -10 -> 30      1     1 Criteria 1
#> 2 M         8 -12 -> 28      2     1 Criteria 1
#> 3 M        20 0 -> 40        3     1 Criteria 1
#> 4 F         5 -15 -> 25      4     4 Criteria 1
#> 5 M         5 -15 -> 25      5     1 Criteria 1
#> 6 M         9 -11 -> 29      6     1 Criteria 1
#> 7 F         7 -13 -> 27      7     4 Criteria 1

# Do not directly use number_line objects as criterias.
# Instead, use it as the sub_criteria to a 'dummy criteria'
dob$dum_var <- 1
bind_cols(dob, record_group(dob, criteria = dum_var, sub_criteria = list(s1a="range")))

#> 
#> Group criteria 1 - `dum_var`
#> 7 of 7 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a group unique ID. 

#> # A tibble: 7 x 7
#>   sex     age range      dum_var    sn   pid pid_cri   
#>   <chr> <dbl> <numbr_ln>   <dbl> <int> <int> <chr>     
#> 1 M        10 -10 -> 30        1     1     1 Criteria 1
#> 2 M         8 -12 -> 28        1     2     1 Criteria 1
#> 3 M        20 0 -> 40          1     3     1 Criteria 1
#> 4 F         5 -15 -> 25        1     4     1 Criteria 1
#> 5 M         5 -15 -> 25        1     5     1 Criteria 1
#> 6 M         9 -11 -> 29        1     6     1 Criteria 1
#> 7 F         7 -13 -> 27        1     7     1 Criteria 1

# Two or more stages of record grouping
pids <- record_group(staff_records, sn = r_id, criteria = c(forename, surname),
data_source = sex, display = FALSE)

#> Record grouping complete - 1 record(s) assigned a group unique ID. 

left_join(staff_records, pids, by=c("r_id"="sn"))

#> # A tibble: 7 x 8
#>    r_id forename surname  sex   dataset      pid pid_cri    pid_dataset
#>   <int> <chr>    <chr>    <chr> <chr>      <dbl> <chr>      <chr>      
#> 1     1 James    Green    M     Staff list     1 Criteria 2 F,M        
#> 2     2 NA       Anderson M     Staff list     2 Criteria 2 M          
#> 3     3 Jamey    Green    M     Pay slips      1 Criteria 2 F,M        
#> 4     4 ""       NA       F     Pay slips      4 None       F          
#> 5     5 Derrick  Anderson M     Staff list     2 Criteria 2 M          
#> 6     6 Darrack  Anderson M     Pay slips      2 Criteria 2 M          
#> 7     7 Christie Green    F     Staff list     1 Criteria 2 F,M        

# Add sex to the second stage to be more certain
staff_records_b <- unite(staff_records, cri_2, c(surname, sex), sep ="-")
pids <- record_group(staff_records_b, r_id, c(forename, cri_2),
data_source = dataset, display = FALSE)

#> Record grouping complete - 2 record(s) assigned a group unique ID. 

bind_cols(staff_records_b, pids)

#> # A tibble: 7 x 8
#>    r_id forename cri_2      dataset       sn   pid pid_cri   pid_dataset        
#>   <int> <chr>    <chr>      <chr>      <int> <dbl> <chr>     <chr>              
#> 1     1 James    Green-M    Staff list     1     1 Criteria~ Pay slips,Staff li~
#> 2     2 NA       Anderson-M Staff list     2     2 Criteria~ Pay slips,Staff li~
#> 3     3 Jamey    Green-M    Pay slips      3     1 Criteria~ Pay slips,Staff li~
#> 4     4 ""       NA-F       Pay slips      4     4 None      Pay slips          
#> 5     5 Derrick  Anderson-M Staff list     5     2 Criteria~ Pay slips,Staff li~
#> 6     6 Darrack  Anderson-M Pay slips      6     2 Criteria~ Pay slips,Staff li~
#> 7     7 Christie Green-F    Staff list     7     7 None      Staff list         

# Using sub-criteria
data(missing_staff_id); missing_staff_id

#> # A tibble: 7 x 8
#>    r_id staff_id   age initials hair_colour branch_office     source_1 source_2
#>   <int>    <dbl> <dbl> <chr>    <chr>       <chr>             <chr>       <dbl>
#> 1     1       NA    30 G.D.     Brown       Republic of Ghana A               3
#> 2     2       NA    30 B.G.     Teal        France            A               1
#> 3     3       NA    30 X.P.     NA          NA                A               1
#> 4     4       NA    30 X.P.     Green       NA                B               1
#> 5     5       NA    30 NA       Green       France            A               1
#> 6     6        2    30 G.D.     Dark brown  Ghana             A               1
#> 7     7        2    30 G.D.     Brown       Republic of Ghana B               2

pids <- record_group(missing_staff_id, r_id, c(staff_id, age),
list(s2a=c("initials","hair_colour","branch_office")), data_source = source_1)

#> 
#> Group criteria 1 - `staff_id`
#> 2 of 7 record(s) have been assigned a group ID. 5 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Group criteria 2 - `age`
#> Matching criteria 2: iteration 2
#> Matching criteria 2: iteration 3
#> 5 of 5 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a group unique ID. 

left_join(missing_staff_id, pids, by=c("r_id"="sn"))

#> # A tibble: 7 x 11
#>    r_id staff_id   age initials hair_colour branch_office source_1 source_2
#>   <int>    <dbl> <dbl> <chr>    <chr>       <chr>         <chr>       <dbl>
#> 1     1       NA    30 G.D.     Brown       Republic of ~ A               3
#> 2     2       NA    30 B.G.     Teal        France        A               1
#> 3     3       NA    30 X.P.     NA          NA            A               1
#> 4     4       NA    30 X.P.     Green       NA            B               1
#> 5     5       NA    30 NA       Green       France        A               1
#> 6     6        2    30 G.D.     Dark brown  Ghana         A               1
#> 7     7        2    30 G.D.     Brown       Republic of ~ B               2
#> # ... with 3 more variables: pid <dbl>, pid_cri <chr>, pid_dataset <chr>

pids <- record_group(missing_staff_id, r_id, c(staff_id, age),
list(s2a=c("initials","hair_colour","branch_office")), data_source = c(source_1, source_2))

#> 
#> Group criteria 1 - `staff_id`
#> 2 of 7 record(s) have been assigned a group ID. 5 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Group criteria 2 - `age`
#> Matching criteria 2: iteration 2
#> Matching criteria 2: iteration 3
#> 5 of 5 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a group unique ID. 

bind_cols(missing_staff_id, pids)

#> # A tibble: 7 x 12
#>    r_id staff_id   age initials hair_colour branch_office source_1 source_2
#>   <int>    <dbl> <dbl> <chr>    <chr>       <chr>         <chr>       <dbl>
#> 1     1       NA    30 G.D.     Brown       Republic of ~ A               3
#> 2     2       NA    30 B.G.     Teal        France        A               1
#> 3     3       NA    30 X.P.     NA          NA            A               1
#> 4     4       NA    30 X.P.     Green       NA            B               1
#> 5     5       NA    30 NA       Green       France        A               1
#> 6     6        2    30 G.D.     Dark brown  Ghana         A               1
#> 7     7        2    30 G.D.     Brown       Republic of ~ B               2
#> # ... with 4 more variables: sn <int>, pid <dbl>, pid_cri <chr>,
#> #   pid_dataset <chr>