Multistage deterministic record linkage

Group matching records from one or more datasets.

record_group(df, sn = NULL, criteria, sub_criteria = NULL,
  data_source = NULL, group_stats = FALSE, display = TRUE,
  to_s4 = TRUE)

Arguments

df
data.frame. One or more datasets appended together.

sn
Unique numerical record identifier. Optional.

criteria
Column names of attributes to match. Each column is one stage of the process and the the order in which they are listed determines the relevance of matches. Matching records are assigned to a record group.

sub_criteria
Matching sub-criteria. Additional matching conditions for each stage (criteria).

data_source
Unique dataset identifier. Useful when df contains data from multiple sources.

group_stats
If TRUE, output will include additional columns with useful stats for each record group.

display
If TRUE, status messages are printed on screen.

to_s4
if TRUE, changes the returned output to a pid object.

Value

pid objects or data.frame if to_s4 is FALSE)

sn - unique record identifier as provided

pid | .Data - unique group identifier

pid_cri - matched criteria for each record in the group

pid_dataset - data sources in each group

pid_total - number of records in each group

pid objects will be the default output from the next release.

Details

Record grouping occurs in stages of matching criteria.

Records are matched in two ways; an exact match - the equivalent of (==), or matching a range of numeric values. An example of range matching is matching a date give or take 5 days, or matching an age give or take 2 years. To do this, create a number_line object to represent the range and supply this to criteria or sub_criteria. The actual value within each range must be assigned to the gid slot of the number_line object.

A match at each stage is considered more relevant than those at subsequent stages. Therefore, criteria should be listed in order of decreasing relevance or certainty.

sub_criteria can be used to force additional matching conditions at each stage. If sub_criteria is not NULL, only records with matching criteria and sub_criteria values are grouped together. If a record has missing values for any criteria, that record skipped at that stage, and another attempt is made at the next stage. If there are no matches for a record at every stage, that record is assigned a unique group ID.

When a data_source identifier is provided, pid_dataset is included in the output. This has the source of every record in each record group.

See also

Examples

library(dplyr)

three_people <- data.frame(forename=c("Obinna","James","Ojay","James","Obinna"),
                           stringsAsFactors = FALSE)

three_people$pids_a <- record_group(three_people, criteria= forename, to_s4 = TRUE)

#> 
#> Group criteria 1 - `forename`
#> 5 of 5 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 1 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 1.
#> 
#> Record grouping complete - 1 record(s) assigned a group unique ID. 

three_people

#>   forename       pids_a
#> 1   Obinna P.1 (CRI 01)
#> 2    James P.2 (CRI 01)
#> 3     Ojay P.3 (No Hit)
#> 4    James P.2 (CRI 01)
#> 5   Obinna P.1 (CRI 01)

# To handle missing or unknown data, recode missing or unknown values to NA or "".
three_people$forename[c(1,4)] <- NA
three_people$pids_b <- record_group(three_people, criteria= forename, to_s4 =TRUE)

#> 
#> Group criteria 1 - `forename`
#> 3 of 5 record(s) have been assigned a group ID. 2 record(s) not yet grouped.
#> 3 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Record grouping complete - 5 record(s) assigned a group unique ID. 

three_people

#>   forename       pids_a       pids_b
#> 1     <NA> P.1 (CRI 01) P.1 (No Hit)
#> 2    James P.2 (CRI 01) P.2 (No Hit)
#> 3     Ojay P.3 (No Hit) P.3 (No Hit)
#> 4     <NA> P.2 (CRI 01) P.4 (No Hit)
#> 5   Obinna P.1 (CRI 01) P.5 (No Hit)

data(staff_records); staff_records

#> # A tibble: 7 x 5
#>    r_id forename surname  sex   dataset   
#>   <int> <chr>    <chr>    <chr> <chr>     
#> 1     1 James    Green    M     Staff list
#> 2     2 <NA>     Anderson M     Staff list
#> 3     3 Jamey    Green    M     Pay slips 
#> 4     4 ""       <NA>     F     Pay slips 
#> 5     5 Derrick  Anderson M     Staff list
#> 6     6 Darrack  Anderson M     Pay slips 
#> 7     7 Christie Green    F     Staff list

# Range matching
dob <- staff_records["sex"]
dob$age <- c(30,28,40,25,25,29,27)

# age range: age + 20 years
dob$range_a <- number_line(dob$age, dob$age+20, gid=dob$age)
dob$pids_a <- record_group(dob, criteria = sex, sub_criteria = list(s1a="range_a"), to_s4 = TRUE)

#> 
#> Group criteria 1 - `sex`
#> Matching criteria 1: iteration 2
#> Matching criteria 1: iteration 3
#> Matching criteria 1: iteration 4
#> Matching criteria 1: iteration 5
#> Matching criteria 1: iteration 6
#> 7 of 7 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a group unique ID. 

dob[c("sex","age","range_a","pids_a")]

#> # A tibble: 7 x 4
#>   sex     age range_a    pids_a      
#>   <chr> <dbl> <numbr_ln> <pid>       
#> 1 M        30 30 -> 50   P.5 (CRI 01)
#> 2 M        28 28 -> 48   P.5 (CRI 01)
#> 3 M        40 40 -> 60   P.5 (CRI 01)
#> 4 F        25 25 -> 45   P.4 (CRI 01)
#> 5 M        25 25 -> 45   P.5 (CRI 01)
#> 6 M        29 29 -> 49   P.5 (CRI 01)
#> 7 F        27 27 -> 47   P.4 (CRI 01)

# age range: age +- 20 years
dob$range_b <- number_line(dob$age-20, dob$age+20, gid=dob$age)
dob$pids_b <- record_group(dob, criteria = sex, sub_criteria = list(s1a="range_b"), to_s4 = TRUE)

#> 
#> Group criteria 1 - `sex`
#> 7 of 7 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a group unique ID. 

dob[c("sex","age","range_b","pids_b")]

#> # A tibble: 7 x 4
#>   sex     age range_b    pids_b      
#>   <chr> <dbl> <numbr_ln> <pid>       
#> 1 M        30 10 -> 50   P.1 (CRI 01)
#> 2 M        28 8 -> 48    P.1 (CRI 01)
#> 3 M        40 20 -> 60   P.1 (CRI 01)
#> 4 F        25 5 -> 45    P.4 (CRI 01)
#> 5 M        25 5 -> 45    P.1 (CRI 01)
#> 6 M        29 9 -> 49    P.1 (CRI 01)
#> 7 F        27 7 -> 47    P.4 (CRI 01)

dob$pids_c <- record_group(dob, criteria = range_b, to_s4 = TRUE)

#> 
#> Group criteria 1 - `dmvr_d1`
#> 7 of 7 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a group unique ID. 

dob[c("age","range_b","pids_c")]

#> # A tibble: 7 x 3
#>     age range_b    pids_c      
#>   <dbl> <numbr_ln> <pid>       
#> 1    30 10 -> 50   P.1 (CRI 01)
#> 2    28 8 -> 48    P.1 (CRI 01)
#> 3    40 20 -> 60   P.1 (CRI 01)
#> 4    25 5 -> 45    P.1 (CRI 01)
#> 5    25 5 -> 45    P.1 (CRI 01)
#> 6    29 9 -> 49    P.1 (CRI 01)
#> 7    27 7 -> 47    P.1 (CRI 01)

# Multistage record grouping
staff_records$pids_a <- record_group(staff_records, sn = r_id, criteria = c(forename, surname),
                                     data_source = sex, display = FALSE, to_s4 = TRUE)

#> Record grouping complete - 1 record(s) assigned a group unique ID. 

staff_records

#> # A tibble: 7 x 6
#>    r_id forename surname  sex   dataset    pids_a      
#>   <int> <chr>    <chr>    <chr> <chr>      <pid>       
#> 1     1 James    Green    M     Staff list P.1 (CRI 02)
#> 2     2 <NA>     Anderson M     Staff list P.2 (CRI 02)
#> 3     3 Jamey    Green    M     Pay slips  P.1 (CRI 02)
#> 4     4 ""       <NA>     F     Pay slips  P.4 (No Hit)
#> 5     5 Derrick  Anderson M     Staff list P.2 (CRI 02)
#> 6     6 Darrack  Anderson M     Pay slips  P.2 (CRI 02)
#> 7     7 Christie Green    F     Staff list P.1 (CRI 02)

# Add `sex` to the second stage (`cri`) to be more certain
staff_records$cri_2 <- paste0(staff_records$surname,"-", staff_records$sex)
staff_records$pids_b <- record_group(staff_records, r_id, c(forename, cri_2),
                                     data_source = dataset, display = FALSE, to_s4 = TRUE)

#> Record grouping complete - 2 record(s) assigned a group unique ID. 

staff_records

#> # A tibble: 7 x 8
#>    r_id forename surname  sex   dataset    pids_a       cri_2      pids_b      
#>   <int> <chr>    <chr>    <chr> <chr>      <pid>        <chr>      <pid>       
#> 1     1 James    Green    M     Staff list P.1 (CRI 02) Green-M    P.1 (CRI 02)
#> 2     2 <NA>     Anderson M     Staff list P.2 (CRI 02) Anderson-M P.2 (CRI 02)
#> 3     3 Jamey    Green    M     Pay slips  P.1 (CRI 02) Green-M    P.1 (CRI 02)
#> 4     4 ""       <NA>     F     Pay slips  P.4 (No Hit) NA-F       P.4 (No Hit)
#> 5     5 Derrick  Anderson M     Staff list P.2 (CRI 02) Anderson-M P.2 (CRI 02)
#> 6     6 Darrack  Anderson M     Pay slips  P.2 (CRI 02) Anderson-M P.2 (CRI 02)
#> 7     7 Christie Green    F     Staff list P.1 (CRI 02) Green-F    P.7 (No Hit)

# Using sub-criteria
data(missing_staff_id); missing_staff_id

#> # A tibble: 7 x 8
#>    r_id staff_id   age initials hair_colour branch_office     source_1 source_2
#>   <int>    <dbl> <dbl> <chr>    <chr>       <chr>             <chr>       <dbl>
#> 1     1       NA    30 G.D.     Brown       Republic of Ghana A               3
#> 2     2       NA    30 B.G.     Teal        France            A               1
#> 3     3       NA    30 X.P.     <NA>        <NA>              A               1
#> 4     4       NA    30 X.P.     Green       <NA>              B               1
#> 5     5       NA    30 <NA>     Green       France            A               1
#> 6     6        2    30 G.D.     Dark brown  Ghana             A               1
#> 7     7        2    30 G.D.     Brown       Republic of Ghana B               2

missing_staff_id$pids <- record_group(missing_staff_id, r_id, c(staff_id, age),
list(s2a=c("initials","hair_colour","branch_office")), data_source = source_1, to_s4 = TRUE)

#> 
#> Group criteria 1 - `staff_id`
#> 2 of 7 record(s) have been assigned a group ID. 5 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Group criteria 2 - `age`
#> Matching criteria 2: iteration 2
#> Matching criteria 2: iteration 3
#> 5 of 5 record(s) have been assigned a group ID. 0 record(s) not yet grouped.
#> 0 record(s) with unique group IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a group unique ID. 

missing_staff_id

#> # A tibble: 7 x 9
#>    r_id staff_id   age initials hair_colour branch_office source_1 source_2
#>   <int>    <dbl> <dbl> <chr>    <chr>       <chr>         <chr>       <dbl>
#> 1     1       NA    30 G.D.     Brown       Republic of ~ A               3
#> 2     2       NA    30 B.G.     Teal        France        A               1
#> 3     3       NA    30 X.P.     <NA>        <NA>          A               1
#> 4     4       NA    30 X.P.     Green       <NA>          B               1
#> 5     5       NA    30 <NA>     Green       France        A               1
#> 6     6        2    30 G.D.     Dark brown  Ghana         A               1
#> 7     7        2    30 G.D.     Brown       Republic of ~ B               2
#> # ... with 1 more variable: pids <pid>