Deterministic linkage for epidemiological analysis

This function links records from one or more datasets into unique group idenfiers.

record_group(df, sn, criteria, sub_criteria = NULL, data_source = NULL,
  group_stats = FALSE, display = TRUE)

Arguments

df
Dataframe. One or more datasets appened together.

sn
numeric unique record indentifier for the dataframe.

criteria
Matching criteria. Records with matching values in these columns are grouped together.

sub_criteria
Matching sub-criteria. Additional conditions for a match at each stage (criteria).

data_source
Unique dataset indentifier. Useful when dataframe contains data from multiple datasets.

group_stats
If TRUE, output will include additional columns with useful stats for each record group.

display
If TRUE, status messages are not printed on screen.

Value

Dataframe

sn - the unique record identifier provided

pid - unique record group indentifier

pid_cri - matching criteria each record

pid_dataset - list of datasets in each record group

pid_total - number of records in each record group

Details

Record linkage occurs in stages as listed in criteria. Linkage at each stage is considered more certain than subsequent stages. Therefore, criteria should be listed in order of decreasing certainty. sub_criteria can be used to force additonal conditions for record linkage at each stage. If sub_criteria is not NULL, only records with matching criteria and sub_criteria values are grouped together. If a criteria value is missing, that record is skipped for that stage of record linkage, and another attempt is made at the next stage. If all criteria values are missing, that record is assigned a unique group ID. When a data_source identifier is included, pid_dataset is included in the output. This shows the datasets included in each record group.

See also

Examples

 library(dplyr)
 library(tidyr)

# One stage record grouping
 df <- data.frame(
   r_id = c(1:5),
   forename = c("Obinna","James","Ojay","James","Obinna"),
   stringsAsFactors = FALSE
 )

 cbind(df, record_group(df, r_id, forename))

#> 
#> Group criteria 1 - `forename`
#> 5 of 5 record(s) have been assigned a group ID. 0 records not yet grouped.
#> 1 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 1.
#> 
#> Record grouping complete - 1 record(s) assigned a unique ID. 

#>   r_id forename sn pid    pid_cri
#> 1    1   Obinna  1   1 Criteria 1
#> 2    2    James  2   2 Criteria 1
#> 3    3     Ojay  3   3       None
#> 4    4    James  4   2 Criteria 1
#> 5    5   Obinna  5   1 Criteria 1

 # Handling missing or unknown data - Recode missing or unknown values to NA or "".
 # These are excluded from record grouping at the relevant stage

 df %>%
   mutate(forename = ifelse(r_id %in% c(1,4), NA, forename)) %>%
   cbind(record_group(., r_id, forename))

#> 
#> Group criteria 1 - `forename`
#> 3 of 5 record(s) have been assigned a group ID. 2 records not yet grouped.
#> 3 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Record grouping complete - 5 record(s) assigned a unique ID. 

#>   r_id forename sn pid pid_cri
#> 1    1     <NA>  1   1    None
#> 2    2    James  2   2    None
#> 3    3     Ojay  3   3    None
#> 4    4     <NA>  4   4    None
#> 5    5   Obinna  5   5    None

 # Two or more stages of record grouping
 # Criteria should be listed in decreasing order of certainty
data_3 <- data.frame(
  forename = c("James",NA,"Jamey","","Derrick","Darrack","Christie"),
  surname = c("Green","Anderson","Green",NA,"Anderson","Anderson","Green"),
  sex = c("M","M","M","F","M","M","F"),
  dataset = c("Staff list","Staff list", "Pay slips","Pay slips", "Staff list","Pay slips","Staff list"),
  r_id = c(1:7),
  stringsAsFactors = FALSE
)

cbind(data_3, record_group(data_3,r_id, c(forename, surname), data_source = sex, display = FALSE))

#> Record grouping complete - 1 record(s) assigned a unique ID. 

#>   forename  surname sex    dataset r_id sn pid    pid_cri pid_dataset
#> 1    James    Green   M Staff list    1  1   1 Criteria 2         F,M
#> 2     <NA> Anderson   M Staff list    2  2   2 Criteria 2           M
#> 3    Jamey    Green   M  Pay slips    3  3   1 Criteria 2         F,M
#> 4              <NA>   F  Pay slips    4  4   4       None           F
#> 5  Derrick Anderson   M Staff list    5  5   2 Criteria 2           M
#> 6  Darrack Anderson   M  Pay slips    6  6   2 Criteria 2           M
#> 7 Christie    Green   F Staff list    7  7   1 Criteria 2         F,M

# Add sex to the second stage to be more certain
data_3 %>%
  unite(cri_2, c(surname, sex)) %>%
  cbind(record_group(., r_id, c(forename, cri_2), data_source = dataset, display = FALSE))

#> Record grouping complete - 2 record(s) assigned a unique ID. 

#>   forename      cri_2    dataset r_id sn pid    pid_cri          pid_dataset
#> 1    James    Green_M Staff list    1  1   1 Criteria 2 Pay slips,Staff list
#> 2     <NA> Anderson_M Staff list    2  2   2 Criteria 2 Pay slips,Staff list
#> 3    Jamey    Green_M  Pay slips    3  3   1 Criteria 2 Pay slips,Staff list
#> 4                NA_F  Pay slips    4  4   4       None            Pay slips
#> 5  Derrick Anderson_M Staff list    5  5   2 Criteria 2 Pay slips,Staff list
#> 6  Darrack Anderson_M  Pay slips    6  6   2 Criteria 2 Pay slips,Staff list
#> 7 Christie    Green_F Staff list    7  7   7       None           Staff list

# Using sub-criteria
data_4 <- data.frame(
  staff_id = c(NA,NA,NA,NA,NA,2,2),
  age = rep(30,7),
  initials = c("G.D.","B.G.","X.P.","X.P.",NA,"G.D.","G.D."),
  hair_colour = c("Brown","Teal",NA,"Green","Green","Dark brown","Brown"),
  branch_office = c("Republic of Ghana","France",NA,NA,"France","Ghana","Republic of Ghana"),
  ds_1 = c("A","A","A","B","A","A","B"),
  ds_2 = c(3,1,1,1,1,1,2),
  r_id = c(1:7),
  stringsAsFactors = FALSE
)

cbind(data_4, record_group(data_4,r_id, c(staff_id, age), list(s2a=c("initials","hair_colour","branch_office")), data_source = ds_1))

#> 
#> Group criteria 1 - `staff_id`
#> 2 of 7 record(s) have been assigned a group ID. 5 records not yet grouped.
#> 0 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Group criteria 2 - `age`
#> Matching criteria 2: iteration 2
#> Matching criteria 2: iteration 3
#> 5 of 5 record(s) have been assigned a group ID. 0 records not yet grouped.
#> 0 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a unique ID. 

#>   staff_id age initials hair_colour     branch_office ds_1 ds_2 r_id sn pid
#> 1       NA  30     G.D.       Brown Republic of Ghana    A    3    1  1   6
#> 2       NA  30     B.G.        Teal            France    A    1    2  2   2
#> 3       NA  30     X.P.        <NA>              <NA>    A    1    3  3   3
#> 4       NA  30     X.P.       Green              <NA>    B    1    4  4   3
#> 5       NA  30     <NA>       Green            France    A    1    5  5   2
#> 6        2  30     G.D.  Dark brown             Ghana    A    1    6  6   6
#> 7        2  30     G.D.       Brown Republic of Ghana    B    2    7  7   6
#>      pid_cri pid_dataset
#> 1 Criteria 2         A,B
#> 2 Criteria 2           A
#> 3 Criteria 2         A,B
#> 4 Criteria 2         A,B
#> 5 Criteria 2           A
#> 6 Criteria 1         A,B
#> 7 Criteria 1         A,B

cbind(data_4, record_group(data_4,r_id, c(staff_id, age), list(s2a=c("initials","hair_colour","branch_office")), data_source = c(ds_1, ds_2)))

#> 
#> Group criteria 1 - `staff_id`
#> 2 of 7 record(s) have been assigned a group ID. 5 records not yet grouped.
#> 0 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 5.
#> 
#> Group criteria 2 - `age`
#> Matching criteria 2: iteration 2
#> Matching criteria 2: iteration 3
#> 5 of 5 record(s) have been assigned a group ID. 0 records not yet grouped.
#> 0 record(s) with unique IDs untagged for possible matching in the next stage. The number of records not yet grouped is now 0.
#> 
#> Record grouping complete - 0 record(s) assigned a unique ID. 

#>   staff_id age initials hair_colour     branch_office ds_1 ds_2 r_id sn pid
#> 1       NA  30     G.D.       Brown Republic of Ghana    A    3    1  1   6
#> 2       NA  30     B.G.        Teal            France    A    1    2  2   2
#> 3       NA  30     X.P.        <NA>              <NA>    A    1    3  3   3
#> 4       NA  30     X.P.       Green              <NA>    B    1    4  4   3
#> 5       NA  30     <NA>       Green            France    A    1    5  5   2
#> 6        2  30     G.D.  Dark brown             Ghana    A    1    6  6   6
#> 7        2  30     G.D.       Brown Republic of Ghana    B    2    7  7   6
#>      pid_cri pid_dataset
#> 1 Criteria 2 A-1,A-3,B-2
#> 2 Criteria 2         A-1
#> 3 Criteria 2     A-1,B-1
#> 4 Criteria 2     A-1,B-1
#> 5 Criteria 2         A-1
#> 6 Criteria 1 A-1,A-3,B-2
#> 7 Criteria 1 A-1,A-3,B-2